### 2025-01-14 #7666 @editorialbot
[PRE REVIEW]: ASSUME - Agent-based Simulation for Studying and Understanding Market Evolution
**Submitting author:** @maurerle (<a href="http://orcid.org/0000-0001-8345-3889">Florian Maurer</a>)
**Repository:** https://github.com/assume-framework/assume
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.3
