+++ assume/scenario/loader_oeds.py
@@ -8,6 +8,7 @@ import shutil
 from datetime import datetime, timedelta
 from pathlib import Path
 
+import numpy as np
 import pandas as pd
 from dateutil import rrule as rr
 
@@ -29,6 +30,7 @@ def load_oeds(
     marketdesign: list[MarketConfig],
     bidding_strategies: dict[str, str],
     nuts_config: list[str] = [],
+    random=True,
 ):
     """
     This initializes a scenario using the open-energy-data-server
@@ -50,11 +52,18 @@ def load_oeds(
         freq="h",
     )
     sim_id = f"{scenario}_{study_case}"
+    year = start.year
     logger.info(f"loading scenario {sim_id} with {nuts_config}")
     infra_interface = InfrastructureInterface("test", infra_uri)
 
-    if not nuts_config:
+    if not nuts_config or nuts_config == "nuts3":
         nuts_config = list(infra_interface.plz_nuts["nuts3"].unique())
+    elif nuts_config == "nuts0":
+        nuts_config = ["DE"]
+    elif nuts_config == "nuts1":
+        nuts_config = list(infra_interface.plz_nuts["nuts1"].unique())
+    elif nuts_config == "nuts2":
+        nuts_config = list(infra_interface.plz_nuts["nuts2"].unique())
 
     world.setup(
         start=start,
@@ -69,31 +78,73 @@ def load_oeds(
     for market_config in marketdesign:
         world.add_market(mo_id, market_config)
 
-    co2_price = 1
-    fuel_prices = {
-        "hard coal": 8.6,
+    fuel_prices = {  # €/MWh
+        "hard coal": infra_interface.get_coal_price(start, end),  # 8.6
         "lignite": 1.8,
-        "oil": 22,
-        "gas": 26,
+        "oil": infra_interface.get_oil_price(start, end),  # 22
+        "gas": infra_interface.get_gas_price(start, end),  # 26
         "biomass": 20,
         "nuclear": 1,
-        "co2": 20,
+        "co2": infra_interface.get_co2_price(start, end),  # 1€/tCO2
     }
+    for name in fuel_prices.keys():
+        if not isinstance(fuel_prices[name], float | int):
+            fuel_prices[name] = (
+                fuel_prices[name].reindex(index, method="nearest").values
+            )
 
-    # for each area - add demand and generation
+    offshore_wind = infra_interface.get_offshore_wind_series(start, end)
+    if offshore_wind.max() > 0:
+        world.add_unit_operator("renewables_offshore")
+        world.add_unit(
+            "renewables_off_wind",
+            "power_plant",
+            "renewables_offshore",
+            # the unit_params have no hints
+            {
+                "min_power": 0,
+                "max_power": offshore_wind.max(),
+                "bidding_strategies": bidding_strategies["wind"],
+                "technology": "wind_offshore",
+                "location": (8.18, 54.4),
+                "node": "DEF",
+            },
+            NaiveForecast(
+                index,
+                availability=offshore_wind / offshore_wind.max(),
+                fuel_price=0.2,
+                co2_price=0,
+            ),
+        )
+    # total german demand if area is not set
+    demand = infra_interface.get_country_demand(start, end, "DE")
+    demand = demand.resample("h").mean() / len(nuts_config)
+
     for area in nuts_config:
         logger.info(f"loading config {area} for {year}")
         config_path = Path.home() / ".assume" / f"{area}_{year}"
         if not config_path.is_dir():
             logger.info("query database time series")
-            demand = infra_interface.get_demand_series_in_area(area, year)
+            if area == "DE":
+                # for each area - add demand and generation
+                demand = infra_interface.get_country_demand(start, end, "DE")
[…]
