+++ assume/scenario/oeds/infrastructure.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import logging
+import math
 from datetime import datetime, timedelta
 
 import holidays
@@ -34,11 +35,27 @@ logger = logging.getLogger(__name__)
 
 
 class InfrastructureInterface:
+    """
+    The infrastructure interface abstracts access to the open-energy-data-server database.
+    The database contains different schemas to access the MaStR, ENTSO-E or ECMWF weather.
+    It allows the construction of a simulation solely from public data.
+    """
+
     def __init__(
         self,
         name,
         db_server_uri,
-        structure_databases=("mastr", "oep", "windmodel", "nuts", "scigrid", "weather"),
+        structure_databases=(
+            "mastr",
+            "oep",
+            "windmodel",
+            "nuts",
+            "scigrid",
+            "weather",
+            "opec",
+            "instrat_pl",
+            "entsoe",
+        ),
     ):
         self.databases = {}
         for db in structure_databases:
@@ -48,13 +65,14 @@ class InfrastructureInterface:
             self.databases[db] = create_engine(
                 f"{db_server_uri}?options=--search_path={schema_name}",
                 connect_args={"application_name": name},
+                pool_pre_ping=True,
             )
         self.setup()
 
     def setup(self):
         with self.databases["nuts"].connect() as conn:
             self.plz_nuts = pd.read_sql_query(
-                "select code, nuts3, longitude, latitude from plz",
+                "select code, nuts1, nuts2, nuts3, longitude, latitude from plz",
                 conn,
                 index_col="code",
             )
@@ -194,7 +212,21 @@ class InfrastructureInterface:
             df["kwkPowerElec"] = 0
         return df
 
-    def get_power_plant_in_area(self, area=52353, fuel_type="lignite"):
+    def get_power_plant_in_area(
+        self, area=52353, fuel_type="lignite", created_before=None, stopped_after=None
+    ):
+        """
+        returns the power plants of a given area and given fuel type filtered by commission and decommission date
+
+        Args:
+            area (int, optional): NUTS area or German postal code. Defaults to 52353.
+            fuel_type (str, optional): fuel type to filter for. Defaults to "lignite".
+            created_before (datetime.datetime, optional): time before which the power plants have to be created. Defaults to None.
+            stopped_after (datetime.datetime, optional): time until the power plants have to be running at least. Defaults to None.
+
+        Returns:
+            pandas.DataFrame: dataframe of power plants
+        """
         if isinstance(area, str) and area.startswith("DE"):
             plz_codes = self.get_plz_codes(area)
             if not plz_codes:
@@ -206,7 +238,7 @@ class InfrastructureInterface:
             if plz not in self.plz_nuts.index:
                 raise Exception("invalid plz code")
 
-        longitude, latitude = self.get_lat_lon_area(area)
+        latitude, longitude = self.get_lat_lon_area(area)
 
         query = f"""
             SELECT ev."EinheitMastrNummer" as "unitID",
@@ -218,8 +250,8 @@ class InfrastructureInterface:
             COALESCE(ev."Technologie", 839) as "turbineTyp",
             ev."GenMastrNummer" as "generatorID"
             """
-        plz_codes_str = ", ".join([str(x) for x in plz_codes])
-        plz_codes_str = f"({plz_codes_str})"
[…]
