### 2025-01-14 c95fe5b
learning examples with high training episodes and loop
