### 2025-01-14 a7eab28
(re)set defaults and move parameters to config
