### 2025-01-14 a80dc61
Fix small bug only adding single critic.
