### 2025-01-14 aafd888
fixup affiliations and license issues
