+++ paper/paper.md
@@ -1,9 +1,3 @@
-<!--
-SPDX-FileCopyrightText: ASSUME Developers
-
-SPDX-License-Identifier: AGPL-3.0-or-later
--->
-
 ---
 title: 'ASSUME - Agent-based Simulation for Studying and Understanding Market Evolution'
 tags:
@@ -18,11 +12,10 @@ authors:
     corresponding: true
     equal-contrib: true
     affiliation: 1
-    "affiliation": 1
   - name: Nick Harder
     orcid: 0000-0003-1897-3671
     equal-contrib: true
-    "affiliation": 2
+    affiliation: 2
   - name: Kim K. Miskiw
     orcid: 0009-0009-1389-4844
     equal-contrib: true
@@ -44,8 +37,6 @@ affiliations:
 date: 13 November 2024
 bibliography: paper.bib
 ---
-
-<!-- pandoc -s paper.md -o paper.pdf --bibliography paper.bib --csl=apa.csl --filter pandoc-citeproc --pdf-engine=xelatex-->
 
 # Summary
 
