### 2025-01-14 #7666 @maurerle
Hello @kyleniemeyer, I would suggest:
igarizio, imcatta, SebastianBoblest as reviewers and
fraukewiese as editor?

I also did take a look at the review of #5041 as AMIRIS is an agent-based market-simulation tool as well with a different focus.
