### 2025-01-14 #7666 @maurerle
The review in AMIRIS had a discussion about REUSE compliant license files in a "licenses" folder: https://github.com/openjournals/joss-reviews/issues/5041#issuecomment-1375849684

This was also discussed on the openmod forum: https://forum.openmod.org/t/reuse-incompatible-to-auto-detection-of-license-s/3590

So I think that our best approach to this is to add a duplicate LICENSE file to the root of the repository as well..?

I did so in https://github.com/assume-framework/assume/pull/482/commits/989df95ff74db8bf01a2a5fbb2d185bad56b49e1
