### 2025-01-15 16f13cc
More efficient implementation of unit specific metrics.
