### 2025-01-15 cdaf49f
fixes for tutorial_07 (#530)
* bidding strategies were adjusted in pypsa loader (no release notes, as
this was not released)
* ruff formatting
* fix release notes - add missing release notes from #524
