### 2025-01-15 ddb6680
-start strategy implementation
