### 2025-01-15 0700c96
-move objectives into the respective flexibility functions -move repeating functions into the parent class
