### 2025-01-15 838f901
Adding evaluation data to TensorBoard.
