### 2025-01-15 e1e727b
-add forecaster to hydrogen plant -improve docstrings and arguments -adjust tests
