### 2025-01-15 15e9585
Explicit demand response peak load (#484)
Implementation of Flexibility Measures:
1. Cost-Based Load Shifting: Optimize operations to minimize variable
costs while considering load-shifting capabilities.
2. Congestion Management Flexibility: Adjust operations based on grid
congestion signals to reduce congestion at critical nodes.
3. Peak Load Shifting: Curtail demand during peak load periods to adhere
to capacity constraints and improve grid stability.
4. Renewable Utilisation: Align operational load increases with
renewable energy availability to enhance grid sustainability.

Co-authored-by:
[…]
