+++ assume/units/dsm_load_shift.py
@@ -439,11 +439,11 @@ class DSMFlex:
         }
         df = pd.DataFrame(data)
         df.to_excel(
-            "./examples/outputs/H2_electrolyser_DRI_inflex_6a_optimistic.xlsx",
+            "./examples/outputs/H2_electrolyser_DRI_flex_6a_base.xlsx",
             index=False,
         )
         logger.debug(
-            f"Time series data saved to {"./examples/outputs/H2_electrolyser_DRI_inflex_6a_optimistic.xlsx"}"
+            f"Time series data saved to {"./examples/outputs/H2_electrolyser_DRI_flex_6a_base.xlsx"}"
         )
 
         # Calculate total cost
