+++ README.md
@@ -90,13 +90,13 @@ If you intend to use the reinforcement learning capabilities of ASSUME and train
 
 When running an example with learning capabilities, you can start TensorBoard to observe the learning process. Use the following shell command to start TensorBoard:
 ```shell
-tensorboard --logdir "examples/inputs/path_to_tensorboard_log_dir"
+tensorboard --logdir "examples/inputs/example_name"
 ```
-while replacing `"path_to_tensorboard_log_dir"` with the correct logging path that differs according to the example investigated.
+while replacing `"example_name"` with the correct example name that differs according to the example investigated.
 
 You can then head to `http://localhost:6006/` to view and evaluate the training process parameters.
 
-Please note that TensorBoard should be shut down via `Ctrl + C` every time you want to start a new simulation run in the same folder structure and want to overwrite existing results.
+Please note that TensorBoard should ideally be shut down via `Ctrl + C` every time you want to start a new simulation run in the same folder structure and want to overwrite existing results, as failing to do so may lead to conflicts deleting old logs.
 
 ## Trying out ASSUME and the provided Examples
 
