### 2025-01-15 60a06cf
Refactor TensorBoard query to aggregate metrics by day.
