### 2025-01-16 e1e36f0
-rename max_load_shift to cost_based_shift -fix building tests and add new tests -improve docstrings for dsm class functions -fix building constraints for energy balance -remove unused examples
