+++ examples/examples.py
@@ -95,10 +95,6 @@ available_examples = {
     },
     "large_2019_rl": {"scenario": "example_03a", "study_case": "base_case_2019"},
     "large_2021_rl": {"scenario": "example_03b", "study_case": "base_case_2021"},
-    "experiment": {
-        "scenario": "experiment",
-        "study_case": "eom",
-    },
 }
 
 
@@ -113,7 +109,7 @@ if __name__ == "__main__":
     # select to store the simulation results in a local database or in timescale
     # when using timescale, you need to have docker installed and can access the grafana dashboard
     data_format = "local_db"  # "local_db" or "timescale"
-    example = "experiment"
+    example = "small"
 
     if data_format == "local_db":
         db_uri = "sqlite:///./examples/local_db/assume_db.db"
