+++ assume/units/building.py
@@ -8,6 +8,7 @@ import pandas as pd
 import pyomo.environ as pyo
 
 from assume.common.base import SupportsMinMax
+from assume.common.forecasts import Forecaster
 from assume.common.utils import str_to_bool
 from assume.units.dsm_load_shift import DSMFlex
 
@@ -32,22 +33,28 @@ class Building(DSMFlex, SupportsMinMax):
     the interactions between different energy components, ensuring that energy demands are
     met while adhering to operational constraints.
 
-    Attributes:
+    Args:
         id (str): Unique identifier for the building unit.
         unit_operator (str): Operator managing the building unit.
         bidding_strategies (dict): Strategies used for energy bidding in the market.
-        technology (str): Type of technology the building unit employs.
-        node (str): Network node where the building unit is connected.
-        index (pd.DatetimeIndex): Time index representing the granularity of the data.
-        location (tuple[float, float]): Geographic coordinates (latitude, longitude) of the building.
+        forecaster (Forecaster): A forecaster used to get key variables such as fuel or electricity prices.
         components (dict[str, dict]): Sub-components of the building, such as heat pumps or storage systems.
-        objective (str): Optimization objective, e.g., "min_variable_cost" to minimize operational expenses.
-        flexibility_measure (str): Metric used to assess the building's flexibility, e.g., "max_load_shift".
+        technology (str, optional): Type of technology the building unit employs. Default is "building".
+        objective (str, optional): Optimization objective, e.g., "min_variable_cost" to minimize operational expenses. Default is "min_variable_cost".
+        flexibility_measure (str, optional): Metric used to assess the building's flexibility, e.g., "cost_based_load_shift". Default is "cost_based_load_shift".
+        is_prosumer (str, optional): Indicates whether the building acts as a prosumer (producing and consuming energy). Default is "No".
+        cost_tolerance (float, optional): Maximum allowable cost variation for flexibility measures. Default is 10.
+        node (str, optional): Network node where the building unit is connected. Default is "node0".
+        location (tuple[float, float], optional): Geographic coordinates (latitude, longitude) of the building. Default is (0.0, 0.0).
+        **kwargs: Additional keyword arguments for custom configurations.
+
+    Attributes:
+        required_technologies (list): A list of required technologies for the building unit (empty by default).
+        optional_technologies (list): A list of optional technologies the building unit can incorporate, such as heat pumps and PV plants.
     """
 
-    # List of optional technologies that a building unit can incorporate
+    # List of required and optional technologies for the building unit
     required_technologies = []
-    # List of optional technologies that a building unit can incorporate
     optional_technologies = [
         "heat_pump",
         "boiler",
@@ -62,46 +69,24 @@ class Building(DSMFlex, SupportsMinMax):
         id: str,
         unit_operator: str,
         bidding_strategies: dict,
-        components: dict[str, dict] = None,
+        forecaster: Forecaster,
+        components: dict[str, dict],
         technology: str = "building",
-        objective: str = None,
-        node: str = "node0",
-        location: tuple[float, float] = (0.0, 0.0),
-        index: pd.DatetimeIndex = None,
-        flexibility_measure: str = "",
+        objective: str = "min_variable_cost",
+        flexibility_measure: str = "cost_based_load_shift",
         is_prosumer: str = "No",
         cost_tolerance: float = 10,
+        node: str = "node0",
+        location: tuple[float, float] = (0.0, 0.0),
         **kwargs,
     ):
-        """
-        Initializes a Building instance with specified parameters and sets up the optimization model.
-
-        Args:
-            id (str): Unique identifier for the building unit.
-            unit_operator (str): Operator managing the building unit.
-            index (pd.DatetimeIndex): Time index representing the granularity of the data.
[…]
