### 2025-01-16 eabde86
-move marginal_cost calculation to base DSM class -fix naive DSM strategy
