### 2025-01-16 492cb2e
-adjust release notes
