### 2025-01-16 6e5e0ba
Fix SQL group by.
