### 2025-01-16 9b2640c
develop learning redispatch strategy
