+++ assume/strategies/learning_strategies.py
@@ -1081,14 +1081,14 @@ class StorageRLStrategy(AbstractLearningStrategy):
 
 class RedispatchRLStrategy(RLStrategy):
     def calculate_bids(self, unit, market_config, product_tuples, **kwargs):
-        market_id = market_config.market_id
-
-        if market_id == "EOM":
-            return super().calculate_bids(unit, market_config, product_tuples, **kwargs)
-        elif market_id == "redispatch":
+        if market_config.market_mechanism == "redispatch":
             return self.calculate_redispatch_bids(
                 unit, market_config, product_tuples, **kwargs
             )
+        else:
+            bids = super().calculate_bids(unit, market_config, product_tuples, **kwargs)
+            unit.outputs["eom_bids"].loc[product_tuples[0][0]] = bids[1]["price"]
+            return bids
 
     def calculate_redispatch_bids(
         self,
@@ -1112,23 +1112,23 @@ class RedispatchRLStrategy(RLStrategy):
         """
         start = product_tuples[0][0]
         # end_all = product_tuples[-1][1]
-        previous_power = unit.get_output_before(start)
+        # previous_power = unit.get_output_before(start)
         min_power, max_power = unit.min_power, unit.max_power
 
         bids = []
         for product in product_tuples:
             start = product[0]
             current_power = unit.outputs["energy"].at[start]
-            marginal_cost = unit.calculate_marginal_cost(
-                start, previous_power
-            )  # calculation of the marginal costs
-
+            # marginal_cost = unit.calculate_marginal_cost(
+            #     start, previous_power
+            # )  # calculation of the marginal costs
+            price = unit.outputs["eom_bids"].at[start]
             bids.append(
                 {
                     "start_time": product[0],
                     "end_time": product[1],
                     "only_hours": product[2],
-                    "price": marginal_cost,
+                    "price": price.item(),
                     "volume": current_power,
                     "max_power": max_power,
                     "min_power": min_power,
@@ -1137,3 +1137,155 @@ class RedispatchRLStrategy(RLStrategy):
             )
 
         return bids
+
+    def calculate_reward(self, unit, marketconfig, orderbook):
+        if marketconfig.market_mechanism == "redispatch":
+            return self.calculate_redispatch_reward(unit, marketconfig, orderbook)
+        else:
+            return self.calculate_EOM_profit(unit, marketconfig, orderbook)
+
+    def calculate_redispatch_reward(
+        self,
+        unit,
+        marketconfig: MarketConfig,
+        orderbook: Orderbook,
+    ):
+        revenue = 0
+        costs = 0
+        profit = 0
+
+        reward = 0
+
+        orderbook[0]["accepted_volume"] = -orderbook[0]["volume"]
+        orderbook[0]["accepted_price"] = orderbook[0]["price"]
+
+        # iterate over all orders in the orderbook, to calculate order specific profit
+        for order in orderbook:
+            start = order["start_time"]
+            end = order["end_time"]
+            # end includes the end of the last product, to get the last products' start time we deduct the frequency once
+            end_excl = end - unit.index.freq
+
+            # depending on way the unit calculates marginal costs we take costs
+            marginal_cost = unit.outputs["marginal_cost"].at[start]
+
+            duration = (end - start) / timedelta(hours=1)
+
+            accepted_volume = order.get("accepted_volume", 0)
+            accepted_price = order.get("accepted_price", 0)
+
+            # calculate profit as income - running_cost from this event
+            order_revenue = accepted_price * accepted_volume * duration
+            order_cost = marginal_cost * accepted_volume * duration
+
+            # collect profit and opportunity cost for all orders
+            revenue += order_revenue
+            costs += order_cost
+
+        profit = revenue - costs
+
[…]
