### 2025-01-17 526b678
-remove redundant forecasts loading -remove unused files
