+++ examples/examples.py
@@ -109,6 +109,8 @@ if __name__ == "__main__":
     # select to store the simulation results in a local database or in timescale
     # when using timescale, you need to have docker installed and can access the grafana dashboard
     data_format = "local_db"  # "local_db" or "timescale"
+
+    # select the example to run from the available examples above
     example = "small"
 
     if data_format == "local_db":
