### 2025-01-20 40c8777
Refactoring: renaming learning_params to rl_critic_params.
