+++ pyproject.toml
@@ -51,7 +51,7 @@ learning = [
     "tensorboard >=2.7.0",
 ]
 network = [
-    "pypsa <=0.30.3",
+    "pypsa >=0.32.0",
 ]
 oeds = [
     "demandlib >=0.1.9",
