### 2025-01-20 f024c01
Changes for query to run both on local and on timescale db.
