### 2025-01-21 8460169
-adjust eom bidding startegy -finetune parameters -add check for tensors to redispatch market
