+++ examples/inputs/example_01d/config.yaml
@@ -95,13 +95,13 @@ learning_case:
     trained_policies_save_path: null
     max_bid_price: 100
     algorithm: matd3
-    learning_rate: 0.001
+    learning_rate: 0.0001
     learning_rate_schedule: linear
-    training_episodes: 50
-    episodes_collecting_initial_experience: 5
+    training_episodes: 100
+    episodes_collecting_initial_experience: 2
     train_freq: 24h
     gradient_steps: -1
-    batch_size: 256
+    batch_size: 64
     gamma: 0.99
     device: cpu
     action_noise_schedule: linear
