+++ assume/reinforcement_learning/learning_utils.py
@@ -3,28 +3,17 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import logging
-import os
-
-# Turn off TF onednn optimizations to avoid memory leaks
-os.environ["TF_ENABLE_ONEDNN_OPTS"] = "0"
 
 from collections.abc import Callable
 from datetime import datetime
 from typing import TypedDict
-from sqlalchemy import create_engine
-from torch.utils.tensorboard import SummaryWriter
-from docs.tensorboard_info import tensor_board_intro
-from sqlalchemy.exc import DataError, OperationalError, ProgrammingError
 
 import numpy as np
 import torch as th
-import pandas as pd
 class ObsActRew(TypedDict):
     observation: list[th.Tensor]
     action: list[th.Tensor]
     reward: list[th.Tensor]
-
-logger = logging.getLogger(__name__)
 
 observation_dict = dict[list[datetime], ObsActRew]
 
@@ -163,140 +152,3 @@ def constant_schedule(val: float) -> Schedule:
         return val
 
     return func
-
-class TensorBoardLogger:
-    """
-    Initializes an instance of the TensorBoardLogger class.
-
-    Args:
-    db_uri (str): URI for connecting to the database.
-    simulation_id (str): The unique identifier for the simulation.
-    tensorboard_path (str, optional): Path for storing tensorboard logs.
-    learning_mode (bool, optional): Whether the simulation is in learning mode. Defaults to False.
-    episodes_collecting_initial_experience (int, optional): Number of episodes for initial experience collection. Defaults to 0.
-    perform_evaluation (bool, optional): Whether the simulation is in evaluation mode. Defaults to False.
-    
-    """
-    def __init__(
-        self,
-        db_uri: str,
-        simulation_id: str,
-        tensorboard_path: str = "logs/tensorboard",
-        learning_mode: bool = False,
-        episodes_collecting_initial_experience: int = 0,
-        perform_evaluation: bool = False,
-        ):
-
-        self.simulation_id = simulation_id
-        self.learning_mode = learning_mode
-        self.perform_evaluation = perform_evaluation
-        self.episodes_collecting_initial_experience = episodes_collecting_initial_experience
-        
-        self.writer = SummaryWriter(tensorboard_path)
-        self.db_uri = db_uri
-        if self.db_uri:
-            self.db = create_engine(self.db_uri)
-
-        # get episode number if in learning or evaluation mode
-        self.episode = None
-        if self.learning_mode or self.perform_evaluation:
-            episode = self.simulation_id.split("_")[-1]
-            if episode.isdigit():
-                self.episode = int(episode)
-            
-    def update_tensorboard(self):
-        """Store episodic evaluation data in tensorboard"""
-        if not (self.episode and self.learning_mode):
-            return
-
-        mode = "train" if not self.perform_evaluation else "eval"
-        
-        # columns specific to the training mode.
-        train_columns = """, 
-            AVG(regret) AS regret, 
-            AVG(critic_loss) AS loss, 
-            AVG(learning_rate) AS lr, 
-            AVG(exploration_noise_0) AS noise_0, 
-            AVG(exploration_noise_1) AS noise_1
-            """
-
-        # Use appropriate date function and parameter style based on database type
-        if self.db.dialect.name == 'sqlite':
-            # To aggregate by hour instead of day, replace '%Y-%m-%d' with '%Y-%m-%d %H'
-            # To aggregate by month instead of day, replace '%Y-%m-%d' with '%Y-%m'
-            date_func = f"strftime('%Y-%m-%d', datetime)"
-        elif self.db.dialect.name == 'postgresql':
[…]
