+++ docs/tensorboard_info.py
@@ -1,64 +1,0 @@
-# SPDX-FileCopyrightText: ASSUME Developers
-#
-# SPDX-License-Identifier: AGPL-3.0-or-later
-
-"""Module containing TensorBoard documentation."""
-
-tensor_board_intro = """
-# TensorBoard Guide: Parameter Visualization and Interpretation
-
-Welcome to the TensorBoard visualization interface for the reinforcement learning training process. This guide will help you navigate and interpret the displayed parameters effectively.
-
-## Accessing the Data
-
-To view the training and evaluation results, navigate to the "SCALARS" page in TensorBoard. Here you'll find various metrics tracked during the training process, visualized as interactive plots.
-
-## Available Parameters
-
-The following parameters are being tracked and displayed:
-
-### Training Metrics:
-a) Reward
-b) Profit
-c) Regret
-d) Loss
-e) Learning Rate
-f) Noise Parameters
-
-### Evaluation Metrics:
-a) Reward
-b) Profit
-
-## Visualization Settings
-
-For optimal visualization of the training progress:
-
-- Set the smoothing parameter to 0,999 for all metrics except the learning rate
-- For the learning rate visualization, set smoothing to 0.0 to see the exact values
-- The x-axis represents time in hours, displayed as consecutive integers over the episodes
-- Data display begins after the initial exploration phase, as early results are random due to the exploration nature of the RL algorithm
-
-## Using Regex Filters
-
-To focus on specific metrics or units, use the regex filter option ion the left of the SCALARS page. This allows you to display only the data you're interested in, making it easier to analyze the results.
-You can filter by unit name, display only averaged values or filter different study cases.
-
-## Interactive Features
-
-The TensorBoard interface offers various interactive features to help you analyze the data:
-
-- Zoom functionality for detailed inspection of specific time periods
-- Clickable data points for detailed value inspection
-- Additional data information available in the upper left corner of each plot
-- Customizable display options for better visualization
-
-## Interpreting the Results
-
-To effectively analyze the training progress, focus on the learning trends and performance improvements over time:
-
-- Monitor how reward, regret, and other metrics evolve over time
-- Look for positive trends such as increasing rewards or decreasing regret
-- Check whether the learning rate and noise parameters follow the trend you set in the config file
-
-The data presentation is designed to help you track the algorithm's learning progress and performance improvements over time. Use the interactive features to focus on specific aspects or time periods of interest.
-"""
