### 2025-01-22 17f5e59
inputs and py file for 2-nodes example
