+++ examples/EEM-minimal_example.py
@@ -1,0 +1,151 @@
+'''
+Minimalbeispiel für Assume
+- 2 nodes
+- wind + oil at node 1, gas and coal at node 2
+- hourly EOM with pay as clear
+- redispatch market pay as bid opening subsequent to EOM
+- 2 demand timeseries representing germany divided by 10 demand
+- simulation of 1 month with 1h timesteps
+- study_cases base_case, learning_oil, learning_gas, learning_coal with learning switched on for oil, gas, coal respectively
+'''
+
+#
+#!pip install assume-framework -U
+#
+
+import logging
+import os
+from datetime import datetime, timedelta
+import numpy as np
+import yaml
+
+import pandas as pd
+from dateutil import rrule as rr
+
+from assume import World
+from assume.common.forecasts import NaiveForecast
+from assume.common.market_objects import MarketConfig, MarketProduct
+
+log = logging.getLogger(__name__)
+
+# import the main World class and the load_scenario_folder functions from assume
+from assume import World
+from assume.scenario.loader_csv import load_scenario_folder, run_learning
+from assume.strategies.learning_strategies import RLStrategy
+#from assume.reinforcement_learning.learning_utils import NormalActionNoise
+#from assume.reinforcement_learning.algorithms import actor_architecture_aliases
+
+
+# Define paths for input and output data
+#csv_path = r"C:\Users\Gunter\Code\EEM_paper\outputs"
+#input_path = r"C:\Users\Gunter\Code\EEM_paper\inputs"
+csv_path = "outputs"
+input_path = "inputs"
+scenario = "2_nodes"
+
+for study_case in ["base_case", "learning_oil", "learning_gas", "learning_coal", "learning_all3"]:
+#study_case="base_case"
+#study_case = "learning_coal"
+#study_case = "learning_gas"
+#study_case = "learning_oil"
+#study_case = "learning_all3"
+#study_case = "learning_coal"
+
+    if study_case == "base_case":
+        eom_bidding_list = ["naive_eom"] * 4
+        redispatch_bidding_list = ["naive_redispatch"] * 4
+    elif study_case == "learning_coal":
+        eom_bidding_list = ["naive_eom"] * 3 + ["pp_learning"]
+        redispatch_bidding_list = ["naive_redispatch"] * 3 + ["redispatch_learning"]
+    elif study_case == "learning_gas":
+        eom_bidding_list = ["naive_eom"] * 2 + ["pp_learning"] + ["naive_eom"]
+        redispatch_bidding_list = ["naive_redispatch"] * 2 + ["redispatch_learning"] + ["naive_redispatch"]
+    elif study_case == "learning_oil":
+        eom_bidding_list = ["naive_eom"] + ["pp_learning"] + ["naive_eom"] * 2
+        redispatch_bidding_list = ["naive_redispatch"] + ["redispatch_learning"] + ["naive_redispatch"] * 2
+    elif study_case == "learning_all3":
+        eom_bidding_list = ["naive_eom"] + ["pp_learning"] * 3
+        redispatch_bidding_list = ["naive_redispatch"] + ["redispatch_learning"] * 3
+    elif study_case == "test_redispatch":
+        eom_bidding_list = ["naive_eom"] * 3 + ["pp_learning"]
+        redispatch_bidding_list = ["naive_redispatch"] * 3 + ["redispatch_learning"]
+
+    # as long as two markets learning is not implemented: choose naive redispatch for all
+    #redispatch_bidding_list = ["naive_redispatch"] * 4
+
+    # Create the data
+    powerplant_units_data = {
+        "name": ["Wind", "Oil", "Gas", "Coal"],
+        "technology": ["wind", "oil", "natural gas", "lignite"],
+        "node": ["north", "north", "south", "south"],
[…]
