### 2025-01-22 8239beb
inputs and py file for 2-nodes example
