+++ examples/inputs/example_01d/config.yaml
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: ASSUME Developers
+ # SPDX-FileCopyrightText: ASSUME Developers
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
