### 2025-01-22 aa3c9a8
inputs and py file for 2-nodes example
