### 2025-01-23 d612a4d
-combine improt and export into one unit -define new bidding strategy for exchanges which produces two bids -rename inputs according to new structure -rework tests
