+++ assume/units/__init__.py
@@ -4,7 +4,7 @@
 
 from assume.common.base import BaseUnit
 from assume.units.demand import Demand
-from assume.units.exchanges import Exchanges
+from assume.units.exchange import Exchange
 from assume.units.powerplant import PowerPlant
 from assume.units.storage import Storage
 from assume.units.steel_plant import SteelPlant
@@ -14,7 +14,7 @@ from assume.units.dst_components import demand_side_technologies
 unit_types: dict[str, BaseUnit] = {
     "power_plant": PowerPlant,
     "demand": Demand,
-    "exchanges": Exchanges,
+    "exchange": Exchange,
     "storage": Storage,
     "steel_plant": SteelPlant,
     "hydrogen_plant": HydrogenPlant,
