+++ tests/test_exchanges.py
@@ -5,47 +5,21 @@
 from datetime import datetime, timedelta
 
 import pandas as pd
+import pytest
 from dateutil import rrule as rr
 
 from assume.common.forecasts import NaiveForecast
 from assume.common.market_objects import MarketConfig, MarketProduct
-from assume.strategies import NaiveSingleBidStrategy
-from assume.units.exchanges import Exchanges
+from assume.strategies import NaiveExchangeStrategy
+from assume.units.exchange import Exchange
+
+# === FIXTURES ===
 
 
-def test_exchanges_export():
-    strategies = {"EOM": NaiveSingleBidStrategy()}
-
-    index = pd.date_range(
-        start=datetime(2023, 7, 1),
-        end=datetime(2023, 7, 2),
-        freq="1h",
-    )
-    product_tuple = (
-        datetime(2023, 7, 1, hour=1),
-        datetime(2023, 7, 1, hour=2),
-        None,
-    )
-    forecaster = NaiveForecast(index, exp=100)
-    exch = Exchanges(
-        id="exp",
-        direction="export",
-        unit_operator="eom_de",
-        technology="energy",
-        bidding_strategies=strategies,
-        max_power=1000000,
-        min_power=0,
-        forecaster=forecaster,
-        price=2999,
-    )
-    start = product_tuple[0]
-    end = product_tuple[1]
-    _, max_power = exch.calculate_min_max_power(start, end)
-
-    assert max_power.max() == -100
-    assert exch.calculate_marginal_cost(start, max_power.max()) == 2999
-
-    mc = MarketConfig(
+@pytest.fixture
+def market_config():
+    """Fixture for market configuration."""
+    return MarketConfig(
         market_id="EOM",
         opening_hours=rr.rrule(rr.HOURLY),
         opening_duration=timedelta(hours=1),
@@ -53,66 +27,157 @@ def test_exchanges_export():
         market_products=[MarketProduct(timedelta(hours=1), 1, timedelta(hours=1))],
     )
 
-    bids = exch.calculate_bids(mc, [product_tuple])
+
+@pytest.fixture
+def full_forecast():
+    """Fixture for forecasted import/export values with both import and export data."""
+    index = pd.date_range(start=datetime(2024, 1, 1), periods=24, freq="1h")
+    return NaiveForecast(
+        index, test_unit_import=100, test_unit_export=50
+    )  # Both import and export
+
+
+@pytest.fixture
+def import_only_forecast():
+    """Fixture for forecasted values where only import is available."""
+    index = pd.date_range(start=datetime(2024, 1, 1), periods=24, freq="1h")
+    return NaiveForecast(index, test_unit_import=100)  # Only import, no export
+
+
+@pytest.fixture
+def export_only_forecast():
+    """Fixture for forecasted values where only export is available."""
+    index = pd.date_range(start=datetime(2024, 1, 1), periods=24, freq="1h")
+    return NaiveForecast(index, test_unit_export=50)  # Only export, no import
+
+
+@pytest.fixture
+def exchange_unit(full_forecast):
+    """Fixture for an Exchange unit with full import/export forecast."""
+    return Exchange(
+        id="test_unit",
+        unit_operator="test_operator",
+        bidding_strategies={"EOM": NaiveExchangeStrategy()},
+        forecaster=full_forecast,
+        price_import=10.0,
+        price_export=2000.0,
+    )
+
+
+@pytest.fixture
+def exchange_import_only(import_only_forecast):
+    """Fixture for an Exchange unit that only has import forecasted."""
+    return Exchange(
+        id="test_unit",
+        unit_operator="test_operator",
+        bidding_strategies={"EOM": NaiveExchangeStrategy()},
+        forecaster=import_only_forecast,
[…]
