+++ examples/inputs/example_01a/config.yaml
@@ -6,7 +6,7 @@ base:
   start_date: 2019-01-01 00:00
   end_date: 2019-01-31 00:00
   time_step: 1h
-  exchanges_units: null
+  exchange_units: null
   save_frequency_hours: 720
 
   markets_config:
@@ -30,7 +30,7 @@ base_with_exchanges:
   start_date: 2019-01-01 00:00
   end_date: 2019-01-31 00:00
   time_step: 1h
-  exchanges_units: exchanges_units.csv
+  exchange_units: exchange_units.csv
   save_frequency_hours: 720
 
   markets_config:
@@ -54,7 +54,6 @@ tiny:
   start_date: 2019-01-01 00:00
   end_date: 2019-01-02 00:00
   time_step: 1h
-  exchanges_units: null
   save_frequency_hours: 24
 
   markets_config:
@@ -78,7 +77,7 @@ dam:
   start_date: 2019-01-01 00:00
   end_date: 2019-01-31 00:00
   time_step: 1h
-  exchanges_units: null
+  exchange_units: null
   save_frequency_hours: 720
 
   markets_config:
@@ -102,7 +101,7 @@ dam_with_complex_clearing:
   start_date: 2019-01-01 00:00
   end_date: 2019-04-01 00:00
   time_step: 1h
-  exchanges_units: null
+  exchange_units: null
   save_frequency_hours: 720
 
   markets_config:

[…]
