### 2025-01-23 bee503a
-add release notes -fix readthedocs config
