+++ .readthedocs.yaml
@@ -10,7 +10,7 @@
 version: 2
 sphinx:
   # Path to your Sphinx configuration file.
-  configuration: docs/conf.py
+  configuration: docs/source/conf.py
 
 # Set the version of Python and other tools you might need
 build:
