### 2025-01-23 ee2294f
507 add cross border flows to scenarios (#522)
Add a new type of unit, "Exchanges", corresponding to the commercial exchanges (imports/exports), allowing imports and exports to be added to the market dynamics.

Fix small bugs in the documentation

Co-authored-by: Nick Harder <nick.harder94@gmail.com>
