+++ tests/test_exchanges.py
@@ -1,0 +1,183 @@
+# SPDX-FileCopyrightText: ASSUME Developers
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+
+from datetime import datetime, timedelta
+
+import pandas as pd
+import pytest
+from dateutil import rrule as rr
+
+from assume.common.forecasts import NaiveForecast
+from assume.common.market_objects import MarketConfig, MarketProduct
+from assume.strategies import NaiveExchangeStrategy
+from assume.units.exchange import Exchange
+
+# === FIXTURES ===
+
+
+@pytest.fixture
+def market_config():
+    """Fixture for market configuration."""
+    return MarketConfig(
+        market_id="EOM",
+        opening_hours=rr.rrule(rr.HOURLY),
+        opening_duration=timedelta(hours=1),
+        market_mechanism="not needed",
+        market_products=[MarketProduct(timedelta(hours=1), 1, timedelta(hours=1))],
+    )
+
+
+@pytest.fixture
+def full_forecast():
+    """Fixture for forecasted import/export values with both import and export data."""
+    index = pd.date_range(start=datetime(2024, 1, 1), periods=24, freq="1h")
+    return NaiveForecast(
+        index, test_unit_import=100, test_unit_export=50
+    )  # Both import and export
+
+
+@pytest.fixture
+def import_only_forecast():
+    """Fixture for forecasted values where only import is available."""
+    index = pd.date_range(start=datetime(2024, 1, 1), periods=24, freq="1h")
+    return NaiveForecast(index, test_unit_import=100)  # Only import, no export
+
+
+@pytest.fixture
+def export_only_forecast():
+    """Fixture for forecasted values where only export is available."""
+    index = pd.date_range(start=datetime(2024, 1, 1), periods=24, freq="1h")
+    return NaiveForecast(index, test_unit_export=50)  # Only export, no import
+
+
+@pytest.fixture
+def exchange_unit(full_forecast):
+    """Fixture for an Exchange unit with full import/export forecast."""
+    return Exchange(
+        id="test_unit",
+        unit_operator="test_operator",
+        bidding_strategies={"EOM": NaiveExchangeStrategy()},
+        forecaster=full_forecast,
+        price_import=10.0,
+        price_export=2000.0,
+    )
+
+
+@pytest.fixture
+def exchange_import_only(import_only_forecast):
+    """Fixture for an Exchange unit that only has import forecasted."""
+    return Exchange(
+        id="test_unit",
+        unit_operator="test_operator",
+        bidding_strategies={"EOM": NaiveExchangeStrategy()},
+        forecaster=import_only_forecast,
+        price_import=10.0,
+        price_export=2000.0,
+    )
+
+
+@pytest.fixture
+def exchange_export_only(export_only_forecast):
+    """Fixture for an Exchange unit that only has export forecasted."""
+    return Exchange(
+        id="test_unit",
+        unit_operator="test_operator",
+        bidding_strategies={"EOM": NaiveExchangeStrategy()},
+        forecaster=export_only_forecast,
+        price_import=10.0,
+        price_export=2000.0,
+    )
+
+
+# === TESTS ===
+
+
+def test_exchange_initialization(exchange_unit):
+    """Test if the Exchange unit initializes correctly with both import and export forecast."""
+    assert exchange_unit.id == "test_unit"
+    assert exchange_unit.unit_operator == "test_operator"
+    assert exchange_unit.price_import == 10.0
+    assert exchange_unit.price_export == 2000.0
+
+    # Validate correct forecasting data
+    assert (
+        exchange_unit.volume_import.at["2024-01-01 00:00:00"] == 100
+    ), "Incorrect import volume"
+    assert (
[…]
