+++ assume/scenario/loader_csv.py
@@ -100,6 +100,10 @@ def load_file(
                 return None
 
             if df.index.freq < index.freq:
+                logger.warning(
+                    f"Resolution of {file_name} ({df.index.freq}) is higher than the simulation ({index.freq}). "
+                    "Resampling using mean(). Make sure this is what you want and your data is in units of power."
+                )
                 df = df.resample(index.freq).mean()
                 logger.info(f"Downsampling {file_name} successful.")
 
@@ -456,6 +460,7 @@ def load_config_and_create_forecaster(
     powerplant_units = load_file(path=path, config=config, file_name="powerplant_units")
     storage_units = load_file(path=path, config=config, file_name="storage_units")
     demand_units = load_file(path=path, config=config, file_name="demand_units")
+    exchange_units = load_file(path=path, config=config, file_name="exchange_units")
 
     # Initialize an empty dictionary to combine the DSM units
     dsm_units = {}
@@ -477,8 +482,9 @@ def load_config_and_create_forecaster(
     demand_df = load_file(path=path, config=config, file_name="demand_df", index=index)
     if demand_df is None:
         raise ValueError("No demand time series was provided!")
-    cross_border_flows_df = load_file(
-        path=path, config=config, file_name="cross_border_flows", index=index
+
+    exchanges_df = load_file(
+        path=path, config=config, file_name="exchanges_df", index=index
     )
     availability = load_file(
         path=path, config=config, file_name="availability_df", index=index
@@ -494,17 +500,8 @@ def load_config_and_create_forecaster(
         )
         availability = normalize_availability(powerplant_units, availability)
 
-    electricity_prices_df = load_file(
-        path=path, config=config, file_name="electricity_prices", index=index
-    )
-    price_forecast_df = load_file(
-        path=path, config=config, file_name="price_forecasts", index=index
-    )
     fuel_prices_df = load_file(
         path=path, config=config, file_name="fuel_prices_df", index=index
-    )
-    temperature_df = load_file(
-        path=path, config=config, file_name="temperature", index=index
     )
 
     buses = load_file(path=path, config=config, file_name="buses")
@@ -521,12 +518,9 @@ def load_config_and_create_forecaster(
 
     forecaster.set_forecast(forecasts_df)
     forecaster.set_forecast(demand_df)
-    forecaster.set_forecast(cross_border_flows_df)
+    forecaster.set_forecast(exchanges_df)
     forecaster.set_forecast(availability, prefix="availability_")
-    forecaster.set_forecast(electricity_prices_df)
-    forecaster.set_forecast(price_forecast_df, "price_")
     forecaster.set_forecast(fuel_prices_df, prefix="fuel_price_")
-    forecaster.set_forecast(temperature_df)
     forecaster.calc_forecast_if_needed()
 
     forecaster.convert_forecasts_to_fast_series()
@@ -540,6 +534,7 @@ def load_config_and_create_forecaster(
         "powerplant_units": powerplant_units,
         "storage_units": storage_units,
         "demand_units": demand_units,
+        "exchange_units": exchange_units,
         "dsm_units": dsm_units,
         "forecaster": forecaster,
     }
@@ -582,6 +577,7 @@ def setup_world(
     powerplant_units = scenario_data["powerplant_units"]
     storage_units = scenario_data["storage_units"]
     demand_units = scenario_data["demand_units"]
+    exchange_units = scenario_data["exchange_units"]
     dsm_units = scenario_data["dsm_units"]
     forecaster = scenario_data["forecaster"]
 
@@ -681,7 +677,7 @@ def setup_world(
     logger.info("Read units from file")
 
     units = defaultdict(list)
-    pwp_plants = read_units(
+    powerplant_units = read_units(
[…]

[…]
