### 2025-01-23 8813761
naive strategy fixed
