+++ assume/units/dsm_load_shift.py
@@ -34,20 +34,7 @@ class DSMFlex:
     # Mapping of flexibility measures to their respective functions
     flexibility_map: dict[str, Callable[[pyo.ConcreteModel], None]] = {
         "cost_based_load_shift": lambda self, model: self.cost_based_flexibility(model),
-        "congestion_management_flexibility": lambda self,
-        model: self.grid_congestion_management(model),
-        "peak_load_shifting": lambda self, model: self.peak_load_shifting_flexibility(
-            model
-        ),
-        "renewable_utilisation": lambda self, model: self.renewable_utilisation(
-            model,
-        ),
-    }
-    big_M = 10000000
-
-    # Mapping of flexibility measures to their respective functions
-    flexibility_map: dict[str, Callable[[pyo.ConcreteModel], None]] = {
-        "cost_based_load_shift": lambda self, model: self.cost_based_flexibility(model),
+        "electricity_price_signal_based_flexibility": lambda self, model: self.electricity_price_signal_based_flexibility(model),
         "congestion_management_flexibility": lambda self,
         model: self.grid_congestion_management(model),
         "peak_load_shifting": lambda self, model: self.peak_load_shifting_flexibility(
@@ -185,155 +172,34 @@ class DSMFlex:
                         + self.model.dsm_blocks["dri_plant"].power_in[t]
                     )
             elif self.technology == "cement_plant":
-                if self.has_raw_mill and not (
-                    self.has_clinker_system
-                    or self.has_cement_mill
-                    or self.has_electrolyser
-                    or self.has_ccs_system
-                    or self.has_hydrogen_buffer_storage
-                ):
-                    return (
-                        m.total_power_input[t]
-                        + m.load_shift_pos[t]
-                        - m.load_shift_neg[t]
-                        == self.model.dsm_blocks["raw_material_mill"].power_in[t]
-                    )
-                if self.has_cement_mill and not (
-                    self.has_clinker_system
-                    or self.has_raw_mill
-                    or self.has_electrolyser
-                    or self.has_ccs_system
-                    or self.has_hydrogen_buffer_storage
-                ):
-                    return (
-                        m.total_power_input[t]
-                        + m.load_shift_pos[t]
-                        - m.load_shift_neg[t]
-                        == self.model.dsm_blocks["cement_mill"].power_in[t]
-                    )
-                if (
-                    self.has_clinker_system
-                    and self.has_cement_mill
-                    and not (
-                        self.has_raw_mill
-                        or self.has_electrolyser
-                        or self.has_ccs_system
-                        or self.has_hydrogen_buffer_storage
-                    )
-                ):
-                    return (
-                        m.total_power_input[t]
-                        + m.load_shift_pos[t]
-                        - m.load_shift_neg[t]
-                        == self.model.dsm_blocks["clinker_system"].power_in[t]
-                        + self.model.dsm_blocks["cement_mill"].power_in[t]
-                    )
-                if (
-                    self.has_raw_mill
-                    and self.has_clinker_system
-                    and self.has_cement_mill
-                    and not (
-                        self.has_electrolyser
-                        or self.has_ccs_system
-                        or self.has_hydrogen_buffer_storage
-                    )
-                ):
-                    return (
-                        m.total_power_input[t]
-                        + m.load_shift_pos[t]
-                        - m.load_shift_neg[t]
-                        == self.model.dsm_blocks["clinker_system"].power_in[t]
-                        + self.model.dsm_blocks["cement_mill"].power_in[t]
[…]
