+++ examples/inputs/experiment/config.yaml
@@ -4,7 +4,7 @@
 
 cement_plant_case:
   start_date: 2019-01-01 00:00
-  end_date: 2019-01-01 23:00
+  end_date: 2019-01-03 23:00
   time_step: 1h
   save_frequency_hours: 24
   markets_config:
@@ -24,3 +24,37 @@ cement_plant_case:
       minimum_bid_price: -500
       price_unit: EUR/MWh
       market_mechanism: pay_as_clear
+
+    CRM_pos:
+      operator: CRM_operator
+      product_type: capacity_pos
+      start_date: 2019-01-01 00:00
+      products:
+        - duration: 4h
+          count: 1
+          first_delivery: 1h
+      opening_frequency: 4h
+      opening_duration: 1h
+      volume_unit: MW
+      maximum_bid_volume: 100000
+      maximum_bid_price: 9999
+      minimum_bid_price: -9999
+      price_unit: EUR/MW
+      market_mechanism: pay_as_bid
+
+    CRM_neg:
+      operator: CRM_operator
+      product_type: capacity_neg
+      start_date: 2019-01-01 00:00
+      products:
+        - duration: 4h
+          count: 1
+          first_delivery: 1h
+      opening_frequency: 4h
+      opening_duration: 1h
+      volume_unit: MW
+      maximum_bid_volume: 100000
+      maximum_bid_price: 9999
+      minimum_bid_price: -9999
+      price_unit: EUR/MW
+      market_mechanism: pay_as_bid
