### 2025-01-24 9783ab2
-fix learning dashboard -add extra log during evaluation
