### 2025-01-24 df4ec5b
-update polyak_update function with more efficient function -sample transiitons only once per gradient step -more efficient preparation of all_states and all_next_states -more efficient critic loss calculation -perform polyak update in a batch instead of in a loop for better performance -more efficient actor loss calculation
