+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -426,83 +426,77 @@ class TD3(RLAlgorithm):
         for _ in range(self.gradient_steps):
             self.n_updates += 1
 
+            transitions = self.learning_role.buffer.sample(self.batch_size)
+            states, actions, next_states, rewards = (
+                transitions.observations,
+                transitions.actions,
+                transitions.next_observations,
+                transitions.rewards,
+            )
+
+            with th.no_grad():
+                # Select action according to policy and add clipped noise
+                # Select action according to policy and add clipped noise
+                noise = th.randn_like(actions) * self.target_policy_noise
+                noise = noise.clamp(-self.target_noise_clip, self.target_noise_clip)
+
+                next_actions = th.stack(
+                    [
+                        (
+                            strategy.actor_target(next_states[:, i, :]) + noise[:, i, :]
+                        ).clamp(-1, 1)
+                        for i, strategy in enumerate(
+                            self.learning_role.rl_strats.values()
+                        )
+                    ]
+                )
+
+                next_actions = next_actions.transpose(0, 1).contiguous()
+                next_actions = next_actions.view(-1, n_rl_agents * self.act_dim)
+
+            all_actions = actions.view(self.batch_size, -1)
+
+            # Precompute unique observation parts for all agents
+            unique_obs_from_others = states[
+                :, :, self.obs_dim - self.unique_obs_dim :
+            ].reshape(self.batch_size, n_rl_agents, -1)
+            next_unique_obs_from_others = next_states[
+                :, :, self.obs_dim - self.unique_obs_dim :
+            ].reshape(self.batch_size, n_rl_agents, -1)
+
             for i, u_id in enumerate(self.learning_role.rl_strats.keys()):
                 critic_target = self.learning_role.target_critics[u_id]
                 critic = self.learning_role.critics[u_id]
                 actor = self.learning_role.rl_strats[u_id].actor
-                actor_target = self.learning_role.rl_strats[u_id].actor_target
 
-                # collect new transitions and calculate the next actions only every 100th iteration
-                # to save computation time
-                if i % 100 == 0:
-                    transitions = self.learning_role.buffer.sample(self.batch_size)
-                    states = transitions.observations
-                    actions = transitions.actions
-                    next_states = transitions.next_observations
-                    rewards = transitions.rewards
-
-                    with th.no_grad():
-                        # Select action according to policy and add clipped noise
-                        noise = actions.clone().data.normal_(
-                            0, self.target_policy_noise
-                        )
-                        noise = noise.clamp(
-                            -self.target_noise_clip, self.target_noise_clip
-                        )
-                        next_actions = [
-                            (
-                                self.learning_role.rl_strats[u].actor_target(
-                                    next_states[:, i, :]
-                                )
-                                + noise[:, i, :]
-                            ).clamp(-1, 1)
-                            for i, u in enumerate(self.learning_role.rl_strats.keys())
-                        ]
-                        next_actions = th.stack(next_actions)
-
-                        next_actions = next_actions.transpose(0, 1).contiguous()
-                        next_actions = next_actions.view(-1, n_rl_agents * self.act_dim)
-
-                    all_actions = actions.view(self.batch_size, -1)
-
-                # this takes the unique observations from all other agents assuming that
-                # the unique observations are at the end of the observation vector
-                temp = th.cat(
+                # Efficiently extract unique observations from all other agents
+                other_unique_obs = th.cat(
+                    (unique_obs_from_others[:, :i], unique_obs_from_others[:, i + 1 :]),
+                    dim=1,
+                )
+                other_next_unique_obs = th.cat(
                     (
[…]
