### 2025-01-24 a419fcd
-add gradient clipping to optimizer steps
