### 2025-01-24 4ec58bd
-switch to AdamW optimizer -use improved weight initialization for actor and critic -simplify critic architecture for small number of agents -adjust example_02c config
