+++ assume/reinforcement_learning/neural_network_architecture.py
@@ -29,64 +29,78 @@ class CriticTD3(nn.Module):
         self.obs_dim = obs_dim + unique_obs_dim * (n_agents - 1)
         self.act_dim = act_dim * n_agents
 
-        # Q1 architecture
-        if n_agents <= 50:
-            self.FC1_1 = nn.Linear(self.obs_dim + self.act_dim, 512, dtype=float_type)
-            self.FC1_2 = nn.Linear(512, 256, dtype=float_type)
-            self.FC1_3 = nn.Linear(256, 128, dtype=float_type)
-            self.FC1_4 = nn.Linear(128, 1, dtype=float_type)
+        # Select proper architecture based on `n_agents`
+        if n_agents <= 10:
+            hidden_sizes = [256, 128]  # Shallow network for small `n_agents`
+        elif n_agents <= 50:
+            hidden_sizes = [512, 256, 128]  # Medium network
         else:
-            self.FC1_1 = nn.Linear(self.obs_dim + self.act_dim, 1024, dtype=float_type)
-            self.FC1_2 = nn.Linear(1024, 512, dtype=float_type)
-            self.FC1_3 = nn.Linear(512, 128, dtype=float_type)
-            self.FC1_4 = nn.Linear(128, 1, dtype=float_type)
+            hidden_sizes = [1024, 512, 256, 128]  # Deeper network for large `n_agents`
 
-        # Q2 architecture
-        if n_agents <= 50:
-            self.FC2_1 = nn.Linear(self.obs_dim + self.act_dim, 512, dtype=float_type)
-            self.FC2_2 = nn.Linear(512, 256, dtype=float_type)
-            self.FC2_3 = nn.Linear(256, 128, dtype=float_type)
-            self.FC2_4 = nn.Linear(128, 1, dtype=float_type)
-        else:
-            self.FC2_1 = nn.Linear(self.obs_dim + self.act_dim, 1024, dtype=float_type)
-            self.FC2_2 = nn.Linear(1024, 512, dtype=float_type)
-            self.FC2_3 = nn.Linear(512, 128, dtype=float_type)
-            self.FC2_4 = nn.Linear(128, 1, dtype=float_type)
+        # First Q-network (Q1)
+        self.q1_layers = self._build_q_network(hidden_sizes, float_type)
+
+        # Second Q-network (Q2) for double Q-learning
+        self.q2_layers = self._build_q_network(hidden_sizes, float_type)
+
+        # Initialize weights properly
+        self._init_weights()
+
+    def _build_q_network(self, hidden_sizes, float_type):
+        """
+        Dynamically creates a Q-network given the chosen hidden layer sizes.
+        """
+        layers = nn.ModuleList()
+        input_dim = (
+            self.obs_dim + self.act_dim
+        )  # Input includes all observations and actions
+
+        for h in hidden_sizes:
+            layers.append(nn.Linear(input_dim, h, dtype=float_type))
+            input_dim = h
+        layers.append(nn.Linear(input_dim, 1, dtype=float_type))  # Output Q-value
+
+        return layers
+
+    def _init_weights(self):
+        """Apply Xavier initialization to all layers."""
+
+        def init_layer(m):
+            if isinstance(m, nn.Linear):
+                nn.init.xavier_uniform_(m.weight)
+                nn.init.zeros_(m.bias)
+
+        self.apply(init_layer)
 
     def forward(self, obs, actions):
         """
-        Forward pass through the network, from observation to actions.
+        Forward pass through both Q-networks.
         """
-        xu = th.cat([obs, actions], 1)
+        xu = th.cat([obs, actions], dim=1)  # Concatenate obs & actions
 
[…]
