+++ examples/inputs/example_02c/config.yaml
@@ -7,22 +7,23 @@ base:
   end_date: 2019-04-01 00:00
   time_step: 1h
   save_frequency_hours: null
-  learning_mode: True
+  learning_mode: true
 
   learning_config:
-    continue_learning: False
+    continue_learning: false
     trained_policies_save_path: null
     max_bid_price: 100
     algorithm: matd3
-    actor_architecture: mlp
     learning_rate: 0.001
-    training_episodes: 100
+    learning_rate_schedule: linear
+    training_episodes: 30
     episodes_collecting_initial_experience: 3
-    train_freq: 24h
+    train_freq: 10h
     gradient_steps: 1
     batch_size: 256
     gamma: 0.99
     device: cpu
+    action_noise_schedule: linear
     noise_sigma: 0.1
     noise_scale: 1
     noise_dt: 1
