### 2025-01-27 b5a884b
bidding stategy included
