### 2025-01-27 7078d65
-fix outputs raising an error with float64
