+++ assume/common/outputs.py
@@ -484,7 +484,11 @@ class WriteOutput(Role):
             if df.empty:
                 continue
 
+            # check for tensors and convert them to floats
             df = df.apply(check_for_tensors)
+
+            # check for any float64 columns and convert them to floats
+            df = df.map(lambda x: float(x) if isinstance(x, np.float64) else x)
 
             if self.export_csv_path:
                 data_path = self.export_csv_path / f"{table}.csv"
