### 2025-01-27 ebf620a
-fix critic loss calculation
