### 2025-01-27 8402de8
Ruff linting changes, change learning_operator id, adapt f-string to older python versions.
