+++ assume/scenario/loader_csv.py
@@ -2,9 +2,9 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
-import os
 import copy
 import logging
+import os
 import shutil
 from collections import defaultdict
 from datetime import datetime
