### 2025-01-27 84fa75c
Adapt learning_role test to updated Learning class.
