+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -9,7 +9,6 @@ import torch as th
 from torch.nn import functional as F
 from torch.optim import AdamW
 
-from assume.common.base import LearningStrategy
 from assume.reinforcement_learning.algorithms.base_algorithm import RLAlgorithm
 from assume.reinforcement_learning.learning_utils import polyak_update
 from assume.reinforcement_learning.neural_network_architecture import CriticTD3
@@ -81,13 +80,11 @@ class TD3(RLAlgorithm):
             directory (str): The base directory for saving the parameters.
         """
         os.makedirs(directory, exist_ok=True)
-        for u_id in self.learning_role.rl_strats.keys():
+        for u_id, strategy in self.learning_role.rl_strats.items():
             obj = {
-                "critic": self.learning_role.critics[u_id].state_dict(),
-                "critic_target": self.learning_role.target_critics[u_id].state_dict(),
-                "critic_optimizer": self.learning_role.critics[
-                    u_id
-                ].optimizer.state_dict(),
+                "critic": strategy.critics.state_dict(),
+                "critic_target": strategy.target_critics.state_dict(),
+                "critic_optimizer": strategy.critics.optimizer.state_dict(),
             }
             path = f"{directory}/critic_{u_id}.pt"
             th.save(obj, path)
@@ -104,15 +101,11 @@ class TD3(RLAlgorithm):
             directory (str): The base directory for saving the parameters.
         """
         os.makedirs(directory, exist_ok=True)
-        for u_id in self.learning_role.rl_strats.keys():
+        for u_id, strategy in self.learning_role.rl_strats.items():
             obj = {
-                "actor": self.learning_role.rl_strats[u_id].actor.state_dict(),
-                "actor_target": self.learning_role.rl_strats[
-                    u_id
-                ].actor_target.state_dict(),
-                "actor_optimizer": self.learning_role.rl_strats[
-                    u_id
-                ].actor.optimizer.state_dict(),
+                "actor": strategy.actor.state_dict(),
+                "actor_target": strategy.actor_target.state_dict(),
+                "actor_optimizer": strategy.actor.optimizer.state_dict(),
             }
             path = f"{directory}/actor_{u_id}.pt"
             th.save(obj, path)
@@ -149,18 +142,14 @@ class TD3(RLAlgorithm):
             )
             return
 
-        for u_id in self.learning_role.rl_strats.keys():
+        for u_id, strategy in self.learning_role.rl_strats.items():
             try:
                 critic_params = self.load_obj(
                     directory=f"{directory}/critics/critic_{str(u_id)}.pt"
                 )
-                self.learning_role.critics[u_id].load_state_dict(
-                    critic_params["critic"]
-                )
-                self.learning_role.target_critics[u_id].load_state_dict(
-                    critic_params["critic_target"]
-                )
-                self.learning_role.critics[u_id].optimizer.load_state_dict(
+                strategy.critics.load_state_dict(critic_params["critic"])
+                strategy.target_critics.load_state_dict(critic_params["critic_target"])
+                strategy.critics.optimizer.load_state_dict(
                     critic_params["critic_optimizer"]
                 )
             except Exception:
@@ -184,18 +173,14 @@ class TD3(RLAlgorithm):
             )
             return
 
-        for u_id in self.learning_role.rl_strats.keys():
+        for u_id, strategy in self.learning_role.rl_strats.items():
             try:
                 actor_params = self.load_obj(
                     directory=f"{directory}/actors/actor_{str(u_id)}.pt"
                 )
-                self.learning_role.rl_strats[u_id].actor.load_state_dict(
-                    actor_params["actor"]
-                )
[…]
