### 2025-01-27 af291e4
-reverse examples.py
