### 2025-01-27 6676a4a
-convert tensors to floats when submitting bids to improve performance -improve function to convert from tensors
