+++ assume/common/outputs.py
@@ -22,7 +22,7 @@ from sqlalchemy.exc import DataError, OperationalError, ProgrammingError
 from assume.common.market_objects import MetaDict
 from assume.common.utils import (
     calculate_content_size,
-    check_for_tensors,
+    convert_tensors,
     separate_orders,
 )
 
@@ -485,7 +485,7 @@ class WriteOutput(Role):
                 continue
 
             # check for tensors and convert them to floats
-            df = df.apply(check_for_tensors)
+            df = df.apply(convert_tensors)
 
             # check for any float64 columns and convert them to floats
             df = df.map(lambda x: float(x) if isinstance(x, np.float64) else x)
