### 2025-01-27 5fc11fc
-optimize get_action and get_observation functions -remove unrequired clone and detach -create tensors directly on device instead of moving them
