+++ assume/strategies/learning_strategies.py
@@ -94,7 +94,7 @@ class RLStrategy(AbstractLearningStrategy):
       unit's operational capacity and production costs, helping the agent determine the optimal bid.
 
     Actions are formulated as two values, representing bid prices for both the inflexible and flexible
-    components of the unit's capacity. Actions are scaled from a clamped range of [-1, 1]
+    components of the unit's capacity. Actions are scaled from a range of [-1, 1]
     to real bid prices in the `calculate_bids` method, where they translate into specific bid volumes
     for the inflexible (P_min) and flexible (P_max - P_min) components.
 
@@ -276,12 +276,12 @@ class RLStrategy(AbstractLearningStrategy):
         # 3.1 formulate the bids for Pmin
         # Pmin, the minimum run capacity is the inflexible part of the bid, which should always be accepted
         bid_quantity_inflex = min_power
-        bid_price_inflex = min(bid_prices)
+        bid_price_inflex = th.min(bid_prices)
 
         # 3.1 formulate the bids for Pmax - Pmin
         # Pmin, the minimum run capacity is the inflexible part of the bid, which should always be accepted
         bid_quantity_flex = max_power - bid_quantity_inflex
-        bid_price_flex = max(bid_prices)
+        bid_price_flex = th.max(bid_prices)
 
         # actually formulate bids in orderbook format
         bids = [
@@ -331,7 +331,7 @@ class RLStrategy(AbstractLearningStrategy):
         -----
         If the agent is in learning mode, noise is added to encourage exploration. In initial exploration,
         actions are derived from noise and the marginal cost to explore the action space around the marginal cost. When not in learning mode,
-        actions are generated by the actor network without added noise. Actions are clamped to [-1, 1].
+        actions are generated by the actor network without added noise.
         """
 
         # distinction whether we are in learning mode or not to handle exploration realised with noise
@@ -340,12 +340,12 @@ class RLStrategy(AbstractLearningStrategy):
             # to get a good initial experience, in the area around the costs of the agent
             if self.collect_initial_experience_mode:
                 # define current action as solely noise
-                noise = (
-                    th.normal(
-                        mean=0.0, std=0.2, size=(1, self.act_dim), dtype=self.float_type
-                    )
-                    .to(self.device)
-                    .squeeze()
+                noise = th.normal(
+                    mean=0.0,
+                    std=0.2,
+                    size=(self.act_dim,),
+                    dtype=self.float_type,
+                    device=self.device,
                 )
 
                 # =============================================================================
@@ -356,24 +356,23 @@ class RLStrategy(AbstractLearningStrategy):
                 # add noise to the last dimension of the observation
                 # needs to be adjusted if observation space is changed, because only makes sense
                 # if the last dimension of the observation space are the marginal cost
-                curr_action = noise + base_bid.clone().detach()
+                curr_action = noise + base_bid
 
             else:
                 # if we are not in the initial exploration phase we choose the action with the actor neural net
                 # and add noise to the action
                 curr_action = self.actor(next_observation).detach()
-                noise = th.tensor(
-                    self.action_noise.noise(), device=self.device, dtype=self.float_type
+                noise = self.action_noise.noise(
+                    device=self.device, dtype=self.float_type
                 )
+
                 curr_action += noise
         else:
             # if we are not in learning mode we just use the actor neural net to get the action without adding noise
             curr_action = self.actor(next_observation).detach()
 
             # noise is an tensor with zeros, because we are not in learning mode
-            noise = th.zeros(self.act_dim, dtype=self.float_type)
-
-        curr_action = curr_action.clamp(-1, 1)
+            noise = th.zeros_like(curr_action, dtype=self.float_type)
 
         return curr_action, noise
 
[…]
