### 2025-01-27 c9eae2a
Added test for tensorboard_logger.
