+++ assume/reinforcement_learning/tensorboard_logger.py
@@ -75,6 +75,7 @@ To effectively analyze the training progress, focus on the learning trends and p
 The data presentation is designed to help you track the algorithm's learning progress and performance improvements over time. Use the interactive features to focus on specific aspects or time periods of interest.
 """
 
+
 class TensorBoardLogger:
     """
     Initializes an instance of the TensorBoardLogger class.
@@ -86,8 +87,9 @@ class TensorBoardLogger:
     learning_mode (bool, optional): Whether the simulation is in learning mode. Defaults to False.
     episodes_collecting_initial_experience (int, optional): Number of episodes for initial experience collection. Defaults to 0.
     perform_evaluation (bool, optional): Whether the simulation is in evaluation mode. Defaults to False.
-    
+
     """
+
     def __init__(
         self,
         db_uri: str,
@@ -96,13 +98,14 @@ class TensorBoardLogger:
         learning_mode: bool = False,
         episodes_collecting_initial_experience: int = 0,
         perform_evaluation: bool = False,
-        ):
-
+    ):
         self.simulation_id = simulation_id
         self.learning_mode = learning_mode
         self.perform_evaluation = perform_evaluation
-        self.episodes_collecting_initial_experience = episodes_collecting_initial_experience
-        
+        self.episodes_collecting_initial_experience = (
+            episodes_collecting_initial_experience
+        )
+
         self.writer = SummaryWriter(tensorboard_path)
         self.db_uri = db_uri
         if self.db_uri:
@@ -114,35 +117,35 @@ class TensorBoardLogger:
             episode = self.simulation_id.split("_")[-1]
             if episode.isdigit():
                 self.episode = int(episode)
-            
+
     def update_tensorboard(self):
         """Store episodic evaluation data in tensorboard"""
         if not (self.episode and self.learning_mode):
             return
 
         mode = "train" if not self.perform_evaluation else "eval"
-        
+
         # columns specific to the training mode.
-        train_columns = """, 
-            AVG(regret) AS regret, 
-            AVG(critic_loss) AS loss, 
-            AVG(learning_rate) AS lr, 
-            AVG(exploration_noise_0) AS noise_0, 
+        train_columns = """,
+            AVG(regret) AS regret,
+            AVG(critic_loss) AS loss,
+            AVG(learning_rate) AS lr,
+            AVG(exploration_noise_0) AS noise_0,
             AVG(exploration_noise_1) AS noise_1
             """
 
         # Use appropriate date function and parameter style based on database type
-        if self.db.dialect.name == 'sqlite':
+        if self.db.dialect.name == "sqlite":
             # To aggregate by hour instead of day, replace '%Y-%m-%d' with '%Y-%m-%d %H'
             # To aggregate by month instead of day, replace '%Y-%m-%d' with '%Y-%m'
             date_func = "strftime('%Y-%m-%d', datetime)"
-        elif self.db.dialect.name == 'postgresql':
+        elif self.db.dialect.name == "postgresql":
             # To aggregate by hour instead of day, replace 'YYYY-MM-DD' with 'YYYY-MM-DD HH24'
             # To aggregate by month instead of day, replace 'YYYY-MM-DD' with 'YYYY-MM'
             date_func = "TO_CHAR(datetime, 'YYYY-MM-DD')"
 
         query = f"""
-            SELECT 
+            SELECT
                 {date_func} AS dt,
                 unit,
                 AVG(profit) AS profit,
@@ -175,35 +178,38 @@ class TensorBoardLogger:
             # Process metrics for each timestamp
             for i, time in enumerate(datetimes):
                 time_df = df[df["dt"] == time]
-                
+
                 # Build metrics dictionary
[…]
