### 2025-01-27 4043b7a
use case verändert, preprocessing und config angepasst
