+++ examples/EEM-minimal_example.py
@@ -43,7 +43,14 @@ csv_path = "outputs"
 input_path = "inputs"
 scenario = "2_nodes"
 
-for study_case in ["learning_coal"]:# ["base_case", "learning_oil", "learning_gas", "learning_coal", "learning_all3"]:
+# powerplant units according to example in Hirth and Schlecht (2024)
+num_wind = 1
+num_diesel = 5
+num_gas = 25
+num_coal = 20
+num_all = num_wind + num_diesel + num_gas + num_coal
+
+for study_case in ["base_case", "learning_diesel", "learning_gas", "learning_coal", "learning_all3"]:
 #study_case="base_case"
 #study_case = "learning_coal"
 #study_case = "learning_gas"
@@ -52,41 +59,39 @@ for study_case in ["learning_coal"]:# ["base_case", "learning_oil", "learning_ga
 #study_case = "learning_coal"
 
     if study_case == "base_case":
-        eom_bidding_list = ["naive_eom"] * 4
-        redispatch_bidding_list = ["naive_redispatch"] * 4
+        eom_bidding_list = ["naive_eom"] * num_all
+        redispatch_bidding_list = ["naive_redispatch"] * num_all
     elif study_case == "learning_coal":
-        eom_bidding_list = ["naive_eom"] * 3 + ["redispatch_learning"]
-        redispatch_bidding_list = ["naive_redispatch"] * 3 + ["redispatch_learning"]
+        eom_bidding_list = ["naive_eom"] * (num_all - num_coal) + ["redispatch_learning"] * num_coal
+        redispatch_bidding_list = ["naive_redispatch"] * (num_all - num_coal) + ["redispatch_learning"] * num_coal
     elif study_case == "learning_gas":
-        eom_bidding_list = ["naive_eom"] * 2 + ["redispatch_learning"] + ["naive_eom"]
-        redispatch_bidding_list = ["naive_redispatch"] * 2 + ["redispatch_learning"] + ["naive_redispatch"]
-    elif study_case == "learning_oil":
-        eom_bidding_list = ["naive_eom"] + ["redispatch_learning"] + ["naive_eom"] * 2
-        redispatch_bidding_list = ["naive_redispatch"] + ["redispatch_learning"] + ["naive_redispatch"] * 2
+        eom_bidding_list = ["naive_eom"] * (num_wind + num_diesel) + ["redispatch_learning"] * num_gas + ["naive_eom"] * num_coal
+        redispatch_bidding_list = ["naive_redispatch"] * (num_wind + num_diesel) + ["redispatch_learning"] * num_gas + ["naive_redispatch"] * num_coal
+    elif study_case == "learning_diesel":
+        eom_bidding_list = ["naive_eom"] * num_wind + ["redispatch_learning"] * num_diesel + ["naive_eom"] * (num_gas + num_coal)
+        redispatch_bidding_list = ["naive_redispatch"] * num_wind + ["redispatch_learning"] * num_diesel + ["naive_redispatch"] * (num_gas + num_coal)
     elif study_case == "learning_all3":
-        eom_bidding_list = ["naive_eom"] + ["redispatch_learning"] * 3
-        redispatch_bidding_list = ["naive_redispatch"] + ["redispatch_learning"] * 3
-    elif study_case == "test_redispatch":
-        eom_bidding_list = ["naive_eom"] * 3 + ["pp_learning"]
-        redispatch_bidding_list = ["naive_redispatch"] * 3 + ["redispatch_learning"]
+        eom_bidding_list = ["naive_eom"] * num_wind + ["redispatch_learning"] * (num_all - num_wind)
+        redispatch_bidding_list = ["naive_redispatch"] * num_wind + ["redispatch_learning"] * (num_all - num_wind)
[…]
