+++ examples/inputs/2_nodes/config.yaml
@@ -1,6 +1,6 @@
 base_case:
   start_date: 2019-01-01 00:00
-  end_date: 2019-01-31 23:45
+  end_date: 2019-01-06 00:00
   time_step: 1h
   learning_mode: False
   save_frequency_hours: False
@@ -208,7 +208,7 @@ learning_gas:
         payment_mechanism: pay_as_bid
         backup_marginal_cost: 10000
 
-learning_oil:
+learning_diesel:
   start_date: 2019-01-01 00:00
   end_date: 2019-01-06 00:00
   time_step: 1h
