### 2025-01-28 e5abfe9
-add historical prices to the observations -reduce foresight from 24 to 12 hours
