+++ assume/strategies/learning_strategies.py
@@ -144,7 +144,7 @@ class RLStrategy(AbstractLearningStrategy):
     """
 
     def __init__(self, *args, **kwargs):
-        super().__init__(obs_dim=50, act_dim=2, unique_obs_dim=2, *args, **kwargs)
+        super().__init__(obs_dim=38, act_dim=2, unique_obs_dim=2, *args, **kwargs)
 
         self.unit_id = kwargs["unit_id"]
 
@@ -183,7 +183,7 @@ class RLStrategy(AbstractLearningStrategy):
         # neural network architecture is predefined, and the size of the observations must remain consistent.
         # If you wish to modify the foresight length, remember to also update the 'obs_dim' parameter above,
         # as the observation dimension depends on the foresight value.
-        self.foresight = 24
+        self.foresight = 12
 
         # define allowed order types
         self.order_types = kwargs.get("order_types", ["SB"])
@@ -456,6 +456,25 @@ class RLStrategy(AbstractLearningStrategy):
                 start : start + forecast_len
             ]
 
+        # collect historical past market clearing prices
+        actual_price = unit.outputs["energy_accepted_price"]
+        if start - forecast_len < actual_price.index[0]:
+            # Not enough historical data, use available actual prices and prepend forecasted values for missing past data
+            actual_price_history = actual_price.loc[:start] / self.max_bid_price
+            missing_values = self.foresight - len(actual_price_history)
+
+            if missing_values > 0:
+                forecasted_prices = self.scaled_prices_obs.iloc[:missing_values]
+                actual_price_history = np.concatenate(
+                    [forecasted_prices, actual_price_history]
+                )
+
+        else:
+            # Sufficient historical data exists, collect past actual prices
+            actual_price_history = (
+                actual_price.loc[start - forecast_len : start] / self.max_bid_price
+            )
+
         # get last accepted bid volume and the current marginal costs of the unit
         current_volume = unit.get_output_before(start)
         current_costs = unit.calculate_marginal_cost(start, current_volume)
@@ -472,6 +491,7 @@ class RLStrategy(AbstractLearningStrategy):
             [
                 scaled_res_load_forecast,
                 scaled_price_forecast,
+                actual_price_history,
                 np.array([scaled_total_dispatch, scaled_marginal_cost]),
             ]
         )
