### 2025-01-28 0d79dff
markets: if a markets first schedule is after the end time, do not open at all
