### 2025-01-28 27dee19
-remove option for -1 for gradient steps as it causes confusion
