### 2025-01-28 462d731
-fix initial exploration count
