+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -32,7 +32,7 @@ class TD3(RLAlgorithm):
         self,
         learning_role,
         learning_rate=1e-4,
-        episodes_collecting_initial_experience=100,
+        episodes_collecting_initial_experience=10,
         batch_size=1024,
         tau=0.005,
         gamma=0.99,
