### 2025-01-28 d198323
-fix continue learning parameter
