### 2025-01-29 af7f5b6
-fix redispatch learning
