+++ examples/EEM-minimal_example.py
@@ -1,4 +1,8 @@
-'''
+# SPDX-FileCopyrightText: ASSUME Developers
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+
+"""
 Minimalbeispiel für Assume
 - 2 nodes
 - wind + oil at node 1, gas and coal at node 2
@@ -7,38 +11,27 @@ Minimalbeispiel für Assume
 - 2 demand timeseries representing germany divided by 10 demand
 - simulation of 1 month with 1h timesteps
 - study_cases base_case, learning_oil, learning_gas, learning_coal with learning switched on for oil, gas, coal respectively
-'''
+"""
 
 #
 #!pip install assume-framework -U
 #
 
 import logging
-import os
-from datetime import datetime, timedelta
-import numpy as np
-import yaml
 
 import pandas as pd
-from dateutil import rrule as rr
 
 from assume import World
-from assume.common.forecasts import NaiveForecast
-from assume.common.market_objects import MarketConfig, MarketProduct
+
+# import the main World class and the load_scenario_folder functions from assume
+from assume.scenario.loader_csv import load_scenario_folder, run_learning
+from assume.strategies.learning_strategies import RLStrategy
 
 log = logging.getLogger(__name__)
 
-# import the main World class and the load_scenario_folder functions from assume
-from assume import World
-from assume.scenario.loader_csv import load_scenario_folder, run_learning
-from assume.strategies.learning_strategies import RLStrategy
-#from assume.reinforcement_learning.learning_utils import NormalActionNoise
-#from assume.reinforcement_learning.algorithms import actor_architecture_aliases
-
-
 # Define paths for input and output data
-#csv_path = r"C:\Users\Gunter\Code\EEM_paper\outputs"
-#input_path = r"C:\Users\Gunter\Code\EEM_paper\inputs"
+# csv_path = r"C:\Users\Gunter\Code\EEM_paper\outputs"
+# input_path = r"C:\Users\Gunter\Code\EEM_paper\inputs"
 csv_path = "outputs"
 input_path = "inputs"
 scenario = "2_nodes"
@@ -50,29 +43,59 @@ num_gas = 25
 num_coal = 20
 num_all = num_wind + num_diesel + num_gas + num_coal
 
-for study_case in ["base_case", "learning_diesel", "learning_gas", "learning_coal", "learning_all3"]:
-#study_case="base_case"
-#study_case = "learning_coal"
-#study_case = "learning_gas"
-#study_case = "learning_oil"
-#study_case = "learning_all3"
-#study_case = "learning_coal"
+for study_case in [
+    "base_case",
+    "learning_diesel",
+    "learning_gas",
+    "learning_coal",
+    "learning_all3",
+]:
+    # study_case="base_case"
+    # study_case = "learning_coal"
+    # study_case = "learning_gas"
+    # study_case = "learning_oil"
+    # study_case = "learning_all3"
+    # study_case = "learning_coal"
 
     if study_case == "base_case":
         eom_bidding_list = ["naive_eom"] * num_all
         redispatch_bidding_list = ["naive_redispatch"] * num_all
     elif study_case == "learning_coal":
-        eom_bidding_list = ["naive_eom"] * (num_all - num_coal) + ["redispatch_learning"] * num_coal
-        redispatch_bidding_list = ["naive_redispatch"] * (num_all - num_coal) + ["redispatch_learning"] * num_coal
+        eom_bidding_list = ["naive_eom"] * (num_all - num_coal) + [
+            "redispatch_learning"
+        ] * num_coal
[…]
