### 2025-01-29 f738e77
-restore missing lines -improve error handling
