### 2025-01-31 53fd71b
-improve readme -add comments to example.py -fix config for 2019 large simulation -add check for duplicate names
