### 2025-02-02 977f112
update support policies
this makes it possible to correctly calculate the profits from having PPA, CFD, FIT and MPFIX available
