### 2025-02-02 972c02a
improve oeds loader
 * update fuel costs
 * increase span width of random
 * respect limited availability
 * add option to calculate demand from actual entsoe-load
