### 2025-02-03 c32dc4d
Added logging of gradient clipping.
