+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -380,9 +380,9 @@ class TD3(RLAlgorithm):
             This function implements the TD3 algorithm's key step for policy improvement and exploration.
 
         Returns:
-            tuple: (learning_rate, critic_losses) where:
+            tuple: (learning_rate, unit_params) where:
                 - learning_rate (float): The current learning rate
-                - critic_losses (list[dict]): Critic losses for each agent in each gradient step
+                - unit_params (list[dict]): Critic losses for each agent in each gradient step
         """
 
         logger.debug("Updating Policy")
@@ -390,8 +390,11 @@ class TD3(RLAlgorithm):
         # Stack strategies for easier access
         strategies = list(self.learning_role.rl_strats.values())
         n_rl_agents = len(strategies)
-        critic_losses = [
-            {u_id: None for u_id in self.learning_role.rl_strats.keys()}
+        unit_params = [
+            {
+                u_id: {"loss": None, "total_grad_norm": None, "max_grad_norm": None}
+                for u_id in self.learning_role.rl_strats.keys()
+            }
             for _ in range(self.gradient_steps)
         ]
 
@@ -518,7 +521,7 @@ class TD3(RLAlgorithm):
                 )
 
                 # Store the critic loss for this unit ID
-                critic_losses[step][strategy.unit_id] = critic_loss.item()
+                unit_params[step][strategy.unit_id]["loss"] = critic_loss.item()
                 total_critic_loss += critic_loss
 
             # Single backward pass for all agents' critics
@@ -526,8 +529,18 @@ class TD3(RLAlgorithm):
 
             # Clip the gradients and step each critic optimizer
             for strategy in strategies:
-                th.nn.utils.clip_grad_norm_(strategy.critics.parameters(), max_norm=1.0)
+                parameters = list(strategy.critics.parameters())
+
+                # Determine clipping statistics
+                max_grad_norm = max(p.grad.norm() for p in parameters)
+
+                # Perform clipping
+                total_norm = th.nn.utils.clip_grad_norm_(parameters, max_norm=1.0)
                 strategy.critics.optimizer.step()
+
+                # Store clipping statistics
+                unit_params[step][strategy.unit_id]["total_grad_norm"] = total_norm
+                unit_params[step][strategy.unit_id]["max_grad_norm"] = max_grad_norm
 
             ######################################################################
             # ACTOR UPDATE (DELAYED): Accumulate losses for all agents in one pass
@@ -609,4 +622,4 @@ class TD3(RLAlgorithm):
                 polyak_update(all_critic_params, all_target_critic_params, self.tau)
                 polyak_update(all_actor_params, all_target_actor_params, self.tau)
 
-        return learning_rate, critic_losses
+        return learning_rate, unit_params
