+++ assume/scenario/oeds/static.py
@@ -55,31 +55,31 @@ mastr_solar_azimuth = {
 }
 
 mastr_storage = {
-    "Niederwartha": "591",
-    "Bleiloch": "753",
-    "Hohenwarte 1": "795",
-    "Hohenwarte 2": "2087",
-    "Wendefurth": "532",
-    "Markersbach": "4018",
-    "Geesthacht": "600",
-    "Waldeck 1": "478",
-    "Waldeck 2": "3428",
-    "Bringhausen": "0",
-    "Hemfurth": "0",
-    "illwerke": "0",
-    "Glems": "560",
-    "Schwarzenbach": "198",
-    "Witznau": "220",
-    "Säckingen": "2064",
-    "Häusern": "463",
-    "Waldshut": "402",
-    "Wehr": "6073",
-    "Walchenseekraftwerk": "0",
-    "Happurg": "900",
-    "Schnitzel": "0",
-    "Langenprozelten": "950",
-    "E2307101": "0",
-    "Goldisthal": "8480",
+    "Niederwartha": 591,
+    "Bleiloch": 753,
+    "Hohenwarte 1": 795,
+    "Hohenwarte 2": 2087,
+    "Wendefurth": 532,
+    "Markersbach": 4018,
+    "Geesthacht": 600,
+    "Waldeck 1": 478,
+    "Waldeck 2": 3428,
+    "Bringhausen": 0,
+    "Hemfurth": 0,
+    "illwerke": 0,
+    "Glems": 560,
+    "Schwarzenbach": 198,
+    "Witznau": 220,
+    "Säckingen": 2064,
+    "Häusern": 463,
+    "Waldshut": 402,
+    "Wehr": 6073,
+    "Walchenseekraftwerk": 0,
+    "Happurg": 900,
+    "Schnitzel": 0,
+    "Langenprozelten": 950,
+    "E2307101": 0,
+    "Goldisthal": 8480,
 }
 
 # start cost given in [€/MW]
