### 2025-02-03 d188c77
update support policies - various fixes - performatives (#542)
This makes it possible to correctly calculate the profits from having PPA, CFD, FIT, and MPFIX available.

There were some inaccuracies in the previous calculation:
1. The contractRole had the same market bugs as those solved in #466.
2. The contract works much better now that the following issues are fixed:
   - When contracting the sum, a cashflow is already calculated. To avoid impacting the result, the contract_price is stored separately, while the contract itself has a cost of 0.
   - The execution of the contract was happenin
[…]
