### 2025-02-03 778e138
update release notes from merged PRs
