+++ examples/inputs/example_02c/config.yaml
@@ -17,7 +17,7 @@ base:
     learning_rate: 0.001
     learning_rate_schedule: linear
     training_episodes: 100
-    episodes_collecting_initial_experience: 3
+    episodes_collecting_initial_experience: 10
     train_freq: 10h
     gradient_steps: 1
     batch_size: 256
