+++ examples/inputs/example_03a/config.yaml
@@ -7,7 +7,7 @@ base_case_2019:
   end_date: 2019-03-31 00:00
   time_step: 1h
   learning_mode: true
-  save_frequency_hours: Null
+  save_frequency_hours: null
 
   learning_config:
     continue_learning: false
