+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -4,6 +4,7 @@
 
 import logging
 import os
+import random
 
 import torch as th
 from torch.nn import functional as F
@@ -386,6 +387,9 @@ class TD3(RLAlgorithm):
         strategies = list(self.learning_role.rl_strats.values())
         n_rl_agents = len(strategies)
 
+        # Randomly select a subset of strategies to update
+        selected_strategies = random.sample(strategies, k=int(n_rl_agents / 3))
+
         # update noise decay and learning rate
         updated_noise_decay = self.learning_role.calc_noise_from_progress(
             self.learning_role.get_progress_remaining()
@@ -396,7 +400,7 @@ class TD3(RLAlgorithm):
         )
 
         # loop over all units to avoid update call for every gradient step, as it will be ambiguous
-        for strategy in strategies:
+        for strategy in selected_strategies:
             self.update_learning_rate(
                 [
                     strategy.critics.optimizer,
@@ -449,13 +453,16 @@ class TD3(RLAlgorithm):
             #####################################################################
 
             # Zero-grad for all critics before accumulation
-            for strategy in strategies:
+            for strategy in selected_strategies:
                 strategy.critics.optimizer.zero_grad(set_to_none=True)
 
             total_critic_loss = 0.0
 
             # Loop over all agents and accumulate critic loss
             for i, strategy in enumerate(strategies):
+                if strategy not in selected_strategies:
+                    continue
+
                 actor = strategy.actor
                 critic = strategy.critics
                 critic_target = strategy.target_critics
@@ -513,7 +520,7 @@ class TD3(RLAlgorithm):
             total_critic_loss.backward()
 
             # Clip the gradients and step each critic optimizer
-            for strategy in strategies:
+            for strategy in selected_strategies:
                 th.nn.utils.clip_grad_norm_(strategy.critics.parameters(), max_norm=1.0)
                 strategy.critics.optimizer.step()
 
@@ -522,13 +529,16 @@ class TD3(RLAlgorithm):
             ######################################################################
             if self.n_updates % self.policy_delay == 0:
                 # Zero-grad for all actors first
-                for strategy in strategies:
+                for strategy in selected_strategies:
                     strategy.actor.optimizer.zero_grad(set_to_none=True)
 
                 total_actor_loss = 0.0
 
                 # We'll compute each agent's actor loss, accumulate, then do one backprop
-                for i, strategy in enumerate(strategies):
+                for i, strategy in enumerate(selected_strategies):
+                    if strategy not in selected_strategies:
+                        continue
+
                     actor = strategy.actor
                     critic = strategy.critics
 
@@ -571,7 +581,7 @@ class TD3(RLAlgorithm):
                 total_actor_loss.backward()
 
                 # Clip and step each actor optimizer
-                for strategy in strategies:
+                for strategy in selected_strategies:
                     th.nn.utils.clip_grad_norm_(
                         strategy.actor.parameters(), max_norm=1.0
                     )
@@ -584,7 +594,7 @@ class TD3(RLAlgorithm):
                 all_actor_params = []
                 all_target_actor_params = []
 
-                for strategy in strategies:
+                for strategy in selected_strategies:
                     all_critic_params.extend(strategy.critics.parameters())
                     all_target_critic_params.extend(
                         strategy.target_critics.parameters()
