### 2025-02-05 5a49be5
bidding strategies included
