+++ assume/strategies/naive_strategies.py
@@ -146,8 +146,8 @@ class NaiveProfileStrategy(BaseStrategy):
 
 class NaiveDADSMStrategy(BaseStrategy):
     """
-    A naive strategy of a Demand Side Management (DSM) unit. The bid volume is the optimal power requirement of
-    the unit at the start time of the product. The bid price is the marginal cost of the unit at the start time of the product.
+    A naive strategy of a Demand Side Management (DSM) unit for the Day-Ahead Market.
+    The bid volume is the optimal power requirement above the baseline threshold.
     """
 
     def calculate_bids(
@@ -157,10 +157,22 @@ class NaiveDADSMStrategy(BaseStrategy):
         product_tuples: list[Product],
         **kwargs,
     ) -> Orderbook:
-        # calculate the optimal operation of the unit
-        # unit.determine_optimal_operation_without_flex()
-        unit.determine_optimal_operation_with_flex()
-        self.plot_power_requirements(unit)
+        """
+        Formulate bids for the Day-Ahead Market.
+
+        Args:
+            unit: The demand-side agent (e.g., cement plant) for which bids are being formulated.
+            market_config: Market configuration containing product details and constraints.
+            product_tuples: List of products for the Day-Ahead Market (duration and time steps).
+            **kwargs: Additional arguments for bid calculation.
+
+        Returns:
+            Orderbook: Contains bids for the Day-Ahead Market.
+        """
+
+        # Calculate the baseline threshold (80% of maximum power requirement)
+        max_power = max(unit.opt_power_requirement)
+        baseline_threshold = 0.7 * max_power
 
         bids = []
         for product in product_tuples:
@@ -169,16 +181,29 @@ class NaiveDADSMStrategy(BaseStrategy):
             and the volume of the product. Dispatch the order to the market.
             """
             start = product[0]
+            end = product[1]
 
-            volume = unit.opt_power_requirement.at[start]
-            marginal_price = unit.calculate_marginal_cost(start=start, power=volume)
+            # Calculate bid volume (energy above the threshold)
+            opt_power = unit.opt_power_requirement.at[start]
+            if opt_power > baseline_threshold:
+                bid_volume = opt_power - baseline_threshold
+            else:
+                bid_volume = 0
+
+            # Skip if bid volume is 0
+            if bid_volume <= 0:
+                continue
+
+            # Calculate the marginal cost for the bid volume
+            marginal_price = unit.calculate_marginal_cost(start=start, power=bid_volume)
+
             bids.append(
                 {
                     "start_time": start,
-                    "end_time": product[1],
+                    "end_time": end,
                     "only_hours": product[2],
-                    "price": marginal_price,
-                    "volume": -volume,
+                    "price": 3000,
+                    "volume": -bid_volume,
                 }
             )
 
@@ -244,14 +269,14 @@ class OTC_DSM_Strategy(BaseStrategy):
         Returns:
             Orderbook: Contains bids for the LTM market.
         """
-        # Determine the optimal operation of the unit
-        unit.determine_optimal_operation_without_flex()
+        if unit.optimisation_counter == 0:
+            unit.determine_optimal_operation_with_flex()
+            self.plot_power_requirements(unit)
+            unit.optimisation_counter = 1
 
-        # Calculate the baseline power threshold (80% of max optimal power requirement)
+        # Calculate the baseline threshold (70% of maximum power requirement)
         max_power = max(unit.opt_power_requirement)
-        print(max_power)
-        baseline_threshold = 0.8 * max_power
-        # print(baseline_threshold)
+        baseline_threshold = 0.7 * max_power
 
         bids = []
 
@@ -259,13 +284,187 @@ class OTC_DSM_Strategy(BaseStrategy):
             start = product[0]
             end = product[1]
 
[…]
