### 2025-02-05 3d738c4
-better handle save directory for the strategies -fix release notes
