### 2025-02-10 a1edc91
-fix single agent updates
