### 2025-02-10 856e43b
Moved the TensorBoard logger from world to the learning_role.
