+++ assume/reinforcement_learning/learning_role.py
@@ -17,6 +17,7 @@ from assume.reinforcement_learning.algorithms.base_algorithm import RLAlgorithm
 from assume.reinforcement_learning.algorithms.matd3 import TD3
 from assume.reinforcement_learning.buffer import ReplayBuffer
 from assume.reinforcement_learning.learning_utils import linear_schedule_func
+from assume.reinforcement_learning.tensorboard_logger import TensorBoardLogger
 
 logger = logging.getLogger(__name__)
 
@@ -139,6 +140,17 @@ class Learning(Role):
         self.rl_eval = defaultdict(list)
         # list of avg_changes
         self.avg_rewards = []
+
+        # tensor board logging
+        self.tensor_board_path = (
+            f"{learning_config.get('trained_policies_save_path', 'logs')}/TB"
+        )
+        self.tensor_board_logger = None
+
+        #
+        self.db_addr = None
+        self.freq_timedelta = None
+        self.time_steps = None
 
     def load_inter_episodic_data(self, inter_episodic_data):
         """
@@ -391,6 +403,37 @@ class Learning(Role):
                         return True
             return False
 
+    def init_logging(self):
+        """
+        Initialize the logging for the reinforcement learning agent.
+
+        This method initializes the tensor board logger for the reinforcement learning agent.
+        It also initializes the parameters required for sending data to the output role.
+
+        Notes:
+            This method is typically called at the beginning of the training process to enable logging.
+        """
+
+        self.tensor_board_logger = TensorBoardLogger(
+            simulation_id=self.context.data.get("simulation_id"),
+            db_uri=self.context.data.get("db_uri"),
+            tensorboard_path=self.tensor_board_path,
+            learning_mode=self.learning_mode,
+            episodes_collecting_initial_experience=self.episodes_collecting_initial_experience,
+            perform_evaluation=self.perform_evaluation,
+        )
+
+        # Parameters required for sending data to the output role
+        self.db_addr = self.context.data.get("output_agent_addr")
+
+        self.datetime = pd.to_datetime(self.context.data.get("train_start"))
+
+        self.freq_timedelta = pd.Timedelta(self.context.data.get("freq"))
+        # This is for padding the output list when the gradient steps are not identical to the train frequency
+        self.time_steps = int(
+            pd.Timedelta(self.train_freq) / (self.freq_timedelta * self.gradient_steps)
+        )
+
     def write_rl_critic_params_to_output(
         self, learning_rate: float, unit_params_list: list[dict]
     ) -> None:
@@ -409,17 +452,6 @@ class Learning(Role):
             A list of dictionaries containing critic losses for each time step.
             Each dictionary maps critic names to their corresponding loss values.
         """
-        db_addr = self.context.data.get("output_agent_addr")
-
-        # Initialize datetime the first time the function is called
-        if self.datetime is None:
-            self.datetime = pd.to_datetime(self.context.data.get("train_start"))
-
-        freq_timedelta = pd.Timedelta(self.context.data.get("freq"))
-        # This is for padding the output list when the gradient steps are not identical to the train frequency
-        time_steps = int(
-            pd.Timedelta(self.train_freq) / (freq_timedelta * self.gradient_steps)
-        )
 
         output_list = [
             {
@@ -429,16 +461,19 @@ class Learning(Role):
                 "max_grad_norm": params["max_grad_norm"],
                 "learning_rate": learning_rate,
                 "datetime": self.datetime
-                + ((time_step * self.gradient_steps + gradient_step) * freq_timedelta),
+                + (
+                    (time_step * self.gradient_steps + gradient_step)
+                    * self.freq_timedelta
+                ),
             }
-            for time_step in range(time_steps)
+            for time_step in range(self.time_steps)
[…]
