### 2025-02-11 ff6fe1a
-remove collecting episodes from output role -set initial exploration as bool in strategies -pass exploration mode from rl unit operator to outputs role -remove unrequired context data from learning role
