+++ assume/world.py
@@ -78,7 +78,6 @@ class World:
         perform_evaluation (bool): A boolean indicating whether the evaluation mode is enabled.
         forecaster (Forecaster, optional): The forecaster used for custom unit types.
         learning_mode (bool): A boolean indicating whether the learning mode is enabled.
-        episodes_collecting_initial_experience (int): The number of episodes for collecting initial experience.
         output_agent_addr (tuple[str, str]): The address of the output agent.
         bidding_params (dict): Parameters for bidding.
         index (pandas.Series): The index for the simulation.
@@ -218,11 +217,7 @@ class World:
             **container_kwargs,
         )
         self.learning_mode = self.learning_config.get("learning_mode", False)
-        # if we do not have initial experience collected we will get an error as no samples are available on the
-        # buffer from which we can draw experience to adapt the strategy, hence we set it to minimum one episode
-        self.episodes_collecting_initial_experience = max(
-            learning_config.get("episodes_collecting_initial_experience", 5), 1
-        )
+
         if not self.db_uri and not self.export_csv_path:
             self.output_agent_addr = None
         else:
@@ -272,18 +267,13 @@ class World:
             )
             rl_agent.suspendable_tasks = False
 
-            rl_agent._role_context.data.update(
-                {
-                    "output_agent_addr": self.output_agent_addr,
-                    "simulation_id": simulation_id,
-                    "db_uri": self.db_uri,
-                    "train_start": self.start,
-                    "train_end": self.end,
-                    "freq": self.forecaster.index.freq,
-                }
+            self.learning_role.init_logging(
+                simulation_id=simulation_id,
+                db_uri=self.db_uri,
+                output_agent_addr=self.output_agent_addr,
+                train_start=self.start,
+                freq=self.forecaster.index.freq,
             )
-
-            self.learning_role.init_logging()
 
     def setup_output_agent(self, simulation_id: str, save_frequency_hours: int) -> None:
         """
@@ -309,7 +299,6 @@ class World:
             export_csv_path=self.export_csv_path,
             save_frequency_hours=save_frequency_hours,
             learning_mode=self.learning_mode,
-            episodes_collecting_initial_experience=self.episodes_collecting_initial_experience,
             perform_evaluation=self.perform_evaluation,
             additional_kpis=self.additional_kpis,
         )
