### 2025-02-11 f561da8
-fix error asking to delete folder by creating writer only when needed
