### 2025-02-11 605b68b
-store only average values in tensorboard -store everything per run to be able to compare -use tensorboard folder in root path to store everything
