### 2025-02-11 a9555c0
-dynamically collect noise and regret columns -collect values also during initial exploration -log average reward per episode -collect rewards and profits as sums
