+++ assume/reinforcement_learning/tensorboard_logger.py
@@ -130,51 +130,92 @@ class TensorBoardLogger:
 
         mode = "train" if not self.perform_evaluation else "eval"
 
-        # columns specific to the training mode.
-        train_columns = """,
-            AVG(regret) AS regret,
-            AVG(critic_loss) AS loss,
+        # Dynamically detect noise columns in database
+        query_columns = (
+            "PRAGMA table_info(rl_params)"
+            if self.db.dialect.name == "sqlite"
+            else """
+            SELECT column_name FROM information_schema.columns
+            WHERE table_name = 'rl_params'
+        """
+        )
+        columns_df = pd.read_sql(query_columns, self.db)
+
+        column_names = (
+            columns_df["name"].tolist()
+            if self.db.dialect.name == "sqlite"
+            else columns_df["column_name"].tolist()
+        )
+        noise_columns = [
+            col for col in column_names if col.startswith("exploration_noise_")
+        ]
+        noise_sql = (
+            ", ".join([f"AVG({col}) AS {col}" for col in noise_columns])
+            if noise_columns
+            else ""
+        )
+
+        # Define the SQL query dynamically
+        critic_columns = (
+            """AVG(critic_loss) AS loss,
             AVG(total_grad_norm) AS total_grad_norm,
             MAX(max_grad_norm) AS max_grad_norm,
-            AVG(learning_rate) AS lr,
-            AVG(exploration_noise_0) AS noise_0,
-            AVG(exploration_noise_1) AS noise_1
-            """
+            AVG(learning_rate) AS lr
+        """
+            if mode == "train"
+            and self.episode > self.episodes_collecting_initial_experience
+            else ""
+        )
 
-        # Use appropriate date function and parameter style based on database type
-        if self.db.dialect.name == "sqlite":
-            # To aggregate by hour instead of day, replace '%Y-%m-%d' with '%Y-%m-%d %H'
-            # To aggregate by month instead of day, replace '%Y-%m-%d' with '%Y-%m'
-            date_func = "strftime('%Y-%m-%d', datetime)"
-        elif self.db.dialect.name == "postgresql":
-            # To aggregate by hour instead of day, replace 'YYYY-MM-DD' with 'YYYY-MM-DD HH24'
-            # To aggregate by month instead of day, replace 'YYYY-MM-DD' with 'YYYY-MM'
-            date_func = "TO_CHAR(datetime, 'YYYY-MM-DD')"
+        date_func = (
+            "strftime('%Y-%m-%d', datetime)"
+            if self.db.dialect.name == "sqlite"
+            else "TO_CHAR(datetime, 'YYYY-MM-DD')"
+        )
 
+        # Dynamically build the SQL query, ensuring proper commas
+        query_parts = [
+            f"{date_func} AS dt",
+            "unit",
+            "SUM(profit) AS profit",
+            "SUM(reward) AS reward",
+        ]
+
+        # Add regret column if it exists
+        if "regret" in column_names:
+            query_parts.append("SUM(regret) AS regret")
+
+        # Add critic-related columns (if any)
+        if critic_columns:
+            query_parts.append(critic_columns)
+
+        # Add noise columns (if any)
+        if noise_sql:
+            query_parts.append(noise_sql)
+
+        # Build the final SQL query
         query = f"""
             SELECT
-                {date_func} AS dt,
-                unit,
-                AVG(profit) AS profit,
-                AVG(reward) AS reward
-                {train_columns if mode == 'train' else ''}
+                {", ".join(query_parts)}
             FROM rl_params
             WHERE episode = '{self.episode}'
             AND simulation = '{self.simulation_id}'
             AND perform_evaluation = {self.perform_evaluation}
-            {'AND initial_exploration = False' if mode == 'train' else ''}
[…]
