+++ README.md
@@ -88,11 +88,10 @@ This will launch a container for TimescaleDB and Grafana with preconfigured dash
 
 When running an example with learning capabilities, you can start TensorBoard to observe the learning process. Use the following shell command to start TensorBoard:
 ```shell
-tensorboard --logdir "examples/inputs/example_name"
+tensorboard --logdir tensorboard
 ```
-while replacing `"example_name"` with the correct example name that differs according to the example investigated.
 
-You can then head to `http://localhost:6006/` to view and evaluate the training process parameters.
+You can then head to `http://localhost:6006/` to view and evaluate the training progress.
 
 Please note that TensorBoard should ideally be shut down via `Ctrl + C` every time you want to start a new simulation run in the same folder structure and want to overwrite existing results, as failing to do so may lead to conflicts deleting old logs.
 
