+++ assume/units/dsm_load_shift.py
@@ -270,7 +270,7 @@ class DSMFlex:
             model (pyomo.ConcreteModel): The Pyomo model being optimized.
         """
 
-        max_load = max(self.opt_power_requirement)
+        max_load = max(self.opt_power)
 
         peak_load_cap_value = max_load * (
             self.peak_load_cap / 100
@@ -279,7 +279,7 @@ class DSMFlex:
         model.peak_load_cap_value = pyo.Param(initialize=peak_load_cap_value)
 
         # Parameters
-        model.cost_tolerance = pyo.Param(initialize=self.cost_tolerance)
+        # model.cost_tolerance = pyo.Param(initialize=self.cost_tolerance)
         model.total_cost = pyo.Param(initialize=0.0, mutable=True)
 
         # Variables for load shifting
@@ -314,18 +314,37 @@ class DSMFlex:
         # Power input constraint with flexibility based on congestion
         @model.Constraint(model.time_steps)
         def total_power_input_constraint_with_peak_shift(m, t):
-            if self.has_electrolyser:
+            if self.technology == "steel_plant":
+                if self.has_electrolyser:
+                    return (
+                        m.total_power_input[t] + m.load_shift_pos[t] - m.load_shift_neg[t]
+                        == m.dsm_blocks["electrolyser"].power_in[t]
+                        + m.dsm_blocks["eaf"].power_in[t]
+                        + m.dsm_blocks["dri_plant"].power_in[t]
+                    )
+                else:
+                    return (
+                        m.total_power_input[t] + m.load_shift_pos[t] - m.load_shift_neg[t]
+                        == m.dsm_blocks["eaf"].power_in[t]
+                        + m.dsm_blocks["dri_plant"].power_in[t]
+                    )
+            if self.technology == "cement_plant":
+                power_input = 0
+                if self.has_raw_mill:
+                    power_input += self.model.dsm_blocks["raw_material_mill"].power_in[
+                        t
+                    ]
+                if self.has_cement_mill:
+                    power_input += self.model.dsm_blocks["cement_mill"].power_in[t]
+                if self.has_clinker_system:
+                    power_input += self.model.dsm_blocks["clinker_system"].power_in[t]
+                if self.has_ccs_system:
+                    power_input += self.model.dsm_blocks["ccs_system"].power_in[t]
+                if self.has_electrolyser:
+                    power_input += self.model.dsm_blocks["electrolyser"].power_in[t]
                 return (
                     m.total_power_input[t] + m.load_shift_pos[t] - m.load_shift_neg[t]
-                    == m.dsm_blocks["electrolyser"].power_in[t]
-                    + m.dsm_blocks["eaf"].power_in[t]
-                    + m.dsm_blocks["dri_plant"].power_in[t]
-                )
-            else:
-                return (
-                    m.total_power_input[t] + m.load_shift_pos[t] - m.load_shift_neg[t]
-                    == m.dsm_blocks["eaf"].power_in[t]
-                    + m.dsm_blocks["dri_plant"].power_in[t]
+                    == power_input
                 )
 
         @model.Constraint(model.time_steps)
@@ -333,18 +352,37 @@ class DSMFlex:
             """
             Ensures that the power input during peak periods does not exceed the peak threshold value.
             """
-            if self.has_electrolyser:
+            if self.technology == "steel_plant":
+                if self.has_electrolyser:
+                    return (
+                        m.dsm_blocks["electrolyser"].power_in[t]
+                        + m.dsm_blocks["eaf"].power_in[t]
+                        + m.dsm_blocks["dri_plant"].power_in[t]
+                        <= peak_load_cap_value
+                    )
+                else:
+                    return (
+                        m.dsm_blocks["eaf"].power_in[t]
[…]
