+++ examples/inputs/paper/config.yaml
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 cement_plant_case:
-  start_date: 2019-01-01 00:00
-  end_date: 2019-01-10 23:00
+  start_date: 2021-01-01 00:00
+  end_date: 2021-04-30 23:00
   time_step: 1h
   save_frequency_hours: 24
 
@@ -16,7 +16,7 @@ cement_plant_case:
     LTM_OTC:
       operator: LTM_operator
       product_type: energy
-      start_date: 2019-01-01 00:00
+      start_date: 2021-01-01 00:00
       products:
         - duration: 24h
           count: 1
@@ -31,7 +31,7 @@ cement_plant_case:
     CRM_pos:
       operator: CRM_operator
       product_type: capacity_pos
-      start_date: 2019-01-01 03:00
+      start_date: 2021-01-01 03:00
       products:
         - duration: 4h
           count: 1
@@ -48,7 +48,7 @@ cement_plant_case:
     CRM_neg:
       operator: CRM_operator
       product_type: capacity_neg
-      start_date: 2019-01-01 03:00
+      start_date: 2021-01-01 03:00
       products:
         - duration: 4h
           count: 1
@@ -63,7 +63,7 @@ cement_plant_case:
       market_mechanism: pay_as_bid
 
     EOM:
-      start_date: 2019-01-01 05:00
+      start_date: 2021-01-01 05:00
       operator: EOM_operator
       product_type: energy
       products:
@@ -78,6 +78,3 @@ cement_plant_case:
       minimum_bid_price: -500
       price_unit: EUR/MWh
       market_mechanism: pay_as_clear
-
-
-
