### 2025-02-13 a3b603c
Improves built-in forecast calculations - unstead of using only the marginal cost for the first time step, now it uses time varying marginal costs to calculate simple merit order based market clearing price - extracting VRE technologies for residual load calculation is now more robust taking all power plants with wind and solar in the technology, so any combinations like wind_onshore or Wind Offshore will work fine
