### 2025-02-14 33cbb9e
-fix loading strategies -add temp tensorboard solution
