### 2025-02-14 39f8022
-load and setup once
