+++ examples/notebooks/MPEC/uc_problem.py
@@ -1,0 +1,195 @@
+"""
+
+Created by: Nick Harder (nick.harder94@gmail.com)
+Created on August, 21th, 2023
+
+"""
+
+import pandas as pd
+
+# %%
+# Imports
+import pyomo.environ as pyo
+from pyomo.opt import SolverFactory
+
+
+# %%
+def solve_uc_problem(gens_df, demand_df, k_values_df):
+    model = pyo.ConcreteModel()
+
+    # sets
+    model.time = pyo.Set(initialize=demand_df.index)
+    model.gens = pyo.Set(initialize=gens_df.index)
+
+    # primary problem variables
+    model.g = pyo.Var(model.gens, model.time, within=pyo.NonNegativeReals) # Power output of producer 𝑖 at period 𝑡 (MW)
+    model.d = pyo.Var(model.time, within=pyo.NonNegativeReals) # satisfied demand at period 𝑡 (MW)
+    model.c_up = pyo.Var(model.gens, model.time, within=pyo.NonNegativeReals) # Start-up cost of producer 𝑖 at period 𝑡 (€)
+    model.c_down = pyo.Var(model.gens, model.time, within=pyo.NonNegativeReals) # Shut-down cost of producer 𝑖 at period 𝑡 (€)
+    model.u = pyo.Var(model.gens, model.time, within=pyo.Binary) # Binary UC status of producer 𝑖 at period 𝑡 (𝑢 = 1 if it is on, 𝑢 = 0 if it is off)
+
+    # primary problem objective
+    def objective_rule(model):
+        expr = 0
+        for gen in model.gens:
+            for t in model.time:
+                expr += (
+                    k_values_df.at[t, gen] * gens_df.at[gen, "mc"] * model.g[gen, t]
+                    + model.c_up[gen, t]
+                    + model.c_down[gen, t]
+                )
+
+        for t in model.time:
+            expr -= demand_df.at[t, "price"] * model.d[t]
+
+        return expr
+
+    model.objective = pyo.Objective(rule=objective_rule, sense=pyo.minimize)
+
+    # energy balance constraint
+    def balance_rule(model, t):
+        return model.d[t] - sum(model.g[i, t] for i in model.gens) == 0
+
+    model.balance = pyo.Constraint(model.time, rule=balance_rule)
+
+    # max generation constraint
+    def g_max_rule(model, i, t):
+        return model.g[i, t] <= gens_df.at[i, "g_max"] * model.u[i, t]
+
+    model.g_max = pyo.Constraint(model.gens, model.time, rule=g_max_rule)
+
+    # max demand constraint
+    # TODO: Wieso erlauben wir hier demand kleiner als der input demand? 
+    def d_max_rule(model, t):
+        return model.d[t] <= demand_df.at[t, "volume"]
+
+    model.d_max = pyo.Constraint(model.time, rule=d_max_rule)
+
+    # ramp up constraint
+    def ramp_up_rule(model, i, t):
+        if t == 0:
+            return model.g[i, t] - gens_df.at[i, "g_0"] <= gens_df.at[i, "r_up"]
+        else:
+            return model.g[i, t] - model.g[i, t - 1] <= gens_df.at[i, "r_up"]
+
+    model.ramp_up = pyo.Constraint(model.gens, model.time, rule=ramp_up_rule)
+
+    # ramp down constraint
+    def ramp_down_rule(model, i, t):
+        if t == 0:
+            return gens_df.at[i, "g_0"] - model.g[i, t] <= gens_df.at[i, "r_down"]
+        else:
+            return model.g[i, t - 1] - model.g[i, t] <= gens_df.at[i, "r_down"]
+
+    model.ramp_down = pyo.Constraint(model.gens, model.time, rule=ramp_down_rule)
+
+    # start up cost constraint
[…]
