+++ examples/inputs/example_02b/config.yaml
@@ -16,7 +16,7 @@ base:
     algorithm: matd3
     actor_architecture: mlp
     learning_rate: 0.001
-    training_episodes: 100
+    training_episodes: 10
     episodes_collecting_initial_experience: 3
     train_freq: 24h
     gradient_steps: 1

[…]
