### 2025-02-17 fe658d0
Adjust episode indexing during initial exploration phase.
