+++ assume/scenario/loader_csv.py
@@ -4,7 +4,6 @@
 
 import copy
 import logging
-import shutil
 from collections import defaultdict
 from datetime import datetime
 from pathlib import Path
@@ -13,7 +12,6 @@ import dateutil.rrule as rr
 import numpy as np
 import pandas as pd
 import yaml
-from torch.utils.tensorboard import SummaryWriter
 from tqdm import tqdm
 
 from assume.common.base import LearningConfig
@@ -937,13 +935,6 @@ def run_learning(
 
     eval_episode = 1
 
-    # TODO: delete before release
-    # delete old tensorboard logs
-    scenario_id = world.scenario_data["sim_id"]
-    if Path(f"tensorboard/{scenario_id}").is_dir():
-        shutil.rmtree(f"tensorboard/{scenario_id}")
-    writer = SummaryWriter(log_dir=f"tensorboard/{scenario_id}")
-
     for episode in tqdm(
         range(1, world.learning_role.training_episodes + 1),
         desc="Training Episodes",
@@ -958,15 +949,6 @@ def run_learning(
             world.learning_role.load_inter_episodic_data(inter_episodic_data)
 
         world.run()
-
-        total_rewards = world.output_role.get_sum_reward()
-
-        if len(total_rewards) == 0:
-            raise AssumeException("No rewards were collected during evaluation run")
-
-        avg_reward = np.mean(total_rewards)
-
-        writer.add_scalar("Train/AvgReward", avg_reward, episode)
 
         # -----------------------------------------
         # Store updated information across episodes
@@ -1000,8 +982,6 @@ def run_learning(
 
             avg_reward = np.mean(total_rewards)
 
-            writer.add_scalar("Eval/AvgReward", avg_reward, eval_episode)
-
             # check reward improvement in evaluation run
             # and store best run in eval folder
             terminate = world.learning_role.compare_and_save_policies(
@@ -1032,8 +1012,6 @@ def run_learning(
     logger.info("################")
     logger.info("Training finished, Start evaluation run")
     world.export_csv_path = temp_csv_path
-
-    writer.close()
 
     world.reset()
 
