### 2025-02-18 0db2b0f
-do not load scenarios and strategies twice
