### 2025-02-19 1f41b7c
- fix config - improve comments - switch to best policy when loading instead of last policy
