+++ assume/strategies/learning_strategies.py
@@ -534,17 +534,17 @@ class RLStrategy(AbstractLearningStrategy):
 
         The reward is scaled and stored along with other outputs in the unit’s data to support learning.
         """
-        # function is called after the market is cleared and we get the market feedback,
-        # so we can calculate the profit
+        # Function is called after the market is cleared, and we get the market feedback,
+        # allowing us to calculate profit based on the realized transactions.
 
         product_type = marketconfig.product_type
 
         start = orderbook[0]["start_time"]
         end = orderbook[0]["end_time"]
-        # end includes the end of the last product, to get the last products' start time we deduct the frequency once
+        # `end_excl` marks the last product's start time by subtracting one frequency interval.
         end_excl = end - unit.index.freq
 
-        # depending on way the unit calculates marginal costs we take costs
+        # Depending on how the unit calculates marginal costs, retrieve cost values.
         marginal_cost = unit.calculate_marginal_cost(
             start, unit.outputs[product_type].at[start]
         )
@@ -558,23 +558,22 @@ class RLStrategy(AbstractLearningStrategy):
         accepted_volume_total = 0
         offered_volume_total = 0
 
-        # iterate over all orders in the orderbook, to calculate order specific profit
+        # Iterate over all orders in the orderbook to calculate order-specific profit.
         for order in orderbook:
             accepted_volume = order.get("accepted_volume", 0)
             accepted_volume_total += accepted_volume
 
             offered_volume_total += order["volume"]
 
-            # calculate profit as income - running_cost from this event
+            # Calculate profit as income minus operational cost for this event.
             order_income = market_clearing_price * accepted_volume * duration
             order_cost = marginal_cost * accepted_volume * duration
 
-            # collect profit and opportunity cost for all orders
+            # Accumulate income and operational cost for all orders.
             income += order_income
             operational_cost += order_cost
 
-        # consideration of start-up costs, which are evenly divided between the
-        # upward and downward regulation events
+        # Consideration of start-up costs, divided evenly between upward and downward regulation events.
         if (
             unit.outputs[product_type].at[start] != 0
             and unit.outputs[product_type].at[start - unit.index.freq] == 0
@@ -588,34 +587,37 @@ class RLStrategy(AbstractLearningStrategy):
 
         profit = income - operational_cost
 
-        # account for only 10% of the profit if it is positive to not overstimulate the egant
-        # to manipulate the market
+        # Stabilizing learning: Limit positive profit to 10% of its absolute value.
+        # This reduces variance in rewards and prevents overfitting to extreme profit-seeking behavior.
+        # However, this does NOT prevent the agent from exploiting market inefficiencies if they exist.
+        # RL by nature identifies and exploits system weaknesses if they lead to higher profit.
+        # This is not a price cap but rather a stabilizing factor to avoid reward spikes affecting learning stability.
         profit = min(profit, 0.1 * abs(profit))
 
-        # calculate opportunity cost
-        # as the loss of income we have because we are not running at full power
+        # Opportunity cost: The income lost due to not operating at full capacity.
         opportunity_cost = (
             (market_clearing_price - marginal_cost)
             * (offered_volume_total - accepted_volume_total)
             * duration
         )
 
-        # if our opportunity costs are negative, we did not miss an opportunity to earn money and we set them to 0
+        # If opportunity cost is negative, no income was lost, so we set it to zero.
         opportunity_cost = max(opportunity_cost, 0)
 
-        # set regret_scale as 0.1 if dispatched above min. power, 0.5 if not dispatched at all
+        # Dynamic regret scaling:
+        # - If accepted volume is positive, apply lower regret (0.1) to avoid punishment for being on the edge of the merit order.
[…]
