+++ examples/inputs/example_02c/config.yaml
@@ -15,7 +15,7 @@ base:
     max_bid_price: 100
     algorithm: matd3
     learning_rate: 0.0003
-    # learning_rate_schedule: linear
+    learning_rate_schedule: linear
     training_episodes: 100
     episodes_collecting_initial_experience: 10
     train_freq: 100h
@@ -27,9 +27,9 @@ base:
     noise_sigma: 0.1
     noise_scale: 1
     noise_dt: 1
-    # validation_episodes_interval: 5
-    # early_stopping_steps: 10
-    # early_stopping_threshold: 0.05
+    validation_episodes_interval: 5
+    early_stopping_steps: 10
+    early_stopping_threshold: 0.05
 
   markets_config:
     EOM:
