### 2025-02-19 c16cb70
Update assume/world.py
Co-authored-by: Florian Maurer <maurer@fh-aachen.de>
