### 2025-02-19 498f006
Fix storage operation when using DRL - execute reward calculation after the dispatch has been set - fix issue where min soc had to be at least 1 - add soc to the meta and grafana dashboard - adjust tests to fit new structure - remove unrequired function from the learning unit operator
