### 2025-02-19 a5196d2
- fix train freq error where not all observations were written to the buffer - fix missing values for learning role tensorboard when train freq and gradient steps are mismatched - add historical price to the storage startegy - reduce forecast for storage also to 12 hours for compatibility with power plants
