### 2025-02-19 780932c
-fix docs config
