### 2025-02-19 8a8528f
- fix documentation config
