### 2025-02-19 89f1244
- switch to max_power for opportunity cost
