### 2025-02-19 ec99b50
* use self.context.data to handle orders of a market, so that other roles can access this * make sure that accepted_volume and accepted_price is not already set on bids sent to the market * distinguish between registration and registration_reply message
