+++ examples/inputs/example_02e/config.yaml
@@ -4,7 +4,7 @@
 
 base:
   start_date: 2019-03-01 00:00
-  end_date: 2019-04-30 00:00
+  end_date: 2019-05-30 00:00
   time_step: 1h
   save_frequency_hours: null
   learning_mode: True
