### 2025-02-20 cedb269
Improve competitive behavior
Until now large scale MADRL simulations didn't produce desired results. I have reevaluated my approach and introduced some changes to the learning process to improve.

## Changes Proposed
- Introduce randomized policy update instead of updating all policies at once to prevent cooperative behavior and training.
- Refactor reward function to obstruct overprofits to bring the players closer to reality where obvious market manipulation would be punished.
- added smaller fixes to how continue learning is treated to enable learning more sophisticated policies from initia
[…]
