+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -4,6 +4,7 @@
 
 import logging
 import os
+import random
 
 import torch as th
 from torch.nn import functional as F
@@ -32,7 +33,6 @@ class TD3(RLAlgorithm):
         self,
         learning_role,
         learning_rate=1e-4,
-        episodes_collecting_initial_experience=10,
         batch_size=1024,
         tau=0.005,
         gamma=0.99,
@@ -45,7 +45,6 @@ class TD3(RLAlgorithm):
         super().__init__(
             learning_role,
             learning_rate,
-            episodes_collecting_initial_experience,
             batch_size,
             tau,
             gamma,
@@ -384,6 +383,7 @@ class TD3(RLAlgorithm):
         # Stack strategies for easier access
         strategies = list(self.learning_role.rl_strats.values())
         n_rl_agents = len(strategies)
+
         unit_params = [
             {
                 u_id: {"loss": None, "total_grad_norm": None, "max_grad_norm": None}
@@ -391,6 +391,12 @@ class TD3(RLAlgorithm):
             }
             for _ in range(self.gradient_steps)
         ]
+
+        # Randomly select a subset of strategies to update
+        if n_rl_agents > 3:
+            selected_strategies = random.sample(strategies, k=int(n_rl_agents / 3))
+        else:
+            selected_strategies = strategies
 
         # update noise decay and learning rate
         updated_noise_decay = self.learning_role.calc_noise_from_progress(
@@ -402,7 +408,7 @@ class TD3(RLAlgorithm):
         )
 
         # loop over all units to avoid update call for every gradient step, as it will be ambiguous
-        for strategy in strategies:
+        for strategy in selected_strategies:
             self.update_learning_rate(
                 [
                     strategy.critics.optimizer,
@@ -455,13 +461,16 @@ class TD3(RLAlgorithm):
             #####################################################################
 
             # Zero-grad for all critics before accumulation
-            for strategy in strategies:
+            for strategy in selected_strategies:
                 strategy.critics.optimizer.zero_grad(set_to_none=True)
 
             total_critic_loss = 0.0
 
             # Loop over all agents and accumulate critic loss
             for i, strategy in enumerate(strategies):
+                if strategy not in selected_strategies:
+                    continue
+
                 actor = strategy.actor
                 critic = strategy.critics
                 critic_target = strategy.target_critics
@@ -522,7 +531,7 @@ class TD3(RLAlgorithm):
             total_critic_loss.backward()
 
             # Clip the gradients and step each critic optimizer
-            for strategy in strategies:
+            for strategy in selected_strategies:
                 parameters = list(strategy.critics.parameters())
 
                 # Determine clipping statistics
@@ -541,13 +550,16 @@ class TD3(RLAlgorithm):
             ######################################################################
             if self.n_updates % self.policy_delay == 0:
                 # Zero-grad for all actors first
-                for strategy in strategies:
+                for strategy in selected_strategies:
                     strategy.actor.optimizer.zero_grad(set_to_none=True)
 
                 total_actor_loss = 0.0
 
                 # We'll compute each agent's actor loss, accumulate, then do one backprop
-                for i, strategy in enumerate(strategies):
+                for i, strategy in enumerate(selected_strategies):
+                    if strategy not in selected_strategies:
+                        continue
+
                     actor = strategy.actor
                     critic = strategy.critics
 
@@ -590,7 +602,7 @@ class TD3(RLAlgorithm):
                 total_actor_loss.backward()
 
                 # Clip and step each actor optimizer
-                for strategy in strategies:
+                for strategy in selected_strategies:
                     th.nn.utils.clip_grad_norm_(
                         strategy.actor.parameters(), max_norm=1.0
                     )
[…]
