+++ assume/strategies/learning_strategies.py
@@ -205,8 +205,8 @@ class RLStrategy(AbstractLearningStrategy):
                 dt=kwargs.get("noise_dt", 1.0),
             )
 
-        elif Path(kwargs["trained_policies_save_path"]).is_dir():
-            self.load_actor_params(load_path=kwargs["trained_policies_save_path"])
+        elif Path(kwargs["trained_policies_load_path"]).is_dir():
+            self.load_actor_params(load_path=kwargs["trained_policies_load_path"])
         else:
             raise FileNotFoundError(
                 f"No policies were provided for DRL unit {self.unit_id}!. Please provide a valid path to the trained policies."
@@ -534,82 +534,94 @@ class RLStrategy(AbstractLearningStrategy):
 
         The reward is scaled and stored along with other outputs in the unit’s data to support learning.
         """
-        # function is called after the market is cleared and we get the market feedback,
-        # so we can calculate the profit
+        # Function is called after the market is cleared, and we get the market feedback,
+        # allowing us to calculate profit based on the realized transactions.
 
         product_type = marketconfig.product_type
 
-        profit = 0
-        reward = 0
-        opportunity_cost = 0
-        costs = 0
+        start = orderbook[0]["start_time"]
+        end = orderbook[0]["end_time"]
+        # `end_excl` marks the last product's start time by subtracting one frequency interval.
+        end_excl = end - unit.index.freq
 
-        # iterate over all orders in the orderbook, to calculate order specific profit
+        # Depending on how the unit calculates marginal costs, retrieve cost values.
+        marginal_cost = unit.calculate_marginal_cost(
+            start, unit.outputs[product_type].at[start]
+        )
+        market_clearing_price = orderbook[0]["accepted_price"]
+
+        duration = (end - start) / timedelta(hours=1)
+
+        income = 0.0
+        operational_cost = 0.0
+
+        accepted_volume_total = 0
+        offered_volume_total = 0
+
+        # Iterate over all orders in the orderbook to calculate order-specific profit.
         for order in orderbook:
-            start = order["start_time"]
-            end = order["end_time"]
-            # end includes the end of the last product, to get the last products' start time we deduct the frequency once
-            end_excl = end - unit.index.freq
-
-            # depending on way the unit calculates marginal costs we take costs
-            marginal_cost = unit.calculate_marginal_cost(
-                start, unit.outputs[product_type].at[start]
-            )
-
-            duration = (end - start) / timedelta(hours=1)
-
             accepted_volume = order.get("accepted_volume", 0)
-            accepted_price = order.get("accepted_price", 0)
+            accepted_volume_total += accepted_volume
 
-            # calculate profit as income - running_cost from this event
-            order_profit = accepted_price * accepted_volume * duration
+            offered_volume_total += order["volume"]
+
+            # Calculate profit as income minus operational cost for this event.
+            order_income = market_clearing_price * accepted_volume * duration
             order_cost = marginal_cost * accepted_volume * duration
 
-            # collect profit and opportunity cost for all orders
-            profit += order_profit
-            costs += order_cost
+            # Accumulate income and operational cost for all orders.
+            income += order_income
+            operational_cost += order_cost
 
-        # calculate opportunity cost
-        # as the loss of income we have because we are not running at full power
+        # Consideration of start-up costs, divided evenly between upward and downward regulation events.
+        if (
+            unit.outputs[product_type].at[start] != 0
+            and unit.outputs[product_type].at[start - unit.index.freq] == 0
+        ):
+            operational_cost += unit.hot_start_cost / 2
+        elif (
[…]
