### 2025-02-20 9f93a05
- slight changes of warning - removed redundant abs
