### 2025-02-20 aa7cf59
only participate on markets where any unit has a bidding strategy
