### 2025-02-20 02ebb66
- fix mac discharge power calculation - fix all default train_freq to 24h - adjust tests
