### 2025-02-20 037e34f
- reduce the number of required episodes in config for storage learning
