+++ examples/inputs/example_02e/config.yaml
@@ -17,7 +17,7 @@ base:
     actor_architecture: mlp
     learning_rate: 0.001
     learning_rate_schedule: linear
-    training_episodes: 200
+    training_episodes: 100
     episodes_collecting_initial_experience: 5
     train_freq: 1000h
     gradient_steps: 1000
