+++ assume/strategies/naive_strategies.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import matplotlib.pyplot as plt
+import pandas as pd
 
 from assume.common.base import BaseStrategy, SupportsMinMax
 from assume.common.market_objects import MarketConfig, Order, Orderbook, Product
@@ -144,10 +145,341 @@ class NaiveProfileStrategy(BaseStrategy):
         return bids
 
 
-class NaiveDADSMStrategy(BaseStrategy):
+# class NaiveDADSMStrategy(BaseStrategy):
+#     """
+#     A naive strategy of a Demand Side Management (DSM) unit for the Day-Ahead Market.
+#     The bid volume is the optimal power requirement above the baseline threshold.
+#     """
+
+#     def calculate_bids(
+#         self,
+#         unit: SupportsMinMax,
+#         market_config: MarketConfig,
+#         product_tuples: list[Product],
+#         **kwargs,
+#     ) -> Orderbook:
+#         """
+#         Formulate bids for the Day-Ahead Market.
+
+#         Args:
+#             unit: The demand-side agent (e.g., cement plant) for which bids are being formulated.
+#             market_config: Market configuration containing product details and constraints.
+#             product_tuples: List of products for the Day-Ahead Market (duration and time steps).
+#             **kwargs: Additional arguments for bid calculation.
+
+#         Returns:
+#             Orderbook: Contains bids for the Day-Ahead Market.
+#         """
+
+#         # Calculate the baseline threshold (80% of maximum power requirement)
+#         max_power = max(unit.opt_power_requirement)
+#         baseline_threshold = 0.5 * max_power
+
+#         bids = []
+#         for product in product_tuples:
+#             """
+#             for each product, calculate the marginal cost of the unit at the start time of the product
+#             and the volume of the product. Dispatch the order to the market.
+#             """
+#             start = product[0]
+#             end = product[1]
+
+#             # Calculate bid volume (energy above the threshold)
+#             opt_power = unit.opt_power_requirement.at[start]
+#             if opt_power > baseline_threshold:
+#                 bid_volume = opt_power - baseline_threshold
+#             else:
+#                 bid_volume = 0
+
+#             # Skip if bid volume is 0
+#             if bid_volume <= 0:
+#                 continue
+
+#             # Calculate the marginal cost for the bid volume
+#             marginal_price = unit.calculate_marginal_cost(start=start, power=bid_volume)
+
+#             bids.append(
+#                 {
+#                     "start_time": start,
+#                     "end_time": end,
+#                     "only_hours": product[2],
+#                     "price": 3000,
+#                     "volume": -bid_volume,
+#                 }
+#             )
+
+#         return bids
+
+#     def plot_power_requirements(self, unit: SupportsMinMax):
+#         """
+#         Plots the optimal power requirement and flexibility power requirement for comparison.
+
+#         Args:
+#             unit (SupportsMinMax): The unit containing power requirements.
+#         """
+#         # Retrieve power requirements data
+#         opt_power_requirement = unit.opt_power_requirement
+#         flex_power_requirement = unit.flex_power_requirement
+
+#         # Plotting
+#         plt.figure(figsize=(10, 6))
+#         plt.plot(
+#             opt_power_requirement.index,
+#             opt_power_requirement,
+#             label="Optimal Power Requirement",
+#             color="blue",
+#         )
+#         plt.plot(
+#             flex_power_requirement.index,
+#             flex_power_requirement,
+#             label="Flex Power Requirement",
+#             color="orange",
+#             linestyle="--",
+#         )
+
+#         # Labels and title
+#         plt.xlabel("Time")
+#         plt.ylabel("Power Requirement (kW)")
+#         plt.title("Comparison of Optimal and Flexible Power Requirements")
[…]
