### 2025-02-20 35e4c85
add fuel market which bids on the price of fuel for a given hour
