### 2025-02-20 d68fa96
add fuel market which bids on the price of fuel for a given hour
