### 2025-02-20 832316c
Fix storage reward calculation and observation
Closes #547, #523 

## Description
Currently, the storage learning strategy is broken because the
`calculate_reward` function is executed before the actual dispatch,
leading to an incorrect state of charge (SOC). This fix ensures that the
correct SOC is set before reward calculation.

Additionally, a restriction requiring the minimum SOC to be at least 1
has been removed. These improvements also enhance compatibility with the
power plant learning strategy, enabling mixed simulations.

## Changes Proposed
- Execute reward calculation **after** disp
[…]
