### 2025-02-21 71de489
implement market participation lambda to better configure participation in multi market scenarios
