+++ tests/test_utils.py
@@ -620,17 +620,18 @@ def test_set_list():
     print(res_pd)
     assert res_fds < res_pd
 
+
 def test_slicing_fastseries_even():
     start = datetime(2020, 1, 1, 0)
     end = datetime(2020, 1, 1, 5)
     index = FastIndex(start, end, freq="1h")
     fs = FastSeries(index)
-    b = start+timedelta(hours=1)
-    e = start+timedelta(hours=4)
+    b = start + timedelta(hours=1)
+    e = start + timedelta(hours=4)
     result = fs[b:e]
-    
+
     datelist = fs.index.get_date_list(b, e)
-    series = pd.Series(0, index=pd.date_range(start,end, freq="h"))
+    series = pd.Series(0, index=pd.date_range(start, end, freq="h"))
     assert list(series[b:e].index) == datelist
     assert len(series[b:e]) == len(fs[b:e])
 
@@ -640,12 +641,12 @@ def test_slicing_fastseries_uneven():
     end = datetime(2020, 1, 1, 5)
     index = FastIndex(start, end, freq="1h")
     fs = FastSeries(index)
-    b = start+timedelta(seconds= 1)
-    e = start+timedelta(hours=4, seconds= 1)
+    b = start + timedelta(seconds=1)
+    e = start + timedelta(hours=4, seconds=1)
     result = fs[b:e]
-    
+
     datelist = fs.index.get_date_list(b, e)
-    series = pd.Series(0, index=pd.date_range(start,end, freq="h"))
+    series = pd.Series(0, index=pd.date_range(start, end, freq="h"))
     assert list(series[b:e].index) == datelist
     assert len(series[b:e]) == len(fs[b:e])
 
