+++ assume/common/fast_pandas.py
@@ -116,7 +116,7 @@ class FastIndex:
                 else item.start or 0
             )
             stop_idx = (
-                self._get_idx_from_date(item.stop, round_up=False) +1
+                self._get_idx_from_date(item.stop, round_up=False) + 1
                 if isinstance(item.stop, datetime)
                 else item.stop or len(self._date_list)
             )
@@ -201,7 +201,7 @@ class FastIndex:
             self._date_list = [self._start + timedelta(seconds=s) for s in total_dates]
 
         start_idx = self._get_idx_from_date(start or self.start)
-        end_idx = self._get_idx_from_date(end or self.end, round_up=False) +1
+        end_idx = self._get_idx_from_date(end or self.end, round_up=False) + 1
         return self._date_list[start_idx:end_idx]
 
     def as_datetimeindex(self) -> pd.DatetimeIndex:
@@ -217,7 +217,7 @@ class FastIndex:
         return pd.DatetimeIndex(pd.to_datetime(datetimes), name="FastIndex")
 
     @lru_cache(maxsize=1000)
-    def _get_idx_from_date(self, date: datetime, round_up:bool = True) -> int:
+    def _get_idx_from_date(self, date: datetime, round_up: bool = True) -> int:
         """
         Convert a datetime to its corresponding index in the range.
 
@@ -444,7 +444,7 @@ class FastSeries:
                 else 0
             )
             stop_idx = (
-                self.index._get_idx_from_date(item.stop, round_up=False)+1
+                self.index._get_idx_from_date(item.stop, round_up=False) + 1
                 if item.stop is not None
                 else len(self.data)
             )
@@ -511,7 +511,7 @@ class FastSeries:
                 )
             )
             stop_idx = (
-                self.index._get_idx_from_date(item.stop, round_up=False) +1
+                self.index._get_idx_from_date(item.stop, round_up=False) + 1
                 if isinstance(item.stop, datetime)
                 else (
                     len(self.data) + item.stop
