### 2025-02-24 f1c8772
improve documentation and usage of lambda functions
