+++ tests/test_units_operator.py
@@ -285,7 +285,7 @@ def test_participate():
     units_role = UnitsOperator(available_markets=[marketconfig])
 
     index = FastIndex(start=start, end=end + pd.Timedelta(hours=4), freq="1h")
-    
+
     assert not units_role.participate(marketconfig)
 
     params_dict = {
@@ -314,6 +314,49 @@ def test_participate():
 
     assert units_role.participate(marketconfig)
 
+def test_participate_lambda():
+    """
+    Tests that one of the selected lambda functions works correctly in the participation
+    """
+    market_id = "EOM"
+    marketconfig = MarketConfig(
+        market_id=market_id,
+        opening_hours=rr.rrule(rr.HOURLY, dtstart=start, until=end),
+        opening_duration=rd(hours=1),
+        market_mechanism="pay_as_clear",
+        market_products=[MarketProduct(rd(hours=1), 1, rd(hours=1))],
+        eligible_obligations_lambda="only_renewables",
+    )
+    units_role = UnitsOperator(available_markets=[marketconfig])
+    index = FastIndex(start=start, end=end + pd.Timedelta(hours=4), freq="1h")
+
+    assert not units_role.participate(marketconfig)
+
+    params_dict = {
+        "bidding_strategies": {"EOM": NaiveSingleBidStrategy()},
+        "technology": "energy",
+        "unit_operator": "x",
+        "max_power": 10,
+        "min_power": 0,
+        "forecaster": NaiveForecast(index, demand=1000),
+    }
+    unit = PowerPlant("testdemand", **params_dict)
+    units_role.add_unit(unit)
+    assert not units_role.participate(marketconfig)
+
+    params_dict = {
+        "bidding_strategies": {"EOM": NaiveSingleBidStrategy()},
+        "technology": "wind offshore",
+        "unit_operator": "x",
+        "max_power": 1000,
+        "min_power": 0,
+        "forecaster": NaiveForecast(index, demand=1000),
+    }
+    unit = PowerPlant("testdemand", **params_dict)
+    units_role.add_unit(unit)
+
+    assert units_role.participate(marketconfig)
+
 
 def test_participate_custom_lambda():
     """
@@ -326,11 +369,11 @@ def test_participate_custom_lambda():
         opening_duration=rd(hours=1),
         market_mechanism="pay_as_clear",
         market_products=[MarketProduct(rd(hours=1), 1, rd(hours=1))],
-        eligible_obligations_lambda=lambda u: abs(u.get("max_power",0)) > 100,
+        eligible_obligations_lambda=lambda u: abs(u.get("max_power", 0)) > 100,
     )
     units_role = UnitsOperator(available_markets=[marketconfig])
     index = FastIndex(start=start, end=end + pd.Timedelta(hours=4), freq="1h")
-    
+
     assert not units_role.participate(marketconfig)
 
     params_dict = {
