+++ assume/common/market_objects.py
@@ -2,6 +2,7 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
+import re
 from collections.abc import Callable
 from dataclasses import dataclass, field
 from datetime import datetime, timedelta
@@ -282,28 +283,18 @@ market_contract_type = Callable[[Agent, Agent, list], None]
 
 
 def only_renewables(unit):
-    """check that technology is renewable"""
-    return unit["technology"] in [
-        "demand",
-        "wind",
-        "wind_onshore",
-        "wind_offshore",
-        "solar",
-        "biomass",
-    ]
+    """
+    check that technology is renewable.
+    This allows all names containing wind, solar, bio or demand
+    """
+    return re.search(
+        r"\b(?:wind|solar|bio|demand)\b", unit["technology"], flags=re.IGNORECASE
+    )
 
 
 def only_co2emissionless(unit):
     """same as only_renewables plus nuclear"""
-    return unit["technology"] in [
-        "demand",
-        "nuclear",
-        "wind",
-        "wind_onshore",
-        "wind_offshore",
-        "solar",
-        "biomass",
-    ]
+    return only_renewables(unit) or unit["technology"] == "nuclear"
 
 
 def power_plant_not_negative(unit):
