### 2025-02-24 cb4095c
Changes the way rl_params are stored in the database and gets rid of changing simulation id
- change perform_evaluation to proper name evaluation_mode
- remove changing sim_id during learning
- store rl_parals under unified simulation_id and differentiate using episode
- use simulation_id as variable through the whole code for better readibility
