+++ assume/world.py
@@ -51,46 +51,53 @@ logger = logging.getLogger(__name__)
 
 class World:
     """
-    World instance with the provided address, database URI, export CSV path, log level, and distributed role settings.
+    Represents a simulation environment with a specified address, database connection, CSV export path,
+    log level, and optional distributed role settings.
 
-    If a database URI is provided, it establishes a database connection. Additionally, it sets up various dictionaries and attributes for market operators,
-    markets, unit operators, unit types, bidding strategies, and clearing mechanisms. If available, it imports learning strategies and handles any potential import errors.
-    Finally, it sets up the event loop for asynchronous operations.
+    If a database URI is provided, the World instance attempts to establish a database connection.
+    It initializes key attributes for market operators, markets, unit operators, bidding strategies,
+    and clearing mechanisms. Additionally, it checks for learning strategy availability and sets up an event loop.
 
     Attributes:
-        addr (Union[tuple[str, int], str]): The address of the world, represented as a tuple of string and int or a string.
+        addr (tuple[str, int] | str, optional): The address of the world, represented as a tuple (host, port) or a string.
+        distributed_role (bool, optional): Defines the world’s role in distributed execution:
+            - `True`: Acts as a manager world that schedules events.
+            - `False`: Acts as a client world receiving schedules from a manager.
+            - `None` (default): Runs independently without subprocesses.
+        export_csv_path (str, optional): Path for exporting CSV data.
+        log_level (str, optional): The logging level for the world instance.
+        db_uri (sqlalchemy.engine.URL, optional): The processed database URI.
+        db (sqlalchemy.engine.base.Engine, optional): The database connection engine.
         container (mango.Container, optional): The container for the world instance.
-        distributed_role (bool, optional): A boolean indicating whether distributed roles are enabled.
-        export_csv_path (str): The path for exporting CSV data.
-        db (sqlalchemy.engine.base.Engine, optional): The database connection.
-        market_operators (dict[str, mango.RoleAgent]): The market operators for the world instance.
-        markets (dict[str, MarketConfig]): The markets for the world instance.
-        unit_operators (dict[str, UnitsOperator]): The unit operators for the world instance.
-        unit_types (dict[str, BaseUnit]): The unit types for the world instance.
-        bidding_strategies (dict[str, type[BaseStrategy]]): The bidding strategies for the world instance.
-        clearing_mechanisms (dict[str, MarketRole]): The clearing mechanisms for the world instance.
-        addresses (list[str]): The addresses for the world instance.
-        loop (asyncio.AbstractEventLoop): The event loop for the world instance.
-        clock (ExternalClock): The external clock for the world instance.
-        start (datetime.datetime): The start datetime for the simulation.
-        end (datetime.datetime): The end datetime for the simulation.
-        learning_config (LearningConfig): The configuration for the learning process.
-        perform_evaluation (bool): A boolean indicating whether the evaluation mode is enabled.
+        loop (asyncio.AbstractEventLoop, optional): The event loop for asynchronous operations.
+        clock (ExternalClock, optional): External clock instance.
+        start (datetime.datetime, optional): Start datetime for the simulation.
+        end (datetime.datetime, optional): End datetime for the simulation.
+        market_operators (dict[str, mango.RoleAgent], optional): Market operators in the world instance.
+        markets (dict[str, MarketConfig], optional): Market configurations.
+        unit_operators (dict[str, UnitsOperator], optional): Unit operators.
+        unit_types (dict[str, BaseUnit], optional): Available unit types.
+        dst_components (dict[str, DemandSideTechnology], optional): Demand-side technologies.
[…]
