### 2025-02-24 a98c26a
Improve market participation (#551)
This makes sure, that at least one unit has a bidding strategy for a
market, before deciding to participate on that market.

Respect the `eligible_obligations_lambda` set in the `MarketConfig` to
only bid on markets where the UnitsOperator fulfills the requirements.

It is possible to set this as string for `only_renewables` or
`only_co2emissionless`.
