### 2025-02-24 624fd7e
- fixed of MPEC prepaeration, now right k_values are calculated
