### 2025-02-24 ab3eede
- fix logging generation cost
