### 2025-02-24 e14c818
drop storing marginal costs
