### 2025-02-25 1b3e85c
Refactor MPEC Notebook.
