+++ examples/notebooks/MPEC/utils.py
@@ -2,9 +2,16 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
+import os
+
 import pandas as pd
 import pyomo.environ as pyo
+import yaml
+from matplotlib import pyplot as plt
 from pyomo.opt import SolverFactory
+
+from examples.notebooks.MPEC.bilevel_opt import find_optimal_dispatch
+from examples.notebooks.MPEC.uc_problem import solve_uc_problem
 
 
 def calculate_profits(main_df, gens_df, price_column="mcp", supp_df=None):
@@ -182,3 +189,303 @@ def calculate_mc(powerplant, fuel_prices):
     mc = fuel_cost + co2_cost + variable_cost
 
     return mc
+
+
+def load_config(inputs_dir, scenario, study_case=""):
+    """
+
+    Load the config file from the given directory.
+
+    Args:
+        inputs_dir (str): The directory containing the config file.
+        scenario (str): The scenario name.
+        study_case (str): The study case name. If left empty, the entire config file is loaded.
+
+    Returns:
+        dict: The loaded config file.
+    """
+    config_path = os.path.join(inputs_dir, scenario, "config.yaml")
+    with open(config_path) as file:
+        config = yaml.safe_load(file)
+    if study_case != "":
+        config = config[study_case]
+    return config
+
+
+def store_config(config, inputs_dir, scenario):
+    """
+    Store the config file in the given directory.
+
+    Args:
+        config (dict): The config file to store.
+        scenario (str): The scenario name.
+        inputs_dir (str): The directory to store the config file.
+    """
+
+    class NoAliasDumper(yaml.SafeDumper):
+        def ignore_aliases(self, _):
+            return True
+
+    # Store the config file in the given directory
+    config_path = os.path.join(inputs_dir, scenario, "config.yaml")
+    with open(config_path, "w") as file:
+        yaml.dump(
+            config,
+            file,
+            Dumper=NoAliasDumper,
+            default_flow_style=False,
+            sort_keys=False,
+        )
+
+
+def get_no_of_val_episodes(inputs_dir, scenario, study_case):
+    # Read and calculate the number of validation episodes from the config file
+    config = load_config(inputs_dir, scenario, study_case)
+    learning_config = config["learning_config"]
+    no_of_val_episodes = (
+        learning_config["training_episodes"]
+        - learning_config["episodes_collecting_initial_experience"]
+    ) // learning_config.get("validation_episodes_interval", 5)
+    return no_of_val_episodes
+
+
+def retrieve_best_episode_actions(inputs_dir, scenario, study_case, db):
+    """
+    Retrieve the actions of the best episode based on the average reward of each episode.
+    """
+    simulation = f"{scenario}_{study_case}_eval"
+    # Get the number of validation episodes
+    no_of_val_episodes = get_no_of_val_episodes(inputs_dir, scenario, study_case)
+    # Get the average reward for each episode in order to determine the best episode.
+    reward_df = pd.DataFrame(
+        columns=["avg_reward"], index=range(1, no_of_val_episodes + 1)
+    )
+    for episode in range(1, no_of_val_episodes + 1):
+        query = f"SELECT AVG(reward) as avg_reward FROM rl_params where simulation = '{simulation}_{episode}'"
+        reward_df.at[episode, "avg_reward"] = pd.read_sql(query, db).values[0][0]
+
+    # Use the episode with the best reward to get the respective actions
+    best_episode = reward_df["avg_reward"].idxmax()
[…]
