### 2025-02-25 25f2c6c
- made comparison rl and mpec profits - preanalysed unintuitive example02b
