### 2025-02-25 3c67335
After latest update the generation costs are no longer written to the database. This is fixed here
