### 2025-02-25 57b8b5a
- changed filtering in loop to account for entire days not just first hour - fixed opt_gen to index conversion between mpec and rl
