### 2025-02-25 b53a650
Fix writing generation costs to the database (#555)
After latest update in #553 the generation costs are no longer written
to the database. This is fixed here
