### 2025-02-25 7ec1b77
-add soc update into set_dispatch for storage units so the new soc is available when bidding for the next period -fix flexable storage strategy for better theoretic soc handling
