+++ assume/strategies/flexable_storage.py
@@ -58,47 +58,28 @@ class flexableEOMStorage(BaseStrategy):
         """
 
         # =============================================================================
-        # Storage Unit is either charging, discharging, or off
-        # =============================================================================
-        start = product_tuples[0][0]
-        end_all = product_tuples[-1][1]
-
-        previous_power = unit.get_output_before(start)
-
-        # save a theoretic SOC to calculate the ramping
-        theoretic_SOC = unit.get_soc_before(start)
-
-        # calculate min and max power for charging and discharging
-        min_power_charge_values, max_power_charge_values = (
-            unit.calculate_min_max_charge(start, end_all)
-        )
-        min_power_discharge_values, max_power_discharge_values = (
-            unit.calculate_min_max_discharge(start, end_all, soc=theoretic_SOC)
-        )
-
-        # =============================================================================
         # Calculate bids
         # =============================================================================
-        bids = []
+        # save a theoretic SOC to calculate the ramping
+        theoretic_SOC = unit.outputs["soc"].at[product_tuples[0][0]]
 
-        for (
-            product,
-            max_power_discharge,
-            min_power_discharge,
-            max_power_charge,
-            min_power_charge,
-        ) in zip(
-            product_tuples,
-            max_power_discharge_values,
-            min_power_discharge_values,
-            max_power_charge_values,
-            min_power_charge_values,
-        ):
+        bids = []
+        for product in product_tuples:
             start, end = product[0], product[1]
+
+            previous_power = unit.get_output_before(start)
 
             current_power = unit.outputs["energy"].at[start]
             current_power_discharge = max(current_power, 0)
             current_power_charge = min(current_power, 0)
+
+            # calculate min and max power for charging and discharging
+            min_power_charge, max_power_charge = unit.calculate_min_max_charge(
+                start, end, soc=theoretic_SOC
+            )
+            min_power_discharge, max_power_discharge = unit.calculate_min_max_discharge(
+                start, end, soc=theoretic_SOC
+            )
 
             # Calculate ramping constraints using helper function
             max_power_discharge = unit.calculate_ramp_discharge(
@@ -176,8 +157,6 @@ class flexableEOMStorage(BaseStrategy):
 
             theoretic_SOC += delta_soc
             previous_power = bid_quantity + current_power
-
-        bids = self.remove_empty_bids(bids)
 
         return bids
 
@@ -267,7 +246,7 @@ class flexablePosCRMStorage(BaseStrategy):
         end = product_tuples[-1][1]
 
         previous_power = unit.get_output_before(start)
-        theoretic_SOC = unit.get_soc_before(start)
+        theoretic_SOC = unit.outputs["soc"].at[start]
 
         _, max_power_discharge_values = unit.calculate_min_max_discharge(
             start, end, soc=theoretic_SOC
@@ -400,7 +379,7 @@ class flexableNegCRMStorage(BaseStrategy):
 
         previous_power = unit.get_output_before(start)
 
-        theoretic_SOC = unit.get_soc_before(start)
+        theoretic_SOC = unit.outputs["soc"].at[start]
 
         _, max_power_charge_values = unit.calculate_min_max_charge(start, end)
 
