+++ assume/common/base.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 from collections import defaultdict
-from datetime import datetime
+from datetime import datetime, timedelta
 from typing import TypedDict
 
 import numpy as np
@@ -171,6 +171,44 @@ class BaseUnit:
             self.outputs[f"{product_type}_accepted_price"].loc[start:end_excl] = (
                 accepted_price
             )
+
+        # also update the SOC if the unit is a storage unit
+        if isinstance(self, SupportsMinMaxCharge):
+            start = orderbook[0]["start_time"]
+            end = orderbook[-1]["end_time"]
+            time_delta = self.index.freq / timedelta(hours=1)
+
+            for t in self.index[start:end]:
+                next_t = t + self.index.freq
+                # continue if it is the last time step
+                if next_t not in self.index:
+                    continue
+                current_power = self.outputs["energy"].at[t]
+
+                # calculate the change in state of charge
+                delta_soc = 0
+                soc = self.outputs["soc"].at[t]
+
+                # discharging
+                if current_power > 0:
+                    max_soc_discharge = self.calculate_soc_max_discharge(soc)
+
+                    if current_power > max_soc_discharge:
+                        current_power = max_soc_discharge
+
+                    delta_soc = -current_power * time_delta / self.efficiency_discharge
+
+                # charging
+                elif current_power < 0:
+                    max_soc_charge = self.calculate_soc_max_charge(soc)
+
+                    if current_power < max_soc_charge:
+                        current_power = max_soc_charge
+
+                    delta_soc = -current_power * time_delta * self.efficiency_charge
+
+                # update the values of the state of charge and the energy
+                self.outputs["soc"].at[next_t] = soc + delta_soc
 
     def calculate_cashflow_and_reward(
         self,
@@ -538,23 +576,6 @@ class SupportsMinMaxCharge(BaseUnit):
         Returns:
             tuple[np.array, np.array]: The min and max discharging power for the given time period.
         """
-
-    def get_soc_before(self, dt: datetime) -> float:
-        """
-        Returns the State of Charge (SoC) before the given datetime.
-        If datetime is before the start of the index, the initial SoC is returned.
-        The SoC is a float between 0 and 1.
-
-        Args:
-            dt (datetime.datetime): The current datetime.
-
-        Returns:
-            float: The SoC before the given datetime.
-        """
-        if dt - self.index.freq < self.index[0]:
-            return self.initial_soc
-        else:
-            return self.outputs["soc"].at[dt - self.index.freq]
 
     def calculate_ramp_discharge(
         self,
