+++ examples/examples.py
@@ -154,11 +154,6 @@ if __name__ == "__main__":
 
         if world.learning_config.get("learning_mode", False):
             # run learning if learning mode is enabled
-            run_learning(
-                world,
-                inputs_path="examples/inputs",
-                scenario=available_examples[example]["scenario"],
-                study_case=available_examples[example]["study_case"],
-            )
+            run_learning(world)
 
         world.run()
