### 2025-02-26 5a2ce97
Move indexing out of run_MPEC function.
