+++ tests/test_storage.py
@@ -83,21 +83,21 @@ def test_calculate_operational_window(storage_unit):
     start = datetime(2022, 1, 1, 0)
     end = datetime(2022, 1, 1, 1)
     min_power_discharge, max_power_discharge = storage_unit.calculate_min_max_discharge(
-        start, end, product_type="energy"
+        start, end
     )
-    cost_discharge = storage_unit.calculate_marginal_cost(start, max_power_discharge[0])
+    cost_discharge = storage_unit.calculate_marginal_cost(start, max_power_discharge)
 
-    assert min_power_discharge[0] == 0
-    assert max_power_discharge[0] == 100
+    assert min_power_discharge == 0
+    assert max_power_discharge == 100
     assert cost_discharge == 4 / 0.95
 
     min_power_charge, max_power_charge = storage_unit.calculate_min_max_charge(
-        start, end, product_type="energy"
+        start, end
     )
-    cost_charge = storage_unit.calculate_marginal_cost(start, max_power_charge[0])
+    cost_charge = storage_unit.calculate_marginal_cost(start, max_power_charge)
 
-    assert min_power_charge[0] == 0
-    assert max_power_charge[0] == -100
+    assert min_power_charge == 0
+    assert max_power_charge == -100
     assert math.isclose(cost_charge, 3 / 0.9)
 
     assert storage_unit.outputs["energy"].at[start] == 0
@@ -109,14 +109,14 @@ def test_calculate_operational_window(storage_unit):
     min_power_charge, max_power_charge = storage_unit.calculate_min_max_charge(
         start, end
     )
-    assert min_power_charge[0] == -40
-    assert max_power_charge[0] == -60
+    assert min_power_charge == -40
+    assert max_power_charge == -60
 
     min_power_discharge, max_power_discharge = storage_unit.calculate_min_max_discharge(
         start, end
     )
-    assert min_power_discharge[0] == 40
-    assert max_power_discharge[0] == 60
+    assert min_power_discharge == 40
+    assert max_power_discharge == 60
 
     start = start + timedelta(hours=1)
 
@@ -133,20 +133,19 @@ def test_soc_constraint(storage_unit):
     storage_unit.outputs["soc"][start - timedelta(hours=1)] = (
         0.05 * storage_unit.max_soc
     )
-    assert storage_unit.get_soc_before(start) == 0.05 * storage_unit.max_soc
     min_power_discharge, max_power_discharge = storage_unit.calculate_min_max_discharge(
         start, end
     )
-    assert min_power_discharge[0] == 40
-    assert max_power_discharge[0] == 60
+    assert min_power_discharge == 40
+    assert max_power_discharge == 60
 
     storage_unit.outputs["soc"][start] = 0.95 * storage_unit.max_soc
     min_power_charge, max_power_charge = storage_unit.calculate_min_max_charge(
         start, end
     )
-    assert min_power_charge[0] == -40
+    assert min_power_charge == -40
     assert math.isclose(
-        max_power_charge[0], -50 / storage_unit.efficiency_charge, abs_tol=0.1
+        max_power_charge, -50 / storage_unit.efficiency_charge, abs_tol=0.1
     )
 
 
@@ -154,19 +153,19 @@ def test_storage_feedback(storage_unit, mock_market_config):
     start = datetime(2022, 1, 1, 0)
     end = datetime(2022, 1, 1, 1)
     min_power_charge, max_power_charge = storage_unit.calculate_min_max_charge(
-        start, end, product_type="energy"
+        start, end
     )
 
     min_power_discharge, max_power_discharge = storage_unit.calculate_min_max_discharge(
-        start, end, product_type="energy"
+        start, end
     )
[…]
