+++ assume/units/storage.py
@@ -325,12 +325,6 @@ class Storage(SupportsMinMaxCharge):
         max_soc_charge = self.calculate_soc_max_charge(soc)
         max_power_charge = max_power_charge.clip(min=max_soc_charge)
 
-        # convert to single values if only one value is left
-        if len(min_power_charge) == 1:
-            min_power_charge = min_power_charge[0]
-        if len(max_power_charge) == 1:
-            max_power_charge = max_power_charge[0]
-
         return min_power_charge, max_power_charge
 
     def calculate_min_max_discharge(
@@ -376,12 +370,6 @@ class Storage(SupportsMinMaxCharge):
             soc = self.outputs["soc"].at[start]
         max_soc_discharge = self.calculate_soc_max_discharge(soc)
         max_power_discharge = max_power_discharge.clip(max=max_soc_discharge)
-
-        # convert to single values if only one value is left
-        if len(min_power_discharge) == 1:
-            min_power_discharge = min_power_discharge[0]
-        if len(max_power_discharge) == 1:
-            max_power_discharge = max_power_discharge[0]
 
         return min_power_discharge, max_power_discharge
 
