### 2025-02-26 b5e1175
marginal cost fixed, plot code, simulation horizon
