+++ examples/inputs/example_01h/config.yaml
@@ -4,7 +4,7 @@
 
 eom:
   start_date: 2024-06-01 00:00
-  end_date: 2024-06-29 23:45
+  end_date: 2024-06-03 23:45
   time_step: 15min
   save_frequency_hours: 24
 
