### 2025-02-26 cdfc9f0
Refactoring, plot fixes.
