### 2025-02-27 1b22431
- fix wrong handling of early stopping when avg_rewards are in negative
