### 2025-02-27 144bb24
- play around with reward function
