### 2025-02-27 b823506
ready configs for test sim
