+++ examples/inputs/example_02a/config.yaml
@@ -1,138 +1,170 @@
-# SPDX-FileCopyrightText: ASSUME Developers
-#
-# SPDX-License-Identifier: AGPL-3.0-or-later
-
 base:
-  start_date: 2019-03-01 00:00
-  end_date: 2019-03-31 00:00
-  time_step: 1h
-  learning_mode: true
-  save_frequency_hours: null
-
+  end_date: 2019-07-31 00:00
   learning_config:
-    continue_learning: false
-    trained_policies_save_path: null
-    max_bid_price: 100
+    action_noise_schedule: linear
     algorithm: matd3
+    batch_size: 256
+    continue_learning: true
+    device: cpu
+    episodes_collecting_initial_experience: 5
+    gamma: 0.99
+    gradient_steps: 24
     learning_rate: 0.001
     learning_rate_schedule: linear
-    training_episodes: 50
-    episodes_collecting_initial_experience: 5
-    train_freq: 24h
-    gradient_steps: 24
-    batch_size: 256
-    gamma: 0.99
-    device: cpu
-    action_noise_schedule: linear
-    noise_sigma: 0.1
-    noise_scale: 1
+    max_bid_price: 100
     noise_dt: 1
+    noise_scale: 1
+    noise_sigma: 0.1
+    train_freq: 24h
+    trained_policies_load_path: learned_strategies/example_02a_base/last_policies
+    trained_policies_save_path: null
+    training_episodes: 30
     validation_episodes_interval: 5
-
+  learning_mode: false
   markets_config:
     EOM:
-      operator: EOM_operator
-      product_type: energy
-      start_date: 2019-03-01 00:00
-      products:
-        - duration: 1h
-          count: 1
-          first_delivery: 1h
-      opening_frequency: 1h
-      opening_duration: 1h
-      volume_unit: MWh
-      maximum_bid_volume: 100000
-      maximum_bid_price: 3000
-      minimum_bid_price: -500
-      price_unit: EUR/MWh
       market_mechanism: pay_as_clear
-
-base_lstm:
-  start_date: 2019-03-01 00:00
-  end_date: 2019-03-31 00:00
-  time_step: 1h
-  learning_mode: true
+      maximum_bid_price: 3000
+      maximum_bid_volume: 100000
+      minimum_bid_price: -500
+      opening_duration: 1h
+      opening_frequency: 1h
+      operator: EOM_operator
+      price_unit: EUR/MWh
+      product_type: energy
+      products:
+      - count: 1
+        duration: 1h
+        first_delivery: 1h
+      start_date: 2019-03-01 00:00
+      volume_unit: MWh
   save_frequency_hours: null
-
+  start_date: 2019-03-01 00:00
+  time_step: 1h
+base_lstm:
+  end_date: 2019-03-31 00:00
   learning_config:
-    continue_learning: false
-    trained_policies_save_path: null
-    max_bid_price: 100
+    actor_architecture: lstm
     algorithm: matd3
-    learning_rate: 0.001
-    training_episodes: 50
-    episodes_collecting_initial_experience: 5
-    train_freq: 24h
-    gradient_steps: 24
     batch_size: 256
-    gamma: 0.99
+    continue_learning: false
     device: cpu
-    noise_sigma: 0.1
-    noise_scale: 1
-    noise_dt: 1
-    validation_episodes_interval: 5
     early_stopping_steps: 10
     early_stopping_threshold: 0.05
-    actor_architecture: lstm
-
-  markets_config:
-    EOM:
-      operator: EOM_operator
-      product_type: energy
-      start_date: 2019-03-01 00:00
-      products:
-        - duration: 1h
-          count: 1
-          first_delivery: 1h
[…]
