### 2025-02-27 07af900
- add storage example
