+++ examples/examples.py
@@ -93,6 +93,10 @@ available_examples = {
     },
     "large_2019_rl": {"scenario": "example_03a", "study_case": "base_case_2019"},
     "large_2021_rl": {"scenario": "example_03b", "study_case": "base_case_2021"},
+    "large_2019_rl_storage": {
+        "scenario": "example_03c",
+        "study_case": "base_case_2019_with_storage",
+    },
 }
 
 
