+++ examples/inputs/example_03c/config.yaml
@@ -2,8 +2,7 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
-base_case_2019:
-  simulation_id: learning_2019_with_storage
+base_case_2019_with_storage:
   start_date: 2019-03-01 00:00
   end_date: 2019-03-31 00:00
   time_step: 1h
@@ -15,7 +14,7 @@ base_case_2019:
     trained_policies_save_path: null
     max_bid_price: 100
     algorithm: matd3
-    learning_rate: 0.0001
+    learning_rate: 0.0003
     # learning_rate_schedule: linear
     training_episodes: 100
     episodes_collecting_initial_experience: 5
@@ -24,7 +23,7 @@ base_case_2019:
     batch_size: 256
     gamma: 0.999
     device: cuda:0
-    action_noise_schedule: linear
+    # action_noise_schedule: linear
     noise_sigma: 0.1
     noise_scale: 1
     noise_dt: 1
