### 2025-02-27 d17e0f4
-add storages to learning power plants
