### 2025-02-28 f4a2d85
- orderbook is not chronologically sorted, searching for first and last start time is necessary - stick to "typical" deduction of end time for consistency - update outputs["energy"] (according to soc constraints) for data consistency - update doc strings
