### 2025-03-02 fc56b3a
Adapt docker compose
