### 2025-03-03 82e2124
- add imports and exports to the forecasts calculation - remove wrong pp from examples_03a - improve logging at the beginning of the simulation
