### 2025-03-03 9c5cdd1
Add 2021 DE exchange data to example 3b (#558)
