### 2025-03-04 f333ec7
fix issue if no db is set on tensorflow while learning
