### 2025-03-04 62c4afc
Improve learning logging into the database (#554)
We still have the issue, that the grafana dashboard is not usable when
running large simulations, and this is actually very useful and
important when trying to get large trainings to work, especially with
different agents. I have identified, that the problem is due to the
query of all the simulation ids and processing and so on, it just gets
stuck when I have multiple big runs in the database. Hence I have
decided to completely go away from the idea of changing simulation_id
during training and logging only using episode and mode, which improve
[…]
