+++ assume/scenario/loader_csv.py
@@ -461,9 +461,9 @@ def load_config_and_create_forecaster(
         study_case = list(config.keys())[0]
     config = config[study_case]
 
-    sim_id = config.get("simulation_id", f"{scenario}_{study_case}")
+    simulation_id = config.get("simulation_id", f"{scenario}_{study_case}")
 
-    logger.info(f"Simulation ID: {sim_id}")
+    logger.info(f"Simulation ID: {simulation_id}")
 
     start = pd.Timestamp(config["start_date"])
     end = pd.Timestamp(config["end_date"])
@@ -570,7 +570,7 @@ def load_config_and_create_forecaster(
 
     return {
         "config": config,
-        "sim_id": sim_id,
+        "simulation_id": simulation_id,
         "path": path,
         "start": start,
         "end": end,
@@ -585,7 +585,7 @@ def load_config_and_create_forecaster(
 
 def setup_world(
     world: World,
-    perform_evaluation: bool = False,
+    evaluation_mode: bool = False,
     terminate_learning: bool = False,
     episode: int = 1,
     eval_episode: int = 1,
@@ -598,7 +598,7 @@ def setup_world(
     Args:
         world (World): An instance of the World class representing the simulation environment.
         scenario_data (dict): A dictionary containing the configuration and loaded files for the scenario and study case.
-        perform_evaluation (bool, optional): A flag indicating whether evaluation should be performed. Defaults to False.
+        evaluation_mode (bool, optional): A flag indicating whether evaluation should be performed. Defaults to False.
         terminate_learning (bool, optional): An automatically set flag indicating that we terminated the learning process now, either because we reach the end of the episode iteration or because we triggered an early stopping.
         episode (int, optional): The episode number for learning. Defaults to 1.
         eval_episode (int, optional): The episode number for evaluation. Defaults to 1.
@@ -610,7 +610,7 @@ def setup_world(
     # make a deep copy of the scenario data to avoid changing the original data
     scenario_data = copy.deepcopy(world.scenario_data)
 
-    sim_id = scenario_data["sim_id"]
+    simulation_id = scenario_data["simulation_id"]
     config = scenario_data["config"]
     start = scenario_data["start"]
     end = scenario_data["end"]
@@ -653,31 +653,23 @@ def setup_world(
     bidding_strategy_params = config.get("bidding_strategy_params", {})
 
     learning_config["learning_mode"] = config.get("learning_mode", False)
-    learning_config["perform_evaluation"] = perform_evaluation
+    learning_config["evaluation_mode"] = evaluation_mode
 
     if terminate_learning:
         learning_config["learning_mode"] = False
-        learning_config["perform_evaluation"] = False
+        learning_config["evaluation_mode"] = False
 
     if not learning_config.get("trained_policies_save_path"):
-        learning_config["trained_policies_save_path"] = f"learned_strategies/{sim_id}"
+        learning_config["trained_policies_save_path"] = (
+            f"learned_strategies/{simulation_id}"
+        )
 
     if not learning_config.get("trained_policies_load_path"):
         learning_config["trained_policies_load_path"] = (
-            f"learned_strategies/{sim_id}/avg_reward_eval_policies"
+            f"learned_strategies/{simulation_id}/avg_reward_eval_policies"
         )
 
     config = replace_paths(config, scenario_data["path"])
-
-    if learning_config.get("learning_mode", False) and not learning_config.get(
-        "perform_evaluation", False
-    ):
-        sim_id = f"{sim_id}_{episode}"
-
-    elif learning_config.get("learning_mode", False) and learning_config.get(
-        "perform_evaluation", False
-    ):
[…]
