### 2025-03-06 ea9c6c4
Update docs to include bidding strategies, and change bidding strategy names
