+++ examples/distributed_simulation/world_agent.py
@@ -36,7 +36,7 @@ async def create_worker(
             {
                 "min_power": 200 / (n_proc * m_agents),
                 "max_power": 1000 / (n_proc * m_agents),
-                "bidding_strategies": {"EOM": "naive_eom"},
+                "bidding_strategies": {"EOM": "naive_eom_powerplant"},
                 "technology": "nuclear",
             },
             nuclear_forecast,
