+++ assume/strategies/__init__.py
@@ -3,13 +3,20 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 from assume.common.base import BaseStrategy, LearningStrategy
-from assume.strategies.advanced_orders import flexableEOMBlock, flexableEOMLinked
+from assume.strategies.standard_advanced_orders import (
+    EOMBlockPowerplant,
+    EOMLinkedPowerplant,
+)
 from assume.strategies.extended import OTCStrategy
-from assume.strategies.flexable import flexableEOM, flexableNegCRM, flexablePosCRM
-from assume.strategies.flexable_storage import (
-    flexableEOMStorage,
-    flexableNegCRMStorage,
-    flexablePosCRMStorage,
+from assume.strategies.standard_powerplant import (
+    EOMPowerplant,
+    NCRMPowerplant,
+    PCRMPowerplant,
+)
+from assume.strategies.standard_storage import (
+    EOMStorage,
+    NCRMStorage,
+    PCRMStorage,
 )
 from assume.strategies.naive_strategies import (
     NaiveDADSMStrategy,
@@ -25,22 +32,23 @@ from assume.strategies.dmas_storage import DmasStorageStrategy
 
 
 bidding_strategies: dict[str, BaseStrategy] = {
-    "naive_eom": NaiveSingleBidStrategy,
-    "naive_dam": NaiveProfileStrategy,
-    "naive_pos_reserve": NaiveSingleBidStrategy,
-    "naive_neg_reserve": NaiveSingleBidStrategy,
+    "naive_eom_powerplant": NaiveSingleBidStrategy,
+    "naive_eom_demand": NaiveSingleBidStrategy,
+    "naive_eom_block_powerplant": NaiveProfileStrategy,
+    "naive_pcrm": NaiveSingleBidStrategy,
+    "naive_ncrm": NaiveSingleBidStrategy,
     "naive_exchange": NaiveExchangeStrategy,
     "otc_strategy": OTCStrategy,
-    "flexable_eom": flexableEOM,
-    "flexable_eom_block": flexableEOMBlock,
-    "flexable_eom_linked": flexableEOMLinked,
-    "flexable_neg_crm": flexableNegCRM,
-    "flexable_pos_crm": flexablePosCRM,
-    "flexable_eom_storage": flexableEOMStorage,
-    "flexable_neg_crm_storage": flexableNegCRMStorage,
-    "flexable_pos_crm_storage": flexablePosCRMStorage,
+    "eom_powerplant": EOMPowerplant,
+    "eom_block_powerplant": EOMBlockPowerplant,
+    "eom_linked_powerplant": EOMLinkedPowerplant,
+    "ncrm_powerplant": NCRMPowerplant,
+    "pcrm_powerplant": PCRMPowerplant,
+    "eom_storage": EOMStorage,
+    "ncrm_storage": NCRMStorage,
+    "pcrm_storage": PCRMStorage,
     "naive_redispatch": NaiveRedispatchStrategy,
-    "naive_da_dsm": NaiveDADSMStrategy,
+    "naive_eom_block_dsm": NaiveDADSMStrategy,
     "naive_redispatch_dsm": NaiveRedispatchDSMStrategy,
     "manual_strategy": SimpleManualTerminalStrategy,
     "dmas_powerplant": DmasPowerplantStrategy,
@@ -56,8 +64,8 @@ try:
         StorageRLStrategy,
     )
 
-    bidding_strategies["pp_learning"] = RLStrategy
-    bidding_strategies["storage_learning"] = StorageRLStrategy
+    bidding_strategies["learning_powerplant"] = RLStrategy
+    bidding_strategies["learning_storage"] = StorageRLStrategy
[…]
