+++ tests/test_standard_advanced_orders_strategies.py
@@ -9,8 +9,8 @@ import pytest
 
 from assume.common.forecasts import NaiveForecast
 from assume.strategies import (
-    flexableEOMBlock,
-    flexableEOMLinked,
+    EOMBlockPowerplant,
+    EOMLinkedPowerplant,
 )
 from assume.units import PowerPlant
 
@@ -41,7 +41,7 @@ def power_plant() -> PowerPlant:
 def test_eom_with_blocks(mock_market_config, power_plant):
     power_plant.ramp_up = 400
     product_index = pd.date_range("2023-07-01", periods=24, freq="h")
-    strategy = flexableEOMBlock()
+    strategy = EOMBlockPowerplant()
     mc = mock_market_config
     mc.product_type = "energy_eom"
     product_tuples = [
@@ -92,7 +92,7 @@ def test_eom_with_blocks(mock_market_config, power_plant):
 def test_eom_with_links(mock_market_config, power_plant):
     power_plant.ramp_up = 400
     product_index = pd.date_range("2023-07-01", periods=24, freq="h")
-    strategy = flexableEOMLinked()
+    strategy = EOMLinkedPowerplant()
     mc = mock_market_config
     mc.product_type = "energy_eom"
     product_tuples = [

[…]
