### 2025-03-06 4877c66
Format Bidding Strategy doc file
