+++ docs/source/bidding_strategies.rst
@@ -1,0 +1,175 @@
+.. SPDX-FileCopyrightText: ASSUME Developers
+..
+.. SPDX-License-Identifier: AGPL-3.0-or-later
+
+Bidding Strategies
+=====================
+
+Overview
+-------------
+
+The ASSUME framework provides multiple options in terms of Bidding Strategy methodologies:
+
+ ============================= =============================================================
+  Bidding Strategy Category     Description
+ ============================= =============================================================
+  Naive                         Basic form of bids, based on participating in market mechanism that follows the Merit Order principle.
+                                External market factors or forecasts are not accounted for, so the bid volume is of the order of its max. power capacity (given ramping constraints)
+                                and the price is its marginal cost.
+  Standard                      These bidding strategies, based on the flexABLE methodology [`Qussous et al. 2022 <https://doi.org/10.3390/en15020494>`_],
+                                are more refined compared to the Naive strategy. They incorporate market dynamics to a greater degree by forecasting market prices,
+                                operational history, potential power loss due to heat production,
+                                and taking into account the fact that a plant can make bids for multiple markets at the same time.
+  DMAS                          TODO
+  Learning                      A reinforcement learning (RL) approach designed to optimize bidding strategies for an Energy-Only Market. The agent's actions are two bid prices: one for
+                                the inflexible component (P_min) and another for the flexible component (P_max - P_min) of a unit's capacity.
+                                The bids are informed by 50 observations, which include forecasted residual load, forecasted price, total capacity,
+                                and marginal cost, all contributing to decision-making. Noise can be added to the action to encourage exploration and novelty.
+
+                                The reward system includes profits from accepted bids, operational costs, opportunity costs (penalizing underutilized capacity),
+                                and a regret term to minimize missed revenue opportunities. This approach encourages full utilization of the unit's capacity.
+  Other                         Other bidding methods not belonging to a specific category.
+ ============================= =============================================================
+
+For each Bidding Strategy Category, there are different Bidding Strategy Types based on the unit type (Power Plant, Storage, Demand-Side Management (DSM)) that is making the bid,
+and the Market Type the bid is made for (Energy-Only Market (EOM), Positive/Negative Control Reserve Market (PCRM/NCRM), Redispatch Market). The following sections describe these
+Bidding Strategy Types per category.
+
+Naive
+-------------
+
+ ================================================= ================================================ =============================================================
+  Bidding Strategy Type                             Bidding Strategy ID                              Description
+ ================================================= ================================================ =============================================================
+  Naive Energy Only Market Power Plant Bid          naive_eom_powerplant                             This bid is the generic formulation of a bid in a merit order
+                                                                                                     market configuration at any one timepoint (hour) for the EOM,
+                                                                                                     where it uses marginal cost for its bid price, and max. power
+                                                                                                     output (given ramping constraints) as its volume.
+  Naive Energy Only Market Demand Bid               naive_eom_demand                                 The demand bid uses the same strategy as :code:`naive_eom_powerplant`, it is
+                                                                                                     realised as a price-inelastic demand bid by setting
+                                                                                                     the bid price very high, and the volume is negative as it is consuming rather than producing power.
+  Naive Day-Ahead Market Bid                        naive_eom_block_powerplant                       Similar to :code:`naive_eom_powerplant`, however it is a block bid for 24 hours to
+                                                                                                     simulate a bid for the Day-Ahead market.
+  Naive Positive Control Reserve Market Bid         naive_pcrm                                       This bid uses the same strategy as :code:`naive_eom_powerplant`,
+                                                                                                     however the bid is placed in the positive CRM.
+  Naive Negative Control Reserve Market Bid         naive_ncrm                                       This bid uses the same strategy as :code:`naive_eom_powerplant`,
[…]
