### 2025-03-06 84e746f
Small edits to bidding_strategies doc
