### 2025-03-06 9f7e03f
- change naming - remove redundant strategies
