+++ examples/world_script.py
@@ -56,7 +56,7 @@ def init(world, n=1):
         {
             "min_power": 0,
             "max_power": 1000,
-            "bidding_strategies": {"EOM": "naive_eom_demand"},
+            "bidding_strategies": {"EOM": "naive"},
             "technology": "demand",
         },
         NaiveForecast(index, demand=1000),
@@ -72,7 +72,7 @@ def init(world, n=1):
             {
                 "min_power": 200 / n,
                 "max_power": 1000 / n,
-                "bidding_strategies": {"EOM": "naive_eom_powerplant"},
+                "bidding_strategies": {"EOM": "naive"},
                 "technology": "nuclear",
             },
             nuclear_forecast,
