+++ tests/test_rl_strategies.py
@@ -9,7 +9,7 @@ import pytest
 
 try:
     from assume.reinforcement_learning.learning_role import LearningConfig
-    from assume.strategies.learning_strategies import RLStrategy
+    from assume.strategies.learning_strategies import LearningEOMPowerplant
 except ImportError:
     pass
 
@@ -43,7 +43,7 @@ def power_plant_mcp() -> PowerPlant:
         min_power=200,
         efficiency=0.5,
         additional_cost=10,
-        bidding_strategies={"EOM": RLStrategy(**learning_config)},
+        bidding_strategies={"EOM": LearningEOMPowerplant(**learning_config)},
         fuel_type="lignite",
         emission_factor=0.5,
         forecaster=ff,
@@ -74,7 +74,7 @@ def power_plant_lstm() -> PowerPlant:
         min_power=200,
         efficiency=0.5,
         additional_cost=10,
-        bidding_strategies={"EOM": RLStrategy(**learning_config)},
+        bidding_strategies={"EOM": LearningEOMPowerplant(**learning_config)},
         fuel_type="lignite",
         emission_factor=0.5,
         forecaster=ff,
