+++ tests/test_standard_advanced_orders_strategies.py
@@ -9,8 +9,7 @@ import pytest
 
 from assume.common.forecasts import NaiveForecast
 from assume.strategies import (
-    EOMBlockPowerplant,
-    EOMLinkedPowerplant,
+    StandardProfileEOMPowerplant,
 )
 from assume.units import PowerPlant
 
@@ -38,61 +37,10 @@ def power_plant() -> PowerPlant:
     )
 
 
-def test_eom_with_blocks(mock_market_config, power_plant):
-    power_plant.ramp_up = 400
-    product_index = pd.date_range("2023-07-01", periods=24, freq="h")
-    strategy = EOMBlockPowerplant()
-    mc = mock_market_config
-    mc.product_type = "energy_eom"
-    product_tuples = [
-        (start, start + pd.Timedelta(hours=1), None) for start in product_index
-    ]
-    bids = strategy.calculate_bids(power_plant, mc, product_tuples=product_tuples)
-    assert len(bids) == 25
-    assert bids[0]["price"] == 40
-    assert bids[0]["volume"] == 200
-    assert bids[1]["price"] == 40
-    assert bids[1]["volume"] == 600
-    assert bids[2]["price"] == 40
-    assert bids[2]["volume"] == 800
-    assert bids[-1]["price"] == 40
-    assert bids[-1]["volume"] == {time: 200 for time in product_index}
-
-    # add min_down_time, min_operating_time
-    power_plant.index = pd.date_range("2023-07-01", periods=48, freq="h")
-    power_plant.min_down_time = 2
-    power_plant.min_operating_time = 3
-    product_index = pd.date_range("2023-07-02", periods=24, freq="h")
-    product_tuples = [
-        (start, start + pd.Timedelta(hours=1), None) for start in product_index
-    ]
-
-    # set the outputs before the first bid to 0
-    power_plant.outputs["energy"][power_plant.index[0:24]] = 0
-    assert power_plant.get_operation_time(product_index[0]) == -3
-
-    # test min_down_time, set op_time to -1
-    power_plant.outputs["energy"][power_plant.index[10:23]] = 200
-    assert power_plant.get_operation_time(product_index[0]) == -1
-
-    bids = strategy.calculate_bids(power_plant, mc, product_tuples=product_tuples)
-    assert len(bids) == 24
-    assert bids[0]["price"] == 40
-    assert bids[0]["volume"] == 200
-    assert bids[1]["price"] == 40
-    assert bids[1]["volume"] == 600
-    assert bids[2]["price"] == 40
-    assert bids[2]["volume"] == 800
-    assert bids[-1]["price"] == 40
-    assert bids[-1]["volume"] == {
-        product_index[i]: 0 if i == 0 else 200 for i in range(len(product_index))
-    }
-
-
 def test_eom_with_links(mock_market_config, power_plant):
     power_plant.ramp_up = 400
     product_index = pd.date_range("2023-07-01", periods=24, freq="h")
-    strategy = EOMLinkedPowerplant()
+    strategy = StandardProfileEOMPowerplant()
     mc = mock_market_config
     mc.product_type = "energy_eom"
     product_tuples = [
