+++ assume/strategies/__init__.py
@@ -4,19 +4,18 @@
 
 from assume.common.base import BaseStrategy, LearningStrategy
 from assume.strategies.standard_advanced_orders import (
-    EOMBlockPowerplant,
-    EOMLinkedPowerplant,
+    StandardProfileEOMPowerplant,
 )
 from assume.strategies.extended import OTCStrategy
 from assume.strategies.standard_powerplant import (
-    EOMPowerplant,
-    NCRMPowerplant,
-    PCRMPowerplant,
+    StandardEOMPowerplant,
+    StandardNCRMPowerplant,
+    StandardPCRMPowerplant,
 )
 from assume.strategies.standard_storage import (
-    EOMStorage,
-    NCRMStorage,
-    PCRMStorage,
+    StandardEOMStorage,
+    StandardNCRMStorage,
+    StandardPCRMStorage,
 )
 from assume.strategies.naive_strategies import (
     NaiveDADSMStrategy,
@@ -32,41 +31,37 @@ from assume.strategies.dmas_storage import DmasStorageStrategy
 
 
 bidding_strategies: dict[str, BaseStrategy] = {
-    "naive_eom_powerplant": NaiveSingleBidStrategy,
-    "naive_eom_demand": NaiveSingleBidStrategy,
-    "naive_eom_block_powerplant": NaiveProfileStrategy,
-    "naive_pcrm": NaiveSingleBidStrategy,
-    "naive_ncrm": NaiveSingleBidStrategy,
+    "naive": NaiveSingleBidStrategy,
+    "naive_profile": NaiveProfileStrategy,
+    "naive_profile_dsm": NaiveDADSMStrategy,
     "naive_exchange": NaiveExchangeStrategy,
-    "otc_strategy": OTCStrategy,
-    "eom_powerplant": EOMPowerplant,
-    "eom_block_powerplant": EOMBlockPowerplant,
-    "eom_linked_powerplant": EOMLinkedPowerplant,
-    "ncrm_powerplant": NCRMPowerplant,
-    "pcrm_powerplant": PCRMPowerplant,
-    "eom_storage": EOMStorage,
-    "ncrm_storage": NCRMStorage,
-    "pcrm_storage": PCRMStorage,
     "naive_redispatch": NaiveRedispatchStrategy,
-    "naive_eom_block_dsm": NaiveDADSMStrategy,
     "naive_redispatch_dsm": NaiveRedispatchDSMStrategy,
-    "manual_strategy": SimpleManualTerminalStrategy,
+    "standard_eom_powerplant": StandardEOMPowerplant,
+    "standard_profile_eom_powerplant": StandardProfileEOMPowerplant,
+    "standard_pcrm_powerplant": StandardPCRMPowerplant,
+    "standard_ncrm_powerplant": StandardNCRMPowerplant,
+    "standard_eom_storage": StandardEOMStorage,
+    "standard_pcrm_storage": StandardPCRMStorage,
+    "standard_ncrm_storage": StandardNCRMStorage,
     "dmas_powerplant": DmasPowerplantStrategy,
     "dmas_storage": DmasStorageStrategy,
+    "misc_otc": OTCStrategy,
+    "misc_manual": SimpleManualTerminalStrategy,
 }
 
 try:
     from assume.strategies.learning_advanced_orders import (
-        RLAdvancedOrderStrategy,
+        LearningProfileEOMPowerplant,
     )
     from assume.strategies.learning_strategies import (
-        RLStrategy,
-        StorageRLStrategy,
+        LearningEOMPowerplant,
+        LearningEOMStorage,
     )
 
-    bidding_strategies["learning_powerplant"] = RLStrategy
-    bidding_strategies["learning_storage"] = StorageRLStrategy
[…]

[…]
