### 2025-03-07 b514925
- use single action instead of two actions - fix errors when action consists of a single action
