### 2025-03-12 64e2bf5
-improve no solvers logging
