### 2025-03-12 6c27732
- fixed error in collecting and storing rewards which lead to mixed up values and failed training - reversed to 12 hours horizon for forecast for storage units
