+++ assume/reinforcement_learning/learning_unit_operator.py
@@ -157,9 +157,24 @@ class RLUnitsOperator(UnitsOperator):
         obs_dim = self.learning_strategies["obs_dim"]
         act_dim = self.learning_strategies["act_dim"]
         device = self.learning_strategies["device"]
+
         learning_unit_count = len(self.rl_units)
 
-        values_len = len(self.rl_units[0].outputs["rl_rewards"])
+        # Determine the number of reward values for each unit
+        # This tells us how many complete transitions we have
+        values_len_set = {len(unit.outputs["rl_rewards"]) for unit in self.rl_units}
+
+        # Ensure all units have the same number of reward values
+        # If there is more than one unique length, it means some units have a different number of rewards,
+        # which is an error condition, so we raise an exception.
+        if len(values_len_set) > 1:
+            raise ValueError(
+                "The length of the reward values is not the same for all units"
+            )
+
+        # Since all units have the same length, extract the common length
+        values_len = values_len_set.pop()
+
         # return if no data is available
         if values_len == 0:
             return
@@ -175,9 +190,7 @@ class RLUnitsOperator(UnitsOperator):
         for i, unit in enumerate(self.rl_units):
             # Convert pandas Series to torch Tensor
             obs_tensor = th.stack(unit.outputs["rl_observations"][:values_len], dim=0)
-            actions_tensor = th.stack(
-                unit.outputs["rl_actions"][:values_len], dim=0
-            ).reshape(-1, act_dim)
+            actions_tensor = th.stack(unit.outputs["rl_actions"][:values_len], dim=0)
 
             all_observations[:, i, :] = obs_tensor
             all_actions[:, i, :] = actions_tensor
@@ -186,20 +199,11 @@ class RLUnitsOperator(UnitsOperator):
             # reset the outputs
             unit.reset_saved_rl_data()
 
-        # convert all_actions list of tensor to numpy 2D array
-        all_observations = (
-            all_observations.squeeze()
-            .cpu()
-            .numpy()
-            .reshape(-1, learning_unit_count, obs_dim)
-        )
-        all_actions = (
-            all_actions.squeeze()
-            .cpu()
-            .numpy()
-            .reshape(-1, learning_unit_count, act_dim)
-        )
-        all_rewards = np.array(all_rewards).reshape(-1, learning_unit_count)
+        all_observations = all_observations.numpy(force=True)
+        all_actions = all_actions.numpy(force=True)
+
+        all_rewards = np.array(all_rewards).T
+
         rl_agent_data = (all_observations, all_actions, all_rewards)
 
         learning_role_addr = self.context.data.get("learning_agent_addr")
