+++ tests/test_standard_powerplant_strategies.py
@@ -41,7 +41,7 @@ def power_plant() -> PowerPlant:
     )
 
 
-def test_eom_powerplant(mock_market_config, power_plant):
+def test_standard_eom_powerplant(mock_market_config, power_plant):
     end = datetime(2023, 7, 1, 1)
     strategy = StandardEOMPowerplantStrategy()
     mc = mock_market_config
@@ -72,7 +72,7 @@ def test_eom_powerplant(mock_market_config, power_plant):
     assert bids[1]["volume"] == 100
 
 
-def test_flexable_pos_reserve(mock_market_config, power_plant):
+def test_standard_pos_reserve(mock_market_config, power_plant):
     end = datetime(2023, 7, 1, 4, 0, 0)
     strategy = StandardPosCRMPowerplantStrategy()
     mc = mock_market_config
@@ -97,7 +97,7 @@ def test_flexable_pos_reserve(mock_market_config, power_plant):
     assert bids[0]["volume"] == 1000
 
 
-def test_flexable_neg_reserve(mock_market_config, power_plant):
+def test_standard_neg_reserve(mock_market_config, power_plant):
     end = datetime(2023, 7, 1, 4, 0, 0)
     strategy = StandardNegCRMPowerplantStrategy()
     mc = mock_market_config
