### 2025-03-12 7a29061
-increase foresight to 24 hours for better performance of storage learning - improve configs for 02c and 02e examples - add an extra storage to the example_02e
