### 2025-03-13 2bad747
- add tests for data collection to prevent such error from happening again - adjust other tests to new structure
