+++ tests/test_drl_storage_strategy.py
@@ -144,9 +144,9 @@ def test_storage_rl_strategy_sell_bid(mock_market_config, storage_unit):
             expected_costs = (
                 10.0 * bid["volume"] * duration_hours
             )  # 10 * 500 * 1 = 5000
-            scaling_factor = (
-                0.1 / storage_unit.max_power_discharge
-            )  # 0.1 / 500 = 0.0002
+            scaling_factor = 1 / (
+                storage_unit.max_power_discharge * strategy.max_bid_price
+            )  # 1 / (500*100) = 0.00002
             expected_reward = (
                 expected_profit - expected_costs
             ) * scaling_factor  # (10000 - 5000) * 0.0002 = 1.0
@@ -182,8 +182,8 @@ def test_storage_rl_strategy_buy_bid(mock_market_config, storage_unit):
     # Instantiate the StorageRLStrategy
     strategy = storage_unit.bidding_strategies["EOM"]
 
-    # Define the 'buy' action: [0.3, -0.5] -> price=30, direction='buy'
-    buy_action = [0.3, -0.5]
+    # Define the 'buy' action: [-0.3] -> price=30, direction='buy'
+    buy_action = [-0.3]
 
     # Mock the get_actions method to return the buy action
     with patch.object(
@@ -205,13 +205,13 @@ def test_storage_rl_strategy_buy_bid(mock_market_config, storage_unit):
             bid = bids[0]
 
             # Assert the bid price is correctly scaled
-            expected_bid_price = buy_action[0] * strategy.max_bid_price  # 30.0
+            expected_bid_price = abs(buy_action[0]) * strategy.max_bid_price  # 30.0
             assert math.isclose(
                 bid["price"], expected_bid_price, abs_tol=1e3
             ), f"Expected bid price {expected_bid_price}, got {bid['price']}"
 
             # Assert the bid direction is 'buy' and volume is abs(max_power_charge) + 1e-6
-            expected_volume = storage_unit.max_power_charge + 1e-6  # 500 + 1e-6
+            expected_volume = storage_unit.max_power_charge - 1e-6  # 500 + 1e-6
             assert (
                 bid["volume"] == expected_volume
             ), f"Expected bid volume {expected_volume}, got {bid['volume']}"
@@ -236,9 +236,9 @@ def test_storage_rl_strategy_buy_bid(mock_market_config, storage_unit):
             expected_costs = (
                 15.0 * abs(bid["volume"]) * duration_hours
             )  # 15 * 500 * 1 = 7500
-            scaling_factor = (
-                0.1 / storage_unit.max_power_discharge
-            )  # 0.1 / 500 = 0.0002
+            scaling_factor = 1 / (
+                storage_unit.max_power_discharge * strategy.max_bid_price
+            )  # 1 / (500*100) = 0.00002
             expected_reward = (
                 expected_profit - expected_costs
             ) * scaling_factor  # (15000 - 7500) * 0.0002 = 1.5
@@ -275,8 +275,8 @@ def test_storage_rl_strategy_ignore_bid(mock_market_config, storage_unit):
     # Instantiate the StorageRLStrategy
     strategy = storage_unit.bidding_strategies["EOM"]
 
-    # Define the 'ignore' action: [0.0, 0.0] -> price=0, direction='ignore'
-    ignore_action = [0.0, 0.0]
+    # Define the 'ignore' action: [0.0] -> price=0, direction='ignore'
+    ignore_action = [0.0]
 
     # Mock the get_actions method to return the ignore action
     with patch.object(
@@ -322,12 +322,8 @@ def test_storage_rl_strategy_ignore_bid(mock_market_config, storage_unit):
             costs = storage_unit.outputs["total_costs"].loc[product_index]
 
             # Calculate expected values
[…]
