### 2025-03-13 266bf2e
- drop random sampling of strategies to update as this only degrades the performance after the fix in reward collection - revert config file
