### 2025-03-13 50f2ef3
- remove for loop from storage reward calculation as it can cause issues
