### 2025-03-13 597447b
- improve scaling for power plants to avoid too large gradients
