### 2025-03-13 eb6f54a
- improve configs
