### 2025-03-14 e6f395c
- improved config
