### 2025-03-14 6c87442
- added a new strategy compatible with training along storage units
