### 2025-03-15 fcdcbcf
- improve comments - fix storage action description - fix simple market clearing losing orders with volume of 0 - fix complex_clearing requiring bid_type even though a dafeult value is defined - remove ignore bid from storage learning strategy as it has no use and hinders performance
