### 2025-03-16 b07b046
add test to multi hour storage and soc test
