+++ tests/test_storage.py
@@ -16,7 +16,7 @@ from assume.units import Storage
 
 @pytest.fixture
 def storage_unit() -> Storage:
-    index = pd.date_range("2022-01-01", periods=5, freq="h")
+    index = pd.date_range("2022-01-01", periods=4, freq="h")
     forecaster = NaiveForecast(index, availability=1, price_forecast=50)
     return Storage(
         id="Test_Storage",
@@ -59,23 +59,23 @@ def test_reset_function(storage_unit):
     # check if total_power_output is reset
     assert (
         storage_unit.outputs["energy"]
-        == pd.Series(0.0, index=pd.date_range("2022-01-01", periods=5, freq="h"))
+        == pd.Series(0.0, index=pd.date_range("2022-01-01", periods=4, freq="h"))
     ).all()
 
     # the same for pos and neg capacity reserve
     assert (
         storage_unit.outputs["pos_capacity"]
-        == pd.Series(0.0, index=pd.date_range("2022-01-01", periods=5, freq="h"))
+        == pd.Series(0.0, index=pd.date_range("2022-01-01", periods=4, freq="h"))
     ).all()
     assert (
         storage_unit.outputs["neg_capacity"]
-        == pd.Series(0.0, index=pd.date_range("2022-01-01", periods=5, freq="h"))
+        == pd.Series(0.0, index=pd.date_range("2022-01-01", periods=4, freq="h"))
     ).all()
 
     # check if state of charge (soc) is reset correctly
     assert (
         storage_unit.outputs["soc"]
-        == pd.Series(500.0, index=pd.date_range("2022-01-01", periods=5, freq="h"))
+        == pd.Series(500.0, index=pd.date_range("2022-01-01", periods=4, freq="h"))
     ).all()
 
 
@@ -85,25 +85,19 @@ def test_calculate_operational_window(storage_unit):
     min_power_discharge, max_power_discharge = storage_unit.calculate_min_max_discharge(
         start, end
     )
-    min_power_discharge, max_power_discharge = (
-        min_power_discharge[0],
-        max_power_discharge[0],
-    )
-    cost_discharge = storage_unit.calculate_marginal_cost(start, max_power_discharge)
+    cost_discharge = storage_unit.calculate_marginal_cost(start, max_power_discharge[0])
 
-    assert min_power_discharge == 0
-    assert max_power_discharge == 100
+    assert min_power_discharge[0] == 0
+    assert max_power_discharge[0] == 100
     assert cost_discharge == 4 / 0.95
 
     min_power_charge, max_power_charge = storage_unit.calculate_min_max_charge(
         start, end
     )
-    min_power_charge, max_power_charge = min_power_charge[0], max_power_charge[0]
+    cost_charge = storage_unit.calculate_marginal_cost(start, max_power_charge[0])
 
-    cost_charge = storage_unit.calculate_marginal_cost(start, max_power_charge)
-
-    assert min_power_charge == 0
-    assert max_power_charge == -100
+    assert min_power_charge[0] == 0
+    assert max_power_charge[0] == -100
     assert math.isclose(cost_charge, 3 / 0.9)
 
     assert storage_unit.outputs["energy"].at[start] == 0
@@ -115,19 +109,14 @@ def test_calculate_operational_window(storage_unit):
     min_power_charge, max_power_charge = storage_unit.calculate_min_max_charge(
         start, end
     )
-    min_power_charge, max_power_charge = min_power_charge[0], max_power_charge[0]
-    assert min_power_charge == -40
-    assert max_power_charge == -60
+    assert min_power_charge[0] == -40
+    assert max_power_charge[0] == -60
 
[…]
