+++ assume/common/base.py
@@ -141,7 +141,7 @@ class BaseUnit:
     ) -> None:
         """
         Iterates through the orderbook, adding the accepted volumes to the corresponding time slots
-        in the dispatch plan. Updates the SOC for storage units.
+        in the dispatch plan.
 
         Args:
             marketconfig (MarketConfig): The market configuration.
@@ -171,47 +171,6 @@ class BaseUnit:
             self.outputs[f"{product_type}_accepted_price"].loc[start:end_excl] = (
                 accepted_price
             )
-
-        # also update the SOC if the unit is a storage unit
-        if isinstance(self, SupportsMinMaxCharge):
-            start = min(order["start_time"] for order in orderbook)
-            end = max(order["end_time"] for order in orderbook)
-            # end includes the end of the last product, to get the last products' start time we deduct the frequency once
-            end_excl = end - self.index.freq
-            time_delta = self.index.freq / timedelta(hours=1)
-
-            for t in self.index[start:end_excl]:
-                next_t = t + self.index.freq
-                # continue if it is the last time step
-                if next_t not in self.index:
-                    continue
-                current_power = self.outputs["energy"].at[t]
-
-                # calculate the change in state of charge
-                delta_soc = 0
-                soc = self.outputs["soc"].at[t]
-
-                # discharging
-                if current_power > 0:
-                    max_soc_discharge = self.calculate_soc_max_discharge(soc)
-
-                    if current_power > max_soc_discharge:
-                        current_power = max_soc_discharge
-
-                    delta_soc = -current_power * time_delta / self.efficiency_discharge
-
-                # charging
-                elif current_power < 0:
-                    max_soc_charge = self.calculate_soc_max_charge(soc)
-
-                    if current_power < max_soc_charge:
-                        current_power = max_soc_charge
-
-                    delta_soc = -current_power * time_delta * self.efficiency_charge
-
-                # update the values of the state of charge and the energy
-                self.outputs["soc"].at[next_t] = soc + delta_soc
-                self.outputs["energy"].at[t] = current_power
 
     def calculate_cashflow_and_reward(
         self,
@@ -677,6 +636,55 @@ class SupportsMinMaxCharge(BaseUnit):
                     0,
                 )
         return power_charge
+
+    def set_dispatch_plan(
+        self, marketconfig: MarketConfig, orderbook: Orderbook
+    ) -> None:
+        """Updates the SOC for storage units."""
+        super().set_dispatch_plan(marketconfig, orderbook)
+
+        if not orderbook:
+            return
+
+        # also update the SOC when setting the dispatch plan
+        start = min(order["start_time"] for order in orderbook)
+        end = max(order["end_time"] for order in orderbook)
+        # end includes the end of the last product, to get the last products' start time we deduct the frequency once
+        end_excl = end - self.index.freq
+        time_delta = self.index.freq / timedelta(hours=1)
+
+        for t in self.index[start:end_excl]:
+            next_t = t + self.index.freq
+            # continue if it is the last time step
+            if next_t not in self.index:
+                continue
+            current_power = self.outputs["energy"].at[t]
+
+            # calculate the change in state of charge
+            delta_soc = 0
+            soc = self.outputs["soc"].at[t]
+
+            # discharging
+            if current_power > 0:
+                max_soc_discharge = self.calculate_soc_max_discharge(soc)
+
+                if current_power > max_soc_discharge:
+                    current_power = max_soc_discharge
+
+                delta_soc = -current_power * time_delta / self.efficiency_discharge
+
+            # charging
+            elif current_power < 0:
+                max_soc_charge = self.calculate_soc_max_charge(soc)
+
+                if current_power < max_soc_charge:
[…]
