+++ assume/strategies/flexable_storage.py
@@ -167,6 +167,8 @@ class flexableEOMStorage(BaseStrategy):
             theoretic_SOC += delta_soc
             previous_power = bid_quantity + current_power
 
+        bids = self.remove_empty_bids(bids)
+
         return bids
 
     def calculate_reward(
