### 2025-03-17 968190e
adjusted doc string of storage strategy
