+++ assume/strategies/learning_strategies.py
@@ -647,11 +647,10 @@ class StorageRLStrategy(AbstractLearningStrategy):
     dictating the bid type. The storage agent can also decide to stay inactive by submitting a zero bid
     as this can be a valid strategy in some market conditions and also improves the learning process.
 
-    - **Bid Price**: The first action value determines the price at which the agent will bid.
-    - **Bid Direction**: The second action value defines the type of bid:
-        - If `action[1] < -0.1`: The agent submits a **sell bid**.
-        - If `action[1] > 0.1`: The agent submits a **buy bid**.
-        - Otherwise: The agent submits a **zero bid** to stay inactive.
+    - **Bid Price**: The one action value determines the price at which the agent will bid.
+    - **Bid Direction**: This is implictly set base don the action:
+        - If `action < 0`: The agent submits a **buy bid**.
+        - If `action >= 0`: The agent submits a **sell bid**.
 
     Rewards are based on the profit generated by the agent's market bids, with sell bids contributing
     positive profit and buy bids contributing negative profit. Additional components in the reward
