### 2025-03-17 b34ac24
- switch to single bid strategies
