### 2025-03-17 ed604fa
-fix clearing returning double - fix notebooks
