+++ assume/reinforcement_learning/learning_unit_operator.py
@@ -199,13 +199,7 @@ class RLUnitsOperator(UnitsOperator):
             all_observations[:, i, :] = obs_tensor
             all_actions[:, i, :] = actions_tensor
             all_rewards.append(unit.outputs["rl_rewards"])
-            
-            # ensure length reward is one value smaller than observation and action othwerwise throw warning
-            if len(unit.outputs["rl_observations"]) != values_len + 1:
-                logger.warning(
-                    f" Unit {unit.id} has unexpected observation length: expected {values_len + 1}, got {len(unit.outputs['rl_observations'])}. You are risking mismatched experiences."
-                )
-   
+        
 
             # reset the outputs
             unit.reset_saved_rl_data()
