### 2025-03-18 df8e27d
- readded the current_power into loop to consider updates for all product tuepls if different markets are modelled - changed test to check set_dispatch_plan and execute_current_dispatch - docsting adjustments for bettwer understanding
