### 2025-03-18 dcab45b
- minor test change - adjust time stamps of testing between hour 0 and hour 1, since bids are made for hour 1 there should not be any changes to hour 0 as per the new understanding of the SoC
