+++ assume/units/depot.py
@@ -1,0 +1,155 @@
+# SPDX-FileCopyrightText: ASSUME Developers
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+
+import logging
+import pyomo.environ as pyo
+from assume.common.utils import str_to_bool
+from assume.common.base import SupportsMinMax
+from assume.common.forecasts import Forecaster
+from assume.units.dsm_load_shift import DSMFlex
+
+logger = logging.getLogger(__name__)
+
+class BusDepot(DSMFlex, SupportsMinMax):
+    """
+    Represents a bus depot managing multiple electric vehicle (EV) buses and charging stations.
+    This agent optimizes bus charging schedules based on electricity price, grid demand, and available flexibility.
+
+    The `BusDepot` class utilizes the Pyomo optimization library to determine the optimal
+    charging strategy that minimizes costs, grid impact, or maximizes renewable utilization.
+
+    Args:
+        id (str): Unique identifier for the bus depot.
+        unit_operator (str): Operator managing the bus depot.
+        bidding_strategies (dict): Strategies used for energy bidding in the market.
+        forecaster (Forecaster): A forecaster used to get key variables such as electricity prices.
+        charging_stations (dict[str, dict]): Charging stations available in the depot.
+        electric_vehicles (dict[str, dict]): Electric buses assigned to the depot.
+        objective (str, optional): Optimization objective ("min_cost", "min_grid_load", "max_RE_util"). Default is "min_cost".
+        flexibility_measure (str, optional): Metric used to assess the depot's flexibility. Default is "cost_based_load_shift".
+        max_power_capacity (float, optional): Maximum allowable power capacity. Default is 1000 kW.
+        node (str, optional): Network node where the depot is connected. Default is "bus_node".
+        location (tuple[float, float], optional): Geographic coordinates. Default is (0.0, 0.0).
+        **kwargs: Additional keyword arguments for custom configurations.
+
+    Attributes:
+        required_technologies (list): Required technologies for the depot.
+        optional_technologies (list): Optional technologies available.
+    """
+
+    required_technologies = []
+    optional_technologies = ["charging_station", "electric_vehicle"]
+
+    def __init__(
+        self,
+        id: str,
+        unit_operator: str,
+        bidding_strategies: dict,
+        forecaster: Forecaster,
+        components: dict[str, dict],
+        technology: str = "bus_depot",
+        is_prosumer: str = "No",
+        objective: str = "min_cost",
+        flexibility_measure: str = "cost_based_load_shift",
+        max_power_capacity: float = 1000,
+        node: str = "bus_node",
+        location: tuple[float, float] = (0.0, 0.0),
+        **kwargs,
+    ):
+        super().__init__(
+            id=id,
+            unit_operator=unit_operator,
+            technology=technology,
+            components=components,
+            bidding_strategies=bidding_strategies,
+            forecaster=forecaster,
+            node=node,
+            location=location,
+            **kwargs,
+        )
+
+        self.objective = objective
+        self.is_prosumer = str_to_bool(is_prosumer)
+        self.flexibility_measure = flexibility_measure
+        self.max_power_capacity = max_power_capacity
+        self.electricity_price = self.forecaster["price_EOM"]
+        self.bus_availability = self.forecaster["bus_availability"]
+        self.bus_soc_profile = self.forecaster["bus_soc_profile"]
+        self.renewable_availability = self.forecaster["renewable_availability"]
+
+        self.setup_model(presolve=True)
+        self.initialize_process_sequence()
+
+    def define_parameters(self):
+        """Defines parameters including electricity price, power capacity, and bus availability."""
[…]
