+++ assume/units/__init__.py
@@ -10,6 +10,7 @@ from assume.units.storage import Storage
 from assume.units.steel_plant import SteelPlant
 from assume.units.hydrogen_plant import HydrogenPlant
 from assume.units.building import Building
+from assume.units.depot import BusDepot
 from assume.units.dst_components import demand_side_technologies
 
 unit_types: dict[str, BaseUnit] = {
@@ -20,4 +21,5 @@ unit_types: dict[str, BaseUnit] = {
     "steel_plant": SteelPlant,
     "hydrogen_plant": HydrogenPlant,
     "building": Building,
+    "bus_depot": BusDepot,
 }
