### 2025-03-20 bfa73e2
- adjust test description - fix typos - revert database default in examples to local_db
