+++ tests/test_storage.py
@@ -412,7 +412,7 @@ def test_set_dispatch_plan_multi_hours(mock_market_config, storage_unit):
 
     bids = strategy.calculate_bids(storage_unit, mc, product_tuples=product_tuples)
     assert len(bids) == 2
-    # the second and third hour have are charging 60 MW
+    # continue discharging 100 MW in second and third hour
     assert bids[0]["start_time"] == datetime(2022, 1, 1, 1)
     assert bids[0]["volume"] == 100
     bids[0]["accepted_volume"] = 100
@@ -429,21 +429,21 @@ def test_set_dispatch_plan_multi_hours(mock_market_config, storage_unit):
     for i in range(2):
         s = datetime(2022, 1, 1, i + 1)
         s_next = datetime(2022, 1, 1, i + 2)
-        delta_soc_set_dispacth = (
+        delta_soc_set_dispatch = (
             storage_unit.outputs["soc"][s] - storage_unit.outputs["soc"][s_next]
         )
 
-        if delta_soc_set_dispacth <= 0:
-            delta_set_dispacth = (
+        if delta_soc_set_dispatch <= 0:
+            delta_set_dispatch = (
                 storage_unit.outputs["energy"][s] * storage_unit.efficiency_charge
             )
         else:
-            delta_set_dispacth = (
+            delta_set_dispatch = (
                 storage_unit.outputs["energy"][s] / storage_unit.efficiency_discharge
             )
-        assert math.isclose(delta_set_dispacth, delta_soc_set_dispacth)
+        assert math.isclose(delta_set_dispatch, delta_soc_set_dispatch)
 
-    # test if it is executed correctly, whichc should be the same with the mokc makret config only covering one market
+    # test if it is executed correctly, which should be the same with the mock market config only covering one market
     storage_unit.execute_current_dispatch(start, end)
 
     for i in range(2):
@@ -459,8 +459,8 @@ def test_set_dispatch_plan_multi_hours(mock_market_config, storage_unit):
             )
         assert math.isclose(delta, delta_soc)
 
-    # check that deltas are the same, which agina must be due to only one considered
-    assert math.isclose(delta_soc_set_dispacth, delta_soc)
+    # check that deltas are the same, which again must be due to only one considered market
+    assert math.isclose(delta_soc_set_dispatch, delta_soc)
 
 
 if __name__ == "__main__":
