### 2025-03-20 4b608bb
Fix storage strategies (#557)
FlexABLE storage strategies were acting weird when used with a day ahead
bidding. The problem was that the SOC was updated only later during
execute_current_dispatch and was not reflected for the next bidding
period, where the storage units assumed again an initial soc of 0. This
PR adds an soc update part to the set_dispatch function when used by
storage units. This also refactors the flexable storage strategy to make
viable bids on the day ahead market compared to before.

---------

Co-authored-by: mthede <marie.thederan@student.kit.edu>
Co-authored-by: Florian
[…]
