+++ assume/common/base.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 from collections import defaultdict
-from datetime import datetime
+from datetime import datetime, timedelta
 from typing import TypedDict
 
 import numpy as np
@@ -141,7 +141,7 @@ class BaseUnit:
     ) -> None:
         """
         Iterates through the orderbook, adding the accepted volumes to the corresponding time slots
-        in the dispatch plan. It then calculates the cashflow and the reward for the bidding strategies.
+        in the dispatch plan.
 
         Args:
             marketconfig (MarketConfig): The market configuration.
@@ -510,7 +510,7 @@ class SupportsMinMaxCharge(BaseUnit):
     efficiency_discharge: float
 
     def calculate_min_max_charge(
-        self, start: datetime, end: datetime, product_type="energy"
+        self, start: datetime, end: datetime, soc: float = None
     ) -> tuple[np.array, np.array]:
         """
         Calculates the min and max charging power for the given time period.
@@ -518,14 +518,14 @@ class SupportsMinMaxCharge(BaseUnit):
         Args:
             start (datetime.datetime): The start time of the dispatch.
             end (datetime.datetime): The end time of the dispatch.
-            product_type (str, optional): The product type of the unit. Defaults to "energy".
+            soc (float, optional): The current state-of-charge. Defaults to None.
 
         Returns:
             tuple[np.array, np.array]: The min and max charging power for the given time period.
         """
 
     def calculate_min_max_discharge(
-        self, start: datetime, end: datetime, product_type="energy"
+        self, start: datetime, end: datetime, soc: float = None
     ) -> tuple[np.array, np.array]:
         """
         Calculates the min and max discharging power for the given time period.
@@ -533,28 +533,11 @@ class SupportsMinMaxCharge(BaseUnit):
         Args:
             start (datetime.datetime): The start time of the dispatch.
             end (datetime.datetime): The end time of the dispatch.
-            product_type (str, optional): The product type of the unit. Defaults to "energy".
+            soc (float, optional): The current state-of-charge. Defaults to None.
 
         Returns:
             tuple[np.array, np.array]: The min and max discharging power for the given time period.
         """
-
-    def get_soc_before(self, dt: datetime) -> float:
-        """
-        Returns the State of Charge (SoC) before the given datetime.
-        If datetime is before the start of the index, the initial SoC is returned.
-        The SoC is a float between 0 and 1.
-
-        Args:
-            dt (datetime.datetime): The current datetime.
-
-        Returns:
-            float: The SoC before the given datetime.
-        """
-        if dt - self.index.freq < self.index[0]:
-            return self.initial_soc
-        else:
-            return self.outputs["soc"].at[dt - self.index.freq]
 
     def calculate_ramp_discharge(
         self,
@@ -653,6 +636,55 @@ class SupportsMinMaxCharge(BaseUnit):
                     0,
                 )
         return power_charge
+
+    def set_dispatch_plan(
+        self, marketconfig: MarketConfig, orderbook: Orderbook
+    ) -> None:
+        """Updates the SOC for storage units."""
+        super().set_dispatch_plan(marketconfig, orderbook)
+
+        if not orderbook:
+            return
+
+        # also update the SOC when setting the dispatch plan
+        start = min(order["start_time"] for order in orderbook)
+        end = max(order["end_time"] for order in orderbook)
+        # end includes the end of the last product, to get the last products' start time we deduct the frequency once
+        end_excl = end - self.index.freq
+        time_delta = self.index.freq / timedelta(hours=1)
+
+        for t in self.index[start:end_excl]:
+            next_t = t + self.index.freq
+            # continue if it is the last time step
[…]
