+++ docs/source/release_notes.rst
@@ -7,10 +7,14 @@ Release Notes
 #######################
 
 Upcoming Release
-=======================
+================
 .. warning::
   The features in this section are not released yet, but will be part of the next release! To use the features already you have to install the main branch,
   e.g. ``pip install git+https://github.com/assume-framework/assume``
+
+
+0.5.2 - (21th March 2025)
+=========================
 
 **New Features:**
 
