### 2025-03-24 61e1642
- add small strage and pp learning example
