+++ assume/units/steel_plant.py
@@ -3,7 +3,6 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import logging
-from datetime import datetime
 
 import pyomo.environ as pyo
 
@@ -35,7 +34,7 @@ class SteelPlant(DSMFlex, SupportsMinMax):
         technology (str, optional): The technology of the steel plant. Default is "steel_plant".
         components (dict, optional): A dictionary describing the components of the steel plant, such as Electrolyser, DRI Plant, DRI Storage, and Electric Arc Furnace. Default is an empty dictionary.
         objective (str, optional): The objective function for the steel plant, typically focused on minimizing variable costs. Default is "min_variable_cost".
-        flexibility_measure (str, optional): The flexibility measure for the steel plant, such as "cost_based_load_shift". Default is "max_load_shift".
+        flexibility_measure (str, optional): The flexibility measure for the steel plant, such as "cost_based_load_shift". Default is "cost_based_load_shift".
         demand (float, optional): The steel production demand, representing the amount of steel that needs to be produced. Default is 0.
         cost_tolerance (float, optional): The maximum allowable cost variation when shifting the load, used in flexibility measures. Default is 10.
         congestion_threshold (float, optional): The threshold for congestion management in the plant’s energy system. Default is 0.
@@ -62,7 +61,7 @@ class SteelPlant(DSMFlex, SupportsMinMax):
         components: dict[str, dict] = None,
         technology: str = "steel_plant",
         objective: str = "min_variable_cost",
-        flexibility_measure: str = "max_load_shift",
+        flexibility_measure: str = "cost_based_load_shift",
         demand: float = 0,
         cost_tolerance: float = 10,
         congestion_threshold: float = 0,
@@ -205,7 +204,33 @@ class SteelPlant(DSMFlex, SupportsMinMax):
 
     def initialize_process_sequence(self):
         """
-        Initializes the process sequence and constraints for the steel plant. Here, the components/ technologies are connected to establish a process for steel production
+        Establishes the process sequence and constraints for the steel plant, ensuring that
+        different components and technologies interact correctly to support steel production.
+
+        This function defines three key constraints:
+
+        1. **Direct Hydrogen Flow Constraint**:
+        - Ensures that hydrogen produced by the electrolyzer is either supplied directly
+            to the DRI (Direct Reduced Iron) plant or stored in hydrogen buffer storage.
+        - If storage is available, the constraint balances hydrogen inflow and outflow
+            between the electrolyzer, hydrogen storage, and the DRI plant.
+        - If no electrolyzer exists, it ensures that the DRI plant has an alternative hydrogen
+            source (i.e., imported hydrogen).
+
+        2. **Direct DRI Flow Constraint**:
+        - Regulates the flow of directly reduced iron (DRI) from the DRI plant to the Electric Arc
+            Furnace (EAF) or DRI storage.
+        - If DRI storage is present, it ensures that part of the produced DRI can be stored
+            for later use, maintaining balance between production, storage, and consumption in the EAF.
+        - If no storage is available, all produced DRI must go directly to the EAF.
+
+        3. **Material Flow Constraint from DRI Plant to Electric Arc Furnace**:
+        - Ensures that all DRI produced by the DRI plant is consumed by the Electric Arc Furnace.
+        - This constraint enforces that the total material output from the DRI plant must match
+            the required DRI input for the EAF, preventing material imbalances.
+
+        These constraints collectively ensure proper material and energy flow within the steel
+        production process, maintaining energy efficiency and operational feasibility.
         """
 
[…]
