### 2025-03-24 710e0dd
- Add dictionary linking old bidding strategy names to new names. - Update documetation - Fix tutorial 09 error
