+++ assume/strategies/learning_strategies.py
@@ -648,7 +648,7 @@ class LearningEOMStorageStrategy(AbstractLearningStrategy):
     as this can be a valid strategy in some market conditions and also improves the learning process.
 
     - **Bid Price**: The one action value determines the price at which the agent will bid.
-    - **Bid Direction**: This is implicitly set base don the action:
+    - **Bid Direction**: This is implicitly set based on the action:
         - If `action < 0`: The agent submits a **buy bid**.
         - If `action >= 0`: The agent submits a **sell bid**.
 
@@ -939,7 +939,7 @@ class LearningEOMStorageStrategy(AbstractLearningStrategy):
         # that the strategy is not designed for multiple orders and the market configuration should be adjusted
         if len(orderbook) > 1:
             raise ValueError(
-                "StorageRLStrategy is not designed for multiple orders. Please adjust the market configuration or the strategy."
+                "LearningEOMStorageStrategy is not designed for multiple orders. Please adjust the market configuration or the strategy."
             )
 
         order = orderbook[0]
