### 2025-03-26 55d596b
Remove the profile learning strategy, as it currently does not add value to the framework. It only introduces unnecessary weight and complexity by bidding a single price for the entire day, which is illogical. A new 'new day ahead' strategy will be developed in the future. - drop all files and use cases of this strategy
