+++ assume/units/building.py
@@ -113,6 +113,7 @@ class Building(DSMFlex, SupportsMinMax):
         self.natural_gas_price = self.forecaster["fuel_price_natural gas"]
         self.heat_demand = self.forecaster[f"{self.id}_heat_demand"]
         self.ev_load_profile = self.forecaster["ev_load_profile"]
+        self.ev_load_avilability = self.forecaster["ev_load_profile"]
         self.battery_load_profile = self.forecaster["battery_load_profile"]
         self.inflex_demand = self.forecaster[f"{self.id}_load_profile"]
 
@@ -145,6 +146,23 @@ class Building(DSMFlex, SupportsMinMax):
                 if profile_key.endswith("power_profile")
                 else "availability_profile"
             ] = pv_profile
+
+        # Configure EV plant power profile based on availability
+        if self.has_ev:
+            profile_key_ev = (
+                f"{self.id}_ev_power_profile"
+                if not str_to_bool(
+                    self.components["electric_vehicle"].get("charging_profile", "false")
+                )
+                else "availability_ev"
+            )
+            ev_profile = self.forecaster[profile_key_ev]
+            # Assign the aligned profile
+            self.components["electric_vehicle"][
+                "charging_profile"
+                if profile_key_ev.endswith("charging_profile")
+                else "availability_ev"
+            ] = ev_profile
 
         # Initialize the model
         self.setup_model(presolve=True)
