### 2025-03-27 bb10fe9
Fix Notebook 9: xRL
Fix the notebook 9 not working anymore due to updated packages and changes syntax.
- update versions
- update code
- test and apply pre-commit
