### 2025-03-27 51977cb
Update documentation, delete learning profile class
