+++ assume/world.py
@@ -575,9 +575,8 @@ class World:
                 else:
                     logger.warning(
                         f"""FutureWarning: Bidding strategy name {strategy} is deprecated and has been
-                        renamed to {deprecated_to_new_strategy_names[strategy]}.
-                        For the full list of updated names, check the updated names in the bidding_strategies dict
-                        and compare to old names in the deprecated_to_new_strategy_names dict in strategies.__init__.py"""
+                        renamed to {deprecated_to_new_strategy_names[strategy]}. For the full list of up-to-date strategies and names,
+                        please see the documentation https://assume.readthedocs.io/en/latest/bidding_strategies.html"""
                     )
                     strategy = deprecated_to_new_strategy_names[strategy]
 
