+++ assume/strategies/learning_advanced_orders.py
@@ -1,446 +1,0 @@
-# SPDX-FileCopyrightText: ASSUME Developers
-#
-# SPDX-License-Identifier: AGPL-3.0-or-later
-
-from datetime import datetime
-
-import numpy as np
-import torch as th
-
-from assume.common.base import SupportsMinMax
-from assume.common.market_objects import MarketConfig, Orderbook, Product
-from assume.common.utils import get_products_index
-from assume.strategies.learning_strategies import LearningEOMPowerplantStrategy
-
-
-class LearningProfileEOMPowerplantStrategy(LearningEOMPowerplantStrategy):
-    """
-    Reinforcement Learning Strategy for an Energy-Only-Market with simple hourly, block and linked orders.
-
-    Attributes:
-        foresight (int): Number of time steps to look ahead. Default 24.
-        max_bid_price (float): The maximum bid price.
-        max_demand (float): The maximum demand.
-        device (str): The device to run on.
-        float_type (str): The float type to use.
-        learning_mode (bool): Whether to use learning mode.
-        algorithm (str): RL algorithm. Defaults to "matd3".
-        actor_architecture_class (type[torch.nn.Module]): Actor network class. Defaults to "MLPActor".
-        actor (torch.nn.Module): The actor network.
-        order_types (list[str]): The list of order types to use (SB, LB, BB).
-        episodes_collecting_initial_experience (int): Number of episodes to collect initial experience.
-
-    Args:
-        *args: Variable length argument list.
-        **kwargs: Arbitrary keyword arguments.
-
-    Note:
-        This strategy is based on the strategy in standard_advanced_orders.py, but uses the actor network to determine the prices
-        instead of using the marginal costs as in standard_advanced_orders.py.
-        The two prices for flexible and inflexible power are determined by the actor network,
-        which is trained with the MATD3 algorithm.
-        The maximum of those two prices is used for the flexible bid and the minimum for the inflexible bid.
-        The order type is set implicitly (not by the RL agent itself) and the structure depends on
-        the allowed order types:
-        If only simple hourly orders (SB) are allowed, the strategy will only use SB for both inflexible and flexible power.
-        If SB and linked orders (LB) are allowed, the strategy will use SB for the inflexible power and LB for the flexible power.
-        If SB and block orders (BB) are allowed, the strategy will use BB for the inflexible power and SB for the flexible power.
-        If all three order types (SB, BB, LB) are allowed, the strategy will use BB for the inflexible power
-        and LB for the flexible power, except the inflexible power is 0,
-        then it will use SB for the flexible power (as for VREs).
-    """
-
-    def calculate_bids(
-        self,
-        unit: SupportsMinMax,
-        market_config: MarketConfig,
-        product_tuples: list[Product],
-        **kwargs,
-    ) -> Orderbook:
-        """
-        Calculates bids for a unit.
-
-        Args:
-            unit (SupportsMinMax): Unit to calculate bids for
-            market_config (MarketConfig): Market configuration
-            product_tuples (list[Product]): Product tuples
-
-        Returns:
-            Orderbook: Bids containing start time, end time, price, volume and bid type
-
-        """
-
-        start = product_tuples[0][0]
-        end = product_tuples[-1][1]
-
-        # =============================================================================
-        # 1. Get the Observations, which are the basis of the action decision
-        # =============================================================================
-        next_observation = self.create_observation(
-            unit=unit,
-            market_id=market_config.market_id,
-            start=start,
-        )
-
-        # =============================================================================
-        # 2. Get the Actions, based on the observations
-        # =============================================================================
-        actions, noise = self.get_actions(next_observation)
-
-        # =============================================================================
-        # 3. Transform Actions into bids
[…]
