+++ tests/test_units_operator.py
@@ -164,9 +164,6 @@ async def test_formulate_bids(units_operator: UnitsOperator):
 
 @pytest.mark.require_learning
 async def test_write_learning_params(rl_units_operator: RLUnitsOperator):
-    from assume.strategies.learning_advanced_orders import (
-        LearningProfileEOMPowerplantStrategy,
-    )
     from assume.strategies.learning_strategies import LearningEOMPowerplantStrategy
 
     marketconfig = rl_units_operator.available_markets[0]
@@ -176,7 +173,7 @@ async def test_write_learning_params(rl_units_operator: RLUnitsOperator):
 
     params_dict = {
         "bidding_strategies": {
-            "EOM": LearningProfileEOMPowerplantStrategy(
+            "EOM": LearningEOMPowerplantStrategy(
                 unit_id="testplant",
                 learning_mode=True,
             )
