### 2025-03-27 f25578e
- update hyperprameter for 02f
