+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -197,7 +197,11 @@ class TD3(RLAlgorithm):
 
         """
         if actors_and_critics is None:
-            self.create_actors()
+            self.check_policy_dimensions()
+            if self.actor_architecture == "contextual_mlp":
+                self.create_shared_actor()
+            else:
+                self.create_individual_actors()
             self.create_critics()
 
         else:
@@ -212,7 +216,54 @@ class TD3(RLAlgorithm):
             self.act_dim = actors_and_critics["act_dim"]
             self.unique_obs_dim = actors_and_critics["unique_obs_dim"]
 
-    def create_actors(self) -> None:
+    def check_policy_dimensions(self) -> None:
+        obs_dim_list = []
+        context_dim_list = []
+        act_dim_list = []
+        unique_obs_dim_list = []
+        num_timeseries_obs_dim_list = []
+
+        for strategy in self.learning_role.rl_strats.values():
+            obs_dim_list.append(strategy.obs_dim)
+            context_dim_list.append(strategy.context_dim)
+            act_dim_list.append(strategy.act_dim)
+            unique_obs_dim_list.append(strategy.unique_obs_dim)
+            num_timeseries_obs_dim_list.append(strategy.num_timeseries_obs_dim)
+
+        if len(set(obs_dim_list)) > 1:
+            raise ValueError(
+                "All observation dimensions must be the same for all RL agents"
+            )
+        else:
+            self.obs_dim = obs_dim_list[0]
+
+        if len(set(context_dim_list)) > 1:
+            raise ValueError(
+                "All context dimensions must be the same for all RL agents"
+            )
+        else:
+            self.context_dim = context_dim_list[0]
+
+        if len(set(act_dim_list)) > 1:
+            raise ValueError("All action dimensions must be the same for all RL agents")
+        else:
+            self.act_dim = act_dim_list[0]
+
+        if len(set(unique_obs_dim_list)) > 1:
+            raise ValueError(
+                "All unique_obs_dim values must be the same for all RL agents"
+            )
+        else:
+            self.unique_obs_dim = unique_obs_dim_list[0]
+
+        if len(set(num_timeseries_obs_dim_list)) > 1:
+            raise ValueError(
+                "All num_timeseries_obs_dim values must be the same for all RL agents"
+            )
+        else:
+            self.num_timeseries_obs_dim = num_timeseries_obs_dim_list[0]
+
+    def create_individual_actors(self) -> None:
         """
         Create actor networks for reinforcement learning for each unit strategy.
 
@@ -226,9 +277,6 @@ class TD3(RLAlgorithm):
             If you have units with different observation dimensions. They need to have different critics and hence learning roles.
 
         """
-
-        obs_dim_list = []
-        act_dim_list = []
 
         for strategy in self.learning_role.rl_strats.values():
             strategy.actor = self.actor_architecture_class(
@@ -257,20 +305,53 @@ class TD3(RLAlgorithm):
                 ),  # 1=100% of simulation remaining, uses learning_rate from config as starting point
             )
 
-            obs_dim_list.append(strategy.obs_dim)
-            act_dim_list.append(strategy.act_dim)
+    def create_shared_actor(self) -> None:
+        """
+        Create a shared actor network for reinforcement learning used by all learning strategies.
 
-        if len(set(obs_dim_list)) > 1:
-            raise ValueError(
-                "All observation dimensions must be the same for all RL agents"
-            )
-        else:
-            self.obs_dim = obs_dim_list[0]
+        This method initializes actor networks and their corresponding target networks for each unit strategy.
+        The actors are designed to map observations to action probabilities in a reinforcement learning setting.
 
-        if len(set(act_dim_list)) > 1:
[…]
