+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -100,14 +100,26 @@ class TD3(RLAlgorithm):
             directory (str): The base directory for saving the parameters.
         """
         os.makedirs(directory, exist_ok=True)
-        for u_id, strategy in self.learning_role.rl_strats.items():
+        # save a single shared actor for all agents
+        if self.actor_architecture == "contextual_mlp":
             obj = {
-                "actor": strategy.actor.state_dict(),
-                "actor_target": strategy.actor_target.state_dict(),
-                "actor_optimizer": strategy.actor.optimizer.state_dict(),
+                "actor": self.shared_actor.state_dict(),
+                "actor_target": self.shared_actor_target.state_dict(),
+                "actor_optimizer": self.shared_actor.optimizer.state_dict(),
             }
-            path = f"{directory}/actor_{u_id}.pt"
+            path = f"{directory}/shared_actor.pt"
             th.save(obj, path)
+
+        # else save individual actors for each agent
+        else:
+            for u_id, strategy in self.learning_role.rl_strats.items():
+                obj = {
+                    "actor": strategy.actor.state_dict(),
+                    "actor_target": strategy.actor_target.state_dict(),
+                    "actor_optimizer": strategy.actor.optimizer.state_dict(),
+                }
+                path = f"{directory}/actor_{u_id}.pt"
+                th.save(obj, path)
 
     def load_params(self, directory: str) -> None:
         """
@@ -172,18 +184,32 @@ class TD3(RLAlgorithm):
             )
             return
 
-        for u_id, strategy in self.learning_role.rl_strats.items():
+        if self.actor_architecture == "contextual_mlp":
             try:
                 actor_params = self.load_obj(
-                    directory=f"{directory}/actors/actor_{str(u_id)}.pt"
+                    directory=f"{directory}/actors/shared_actor.pt"
                 )
-                strategy.actor.load_state_dict(actor_params["actor"])
-                strategy.actor_target.load_state_dict(actor_params["actor_target"])
-                strategy.actor.optimizer.load_state_dict(
+                self.shared_actor.load_state_dict(actor_params["actor"])
+                self.shared_actor_target.load_state_dict(actor_params["actor_target"])
+                self.shared_actor.optimizer.load_state_dict(
                     actor_params["actor_optimizer"]
                 )
             except Exception:
-                logger.warning(f"No actor values loaded for agent {u_id}")
+                logger.warning("No shared actor values loaded")
+
+        else:
+            for u_id, strategy in self.learning_role.rl_strats.items():
+                try:
+                    actor_params = self.load_obj(
+                        directory=f"{directory}/actors/actor_{str(u_id)}.pt"
+                    )
+                    strategy.actor.load_state_dict(actor_params["actor"])
+                    strategy.actor_target.load_state_dict(actor_params["actor_target"])
+                    strategy.actor.optimizer.load_state_dict(
+                        actor_params["actor_optimizer"]
+                    )
+                except Exception:
+                    logger.warning(f"No actor values loaded for agent {u_id}")
 
     def initialize_policy(self, actors_and_critics: dict = None) -> None:
         """
@@ -205,10 +231,19 @@ class TD3(RLAlgorithm):
             self.create_critics()
 
         else:
-            for u_id, strategy in self.learning_role.rl_strats.items():
-                strategy.actor = actors_and_critics["actors"][u_id]
-                strategy.actor_target = actors_and_critics["actor_targets"][u_id]
+            if self.actor_architecture == "contextual_mlp":
+                self.shared_actor = actors_and_critics["actors"]
+                self.shared_actor_target = actors_and_critics["actor_targets"]
 
[…]
