### 2025-03-28 b70b065
- fix storage strategy
