+++ examples/inputs/example_02c/config.yaml
@@ -15,6 +15,52 @@ base:
     trained_policies_load_path: null
     max_bid_price: 100
     algorithm: matd3
+    actor_architecture: mlp
+    learning_rate: 0.001
+    training_episodes: 100
+    episodes_collecting_initial_experience: 5
+    train_freq: 100h
+    gradient_steps: 10
+    batch_size: 128
+    gamma: 0.99
+    device: cpu
+    action_noise_schedule: linear
+    noise_sigma: 0.1
+    noise_scale: 1
+    noise_dt: 1
+    validation_episodes_interval: 5
+
+  markets_config:
+    EOM:
+      operator: EOM_operator
+      product_type: energy
+      products:
+        - duration: 1h
+          count: 1
+          first_delivery: 1h
+      opening_frequency: 1h
+      opening_duration: 1h
+      volume_unit: MWh
+      maximum_bid_volume: 100000
+      maximum_bid_price: 3000
+      minimum_bid_price: -500
+      price_unit: EUR/MWh
+      market_mechanism: pay_as_clear
+
+contextual:
+  start_date: 2019-03-01 00:00
+  end_date: 2019-04-01 00:00
+  time_step: 1h
+  save_frequency_hours: null
+  learning_mode: true
+
+  learning_config:
+    continue_learning: false
+    trained_policies_save_path: null
+    trained_policies_load_path: null
+    max_bid_price: 100
+    algorithm: matd3
+    actor_architecture: contextual_mlp
     learning_rate: 0.001
     training_episodes: 100
     episodes_collecting_initial_experience: 5
