### 2025-03-28 02bb600
- add new configurable parameters
