+++ assume/units/depot.py
@@ -3,13 +3,16 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import logging
+
 import pyomo.environ as pyo
-from assume.common.utils import str_to_bool
+
 from assume.common.base import SupportsMinMax
 from assume.common.forecasts import Forecaster
+from assume.common.utils import str_to_bool
 from assume.units.dsm_load_shift import DSMFlex
 
 logger = logging.getLogger(__name__)
+
 
 class BusDepot(DSMFlex, SupportsMinMax):
     """
@@ -38,8 +41,8 @@ class BusDepot(DSMFlex, SupportsMinMax):
         optional_technologies (list): Optional technologies available.
     """
 
-    required_technologies = []
-    optional_technologies = ["charging_station", "electric_vehicle"]
+    required_technologies = ["electric_vehicle"]
+    optional_technologies = []  # "charging_station",
 
     def __init__(
         self,
@@ -52,7 +55,7 @@ class BusDepot(DSMFlex, SupportsMinMax):
         is_prosumer: str = "No",
         objective: str = "min_cost",
         flexibility_measure: str = "cost_based_load_shift",
-        max_power_capacity: float = 1000,
+        cost_tolerance: float = 10,
         node: str = "bus_node",
         location: tuple[float, float] = (0.0, 0.0),
         **kwargs,
@@ -68,18 +71,61 @@ class BusDepot(DSMFlex, SupportsMinMax):
             location=location,
             **kwargs,
         )
+        # check if the required components are present in the components dictionary
+        for component in self.required_technologies:
+            if component not in components.keys():
+                raise ValueError(
+                    f"Component {component} is required for the Bus Depot  unit."
+                )
+
+        # check if the provided components are valid and do not contain any unknown components
+        for component in components.keys():
+            if (
+                component not in self.required_technologies
+                and component not in self.optional_technologies
+            ):
+                raise ValueError(
+                    f"Components {component} is not a valid component for the Bus Depot unit."
+                )
 
         self.objective = objective
         self.is_prosumer = str_to_bool(is_prosumer)
+        self.cost_tolerance = cost_tolerance
         self.flexibility_measure = flexibility_measure
-        self.max_power_capacity = max_power_capacity
-        self.electricity_price = self.forecaster["price_EOM"]
-        self.bus_availability = self.forecaster["bus_availability"]
-        self.bus_soc_profile = self.forecaster["bus_soc_profile"]
-        self.renewable_availability = self.forecaster["renewable_availability"]
+        self.is_prosumer = str_to_bool(is_prosumer)
 
+        self.electricity_price = self.forecaster["electricity_price"]
+        self.range = self.forecaster["range"]
+
+        # Check for the presence of components
+        self.has_ev = "electric_vehicle" in self.components.keys()
+        # Configure EV  power profile based on availability or charging profile
+        if self.has_ev:
+            use_charging_profile = str_to_bool(
+                self.components["electric_vehicle"].get("charging_profile", "false")
+            )
+
+            # Remove flag key to avoid accidentally passing string to the EV model
+            self.components["electric_vehicle"].pop("charging_profile", None)
+
+            profile_key_ev = (
+                f"{self.id}_ev_charging_profile"
+                if use_charging_profile
+                else "availability_profile_ev"
+            )
+
+            ev_profile = self.forecaster[profile_key_ev]
+
+            # # ✅ Add this to preview the time series you're passing into EV
+            # print(f"[DEBUG] Loaded EV profile from '{profile_key_ev}':")
+            # print(ev_profile.as_pd_series().head(10))  # ✅ If you want a timestamped view
+
+            self.components["electric_vehicle"][
+                "charging_profile" if use_charging_profile else "availability_profile"
+            ] = ev_profile
+
[…]
