### 2025-03-29 07a6ed2
-improved NN creation
