+++ assume/reinforcement_learning/neural_network_architecture.py
@@ -7,7 +7,34 @@ from torch import nn
 from torch.nn import functional as F
 
 
-class CriticTD3(nn.Module):
+class Critic(nn.Module):
+    """Parent class for critic networks."""
+
+    def __init__(self):
+        super().__init__()
+
+    def _build_q_network(self, input_dim, hidden_sizes, float_type):
+        """
+        Build a Q-network as a sequence of linear layers.
+        """
+        layers = nn.ModuleList()
+        for h in hidden_sizes:
+            layers.append(nn.Linear(input_dim, h, dtype=float_type))
+            input_dim = h
+        layers.append(nn.Linear(input_dim, 1, dtype=float_type))  # Output Q-value
+        return layers
+
+    def _init_weights(self):
+        """
+        Apply Xavier initialization to all linear layers.
+        """
+        for m in self.modules():
+            if isinstance(m, nn.Linear):
+                nn.init.xavier_uniform_(m.weight)
+                nn.init.zeros_(m.bias)
+
+
+class CriticTD3(Critic):
     """Initialize parameters and build model.
 
     Args:
@@ -37,42 +64,17 @@ class CriticTD3(nn.Module):
         else:
             hidden_sizes = [1024, 512, 256, 128]  # Deeper network for large `n_agents`
 
+        input_dim = self.obs_dim + self.act_dim
         # First Q-network (Q1)
-        self.q1_layers = self._build_q_network(hidden_sizes, float_type)
+        self.q1_layers = self._build_q_network(input_dim, hidden_sizes, float_type)
 
         # Second Q-network (Q2) for double Q-learning
-        self.q2_layers = self._build_q_network(hidden_sizes, float_type)
+        self.q2_layers = self._build_q_network(input_dim, hidden_sizes, float_type)
 
         # Initialize weights properly
         self._init_weights()
 
-    def _build_q_network(self, hidden_sizes, float_type):
-        """
-        Dynamically creates a Q-network given the chosen hidden layer sizes.
-        """
-        layers = nn.ModuleList()
-        input_dim = (
-            self.obs_dim + self.act_dim
-        )  # Input includes all observations and actions
-
-        for h in hidden_sizes:
-            layers.append(nn.Linear(input_dim, h, dtype=float_type))
-            input_dim = h
-        layers.append(nn.Linear(input_dim, 1, dtype=float_type))  # Output Q-value
-
-        return layers
-
-    def _init_weights(self):
-        """Apply Xavier initialization to all layers."""
-
-        def init_layer(m):
-            if isinstance(m, nn.Linear):
-                nn.init.xavier_uniform_(m.weight)
-                nn.init.zeros_(m.bias)
-
-        self.apply(init_layer)
-
-    def forward(self, obs, actions):
+    def forward(self, obs, actions, *args, **kwargs):
         """
         Forward pass through both Q-networks.
         """
@@ -92,7 +94,7 @@ class CriticTD3(nn.Module):
 
         return x1, x2
 
-    def q1_forward(self, obs, actions):
+    def q1_forward(self, obs, actions, *args, **kwargs):
         """
         Compute only Q1 (used during actor updates).
         """
@@ -106,6 +108,120 @@ class CriticTD3(nn.Module):
         return x
 
 
+class ContextualCriticTD3(Critic):
+    """
+    A centralized critic that incorporates both the global state (observation) and
+    unit-specific context in separate branches before combining them with the actions.
+
+    Args:
+        n_agents (int): Number of agents.
+        obs_dim (int): Dimension of each agent's observation.
+        context_dim (int): Dimension of each agent's extra context.
+        act_dim (int): Dimension of each agent's action.
[…]
