+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -11,7 +11,10 @@ from torch.optim import AdamW
 
 from assume.reinforcement_learning.algorithms.base_algorithm import RLAlgorithm
 from assume.reinforcement_learning.learning_utils import polyak_update
-from assume.reinforcement_learning.neural_network_architecture import CriticTD3
+from assume.reinforcement_learning.neural_network_architecture import (
+    ContextualCriticTD3,
+    CriticTD3,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -56,6 +59,10 @@ class TD3(RLAlgorithm):
         self.n_updates = 0
         self.grad_clip_norm = 1.0
 
+        self.use_shared_actor_critic = (
+            True if actor_architecture == "contextual_mlp" else False
+        )
+
     def save_params(self, directory):
         """
         This method saves the parameters of both the actor and critic networks associated with the learning role. It organizes the
@@ -79,14 +86,26 @@ class TD3(RLAlgorithm):
             directory (str): The base directory for saving the parameters.
         """
         os.makedirs(directory, exist_ok=True)
-        for u_id, strategy in self.learning_role.rl_strats.items():
+        # save a single shared critic for all agents
+        if self.use_shared_actor_critic:
             obj = {
-                "critic": strategy.critics.state_dict(),
-                "critic_target": strategy.target_critics.state_dict(),
-                "critic_optimizer": strategy.critics.optimizer.state_dict(),
+                "critic": self.shared_critic.state_dict(),
+                "critic_target": self.shared_critic_target.state_dict(),
+                "critic_optimizer": self.shared_critic.optimizer.state_dict(),
             }
-            path = f"{directory}/critic_{u_id}.pt"
+            path = f"{directory}/shared_critic.pt"
             th.save(obj, path)
+
+        # else save individual critics for each agent
+        else:
+            for u_id, strategy in self.learning_role.rl_strats.items():
+                obj = {
+                    "critic": strategy.critics.state_dict(),
+                    "critic_target": strategy.target_critics.state_dict(),
+                    "critic_optimizer": strategy.critics.optimizer.state_dict(),
+                }
+                path = f"{directory}/critic_{u_id}.pt"
+                th.save(obj, path)
 
     def save_actor_params(self, directory):
         """
@@ -101,7 +120,7 @@ class TD3(RLAlgorithm):
         """
         os.makedirs(directory, exist_ok=True)
         # save a single shared actor for all agents
-        if self.actor_architecture == "contextual_mlp":
+        if self.use_shared_actor_critic:
             obj = {
                 "actor": self.shared_actor.state_dict(),
                 "actor_target": self.shared_actor_target.state_dict(),
@@ -153,18 +172,38 @@ class TD3(RLAlgorithm):
             )
             return
 
-        for u_id, strategy in self.learning_role.rl_strats.items():
+        if self.use_shared_actor_critic:
+            # load shared critic parameters
             try:
                 critic_params = self.load_obj(
-                    directory=f"{directory}/critics/critic_{str(u_id)}.pt"
+                    directory=f"{directory}/critics/shared_critic.pt"
                 )
-                strategy.critics.load_state_dict(critic_params["critic"])
-                strategy.target_critics.load_state_dict(critic_params["critic_target"])
-                strategy.critics.optimizer.load_state_dict(
+                self.shared_critic.load_state_dict(critic_params["critic"])
+                self.shared_critic_target.load_state_dict(
+                    critic_params["critic_target"]
+                )
+                self.shared_critic.optimizer.load_state_dict(
                     critic_params["critic_optimizer"]
                 )
-            except Exception:
-                logger.warning(f"No critic values loaded for agent {u_id}")
+            except Exception as e:
[…]
