+++ assume/reinforcement_learning/neural_network_architecture.py
@@ -108,120 +108,6 @@ class CriticTD3(Critic):
         return x
 
 
-class ContextualCriticTD3(Critic):
-    """
-    A centralized critic that incorporates both the global state (observation) and
-    unit-specific context in separate branches before combining them with the actions.
-
-    Args:
-        n_agents (int): Number of agents.
-        obs_dim (int): Dimension of each agent's observation.
-        context_dim (int): Dimension of each agent's extra context.
-        act_dim (int): Dimension of each agent's action.
-        float_type: The torch data type (e.g., torch.float32).
-        unique_obs_dim (int, optional): Extra observation dimensions from other agents (default=0).
-    """
-
-    def __init__(
-        self,
-        n_agents: int,
-        obs_dim: int,
-        context_dim: int,
-        act_dim: int,
-        float_type,
-        unique_obs_dim: int = 0,
-        hidden_sizes=[1024, 512, 256, 128],
-    ):
-        super().__init__()
-
-        # Effective dimensions:
-        # For the observation branch, we use the agent's own observation
-        # plus the unique observations from the other agents.
-        self.effective_obs_dim = obs_dim + unique_obs_dim * (n_agents - 1)
-        # For the actions branch, we have all agents' actions concatenated.
-        self.effective_act_dim = act_dim * n_agents
-        # For the context branch, we assume that each agent has a context vector;
-        # these are concatenated in a fixed order.
-        self.effective_context_dim = context_dim
-
-        # Build separate processing layers for observation and context.
-        self.obs_fc = nn.Linear(
-            self.effective_obs_dim, hidden_sizes[-1], dtype=float_type
-        )
-        self.context_fc = nn.Linear(
-            self.effective_context_dim, hidden_sizes[-1], dtype=float_type
-        )
-
-        # After processing, we concatenate the two embeddings with the actions.
-        # The combined input dimension becomes: 128 (obs) + 128 (context) + effective_act_dim.
-        combined_input_dim = hidden_sizes[-1] * 2 + self.effective_act_dim
-
-        # Build the two Q-networks (Q1 and Q2).
-        self.q1_layers = self._build_q_network(
-            combined_input_dim, hidden_sizes, float_type
-        )
-        self.q2_layers = self._build_q_network(
-            combined_input_dim, hidden_sizes, float_type
-        )
-
-        self._init_weights()
-
-    def forward(self, obs, context, actions):
-        """
-        Forward pass through both Q-networks.
-
-        Args:
-            obs (torch.Tensor): Global observation tensor with shape [batch, effective_obs_dim].
-            context (torch.Tensor): Concatenated context tensor with shape [batch, effective_context_dim].
-            actions (torch.Tensor): Concatenated actions with shape [batch, effective_act_dim].
-
-        Returns:
-            Tuple of Q-value estimates (Q1, Q2) each with shape [batch, 1].
-        """
-
-        # expand context to match batch size
-        if obs.dim() != context.dim():
-            context = context.unsqueeze(0).expand(obs.size(0), -1)
-
-        # Process the observation and context through their respective branches.
-        obs_emb = F.relu(self.obs_fc(obs))
-        context_emb = F.relu(self.context_fc(context))
-
-        # Concatenate the embeddings with the actions.
-        xu = th.cat([obs_emb, context_emb, actions], dim=1)
-
-        # Compute Q1.
-        x1 = xu
-        for layer in self.q1_layers[:-1]:
-            x1 = F.relu(layer(x1))
-        x1 = self.q1_layers[-1](x1)
-
-        # Compute Q2.
-        x2 = xu
-        for layer in self.q2_layers[:-1]:
[…]
