### 2025-04-02 6d64e39
Improve database handling - add rl_meta to store simulation IDs and episodes to speed up grafana dashboards when having many large simulation - add a function to list and delete specific simulations from the database to avoid deleting complete database
