+++ assume/common/__init__.py
@@ -5,5 +5,5 @@
 from assume.common.forecasts import Forecaster
 from assume.common.mango_serializer import mango_codec_factory
 from assume.common.market_objects import MarketConfig, MarketProduct, Orderbook
-from assume.common.outputs import OutputDef, WriteOutput
+from assume.common.outputs import OutputDef, WriteOutput, DatabaseMaintenance
 from assume.common.units_operator import UnitsOperator
