### 2025-04-02 ee5015b
remove commit as context manager already handles it
