### 2025-04-02 #7666 @kyleniemeyer
@maurerle the version is only assigned after review is complete and prior to publication (since it may change during/after review)
