### 2025-04-03 c7df3ed
Improve database handling (#568)
## Changes Proposed
- add rl_meta to store simulation IDs and episodes to speed up grafana
dashboards when having many large simulation
- adjust variables queries in Grafana learning dashboard to make use of
the rl_meta table and speed up querying
- add a function to list and delete specific simulations from the
database to avoid deleting complete database
