+++ assume/common/outputs.py
@@ -124,6 +124,19 @@ class WriteOutput(Role):
         }
         self.kpi_defs.update(additional_kpis)
 
+        # add rl_meta if in learning or evaluation mode
+        if self.learning_mode or self.evaluation_mode:
+            # Add rl_meta entry to write_buffers
+            self.write_buffers["rl_meta"] = [
+                {
+                    "simulation": self.simulation_id,
+                    "evaluation_mode": self.evaluation_mode,
+                    "learning_mode": self.learning_mode,
+                    "episode": self.episode,
+                    "eval_episode": self.eval_episode,
+                }
+            ]
+
     def delete_db_scenario(self, simulation_id: str):
         """
         Deletes all data from the database for the given simulation id.
@@ -140,8 +153,8 @@ class WriteOutput(Role):
             # ignore spatial_ref_sys table
             if table_name == "spatial_ref_sys":
                 continue
-            # only delete rl_params during the first episode of learning
-            if table_name == "rl_params" and not (
+            # only delete rl_params and rl_meta during the first episode of learning
+            if table_name in ["rl_params", "rl_meta"] and not (
                 self.learning_mode and self.episode == 1
             ):
                 continue
@@ -466,6 +479,8 @@ class WriteOutput(Role):
                         df = self.convert_unit_dispatch(data_list)
                     case "rl_params":
                         df = self.convert_rl_params(data_list)
+                    case "rl_meta":
+                        df = pd.DataFrame(data_list)
                     case "grid_flows":
                         dfs = []
                         for data in data_list:
@@ -712,3 +727,156 @@ class WriteOutput(Role):
         rewards_by_unit = np.array(rewards_by_unit)
 
         return rewards_by_unit
+
+
+class DatabaseMaintenance:
+    """
+    A utility class for managing simulation data stored in a database.
+
+    This class creates a database engine from a provided URI and offers methods to:
+      1. Retrieve a list of unique simulation IDs across all tables.
+      2. Delete specific simulations from every table.
+      3. Delete all simulations, or all except those specified, across every table.
+
+    It assumes that each table (except for system tables like "spatial_ref_sys") contains a column
+    named 'simulation' that uniquely identifies the simulation.
+
+    Example:
+        from assume.common import DatabaseMaintenance
+
+        # Select to store the simulation results in a local database or in TimescaleDB.
+        # When using TimescaleDB, ensure Docker is installed and the Grafana dashboard is accessible.
+        data_format = "timescale"  # Options: "local_db" or "timescale"
+
+        if data_format == "local_db":
+            db_uri = "sqlite:///./examples/local_db/assume_db.db"
+        elif data_format == "timescale":
+            db_uri = "postgresql://assume:assume@localhost:5432/assume"
+
+        maintenance = DatabaseMaintenance(db_uri=db_uri)
+
+        # 1. Retrieve unique simulation IDs:
+        unique_ids = maintenance.get_unique_simulation_ids()
+        print("Unique simulation IDs:", unique_ids)
+
+        # 2. Delete specific simulations:
+        maintenance.delete_simulations(["example_01", "example_02"])
+
+        # 3. Delete all simulations except a few:
+        maintenance.delete_all_simulations(exclude=["example_01"])
+
+    Args:
+        db_uri (str): The URI of the database engine used to create a SQLAlchemy engine.
+    """
+
+    def __init__(self, db_uri: str):
+        """
+        Initializes the DatabaseMaintenance instance by creating a database engine.
+
+        Args:
+            db_uri (str): The URI of the database engine.
+        """
+        self.db_uri = db_uri
+        self.db = create_engine(self.db_uri)
+
+    def get_unique_simulation_ids(self) -> list[str]:
+        """
[…]
