### 2025-04-03 192771b
Improves documentation for the outputs handling and overall structure
- enable configuring buffer size for the outputs buffer
- add documentation for the outputs including methods and tables with their columns
- restructure the index for better grouping
- move reply buffer inside RL section for better readability
- change representation of redispatch and DSM units for nicer documentation
- correct level errors and general bugs in documentation
- other small fixes
