+++ docs/source/redispatch_modeling.rst
@@ -2,28 +2,31 @@
 ..
 .. SPDX-License-Identifier: AGPL-3.0-or-later
 
-==============================================
+
 Congestion Management and Redispatch Modelling
-==============================================
+===============================================
 
 This section demonstrates the modelling and simulation of the redispatch mechanism using PyPSA as a plug-and-play module within the ASSUME framework.
 The model primarily considers grid constraints to identify bottlenecks in the grid, resolve them using the redispatch algorithm, and account for dispatches from the EOM (Energy-Only Market).
 
 Concept of Redispatch
-=====================
-The locational mismatch between electricity demand and generation requires the transmission of electricity from low-demand regions to high-demand regions. The transmission capacity limits the maximum amount of electricity that can be transmitted at any point in time.
-When transmission capacity is insufficient to meet demand, generation must be reduced at locations with low demand and increased at locations with high demand.
-This process is known as **Redispatch**. In addition to spot markets, the redispatch mechanism is used to regulate grid flows and avoid congestion issues. It is operated and controlled by the system operators (SO).
+----------------------
 
------------------------------------------
+The locational mismatch between electricity demand and generation requires the transmission of electricity from low-demand regions to high-demand regions. The transmission capacity limits the maximum amount of electricity that can be transmitted at any point in time.
+
+When transmission capacity is insufficient to meet demand, generation must be reduced at locations with low demand and increased at locations with high demand. This process is known as **Redispatch**. In addition to spot markets, the redispatch mechanism is used to regulate grid flows and avoid congestion issues. It is operated and controlled by the system operators (SO).
+
+
 Overview of Redispatch Modelling in PyPSA
------------------------------------------
+------------------------------------------
+
 The PyPSA network model can be created to visualize line flows using EOM clearing outcomes of generation and loads at different nodes (locations).
 
 PyPSA uses the following terminology to define grid infrastructure:
 
-Attributes:
+Attributes
 -----------
+
 1. **Bus**: Nodes (locations) where power plants and loads are connected
     - **name**: Unique identifier of the bus
     - **v_nom**: Nominal voltage of the node
@@ -65,47 +68,52 @@ Attributes:
     - **name**: Unique identifier of the network
     - **snapshots**: List of timesteps (e.g., hourly, quarter-hourly, daily, etc.)
 
-A PyPSA network model can be created by defining nodes as locations for power generation and consumption, interconnected by transmission lines with nominal transmission capacity (s_nom). These components can be further constrained operationally, for instance, by nominal power, efficiency, ramp rates, and other factors.
+A PyPSA network model can be created by defining nodes as locations for power generation and consumption, interconnected by transmission lines with nominal transmission capacity (`s_nom`). These components can be further constrained operationally, for instance, by nominal power, efficiency, ramp rates, and other factors.
 
 Currently, a limitation of the PyPSA model is the inability to define flexible loads.
 
+Modelling Redispatch in ASSUME
+--------------------------------
+
 Modelling redispatch in the ASSUME framework using PyPSA primarily includes two parts:
 
---------------------------------
-1. Congestion Identification
---------------------------------
+Congestion Identification
+--------------------------
 
 The first step is to check for congestion in the network. The linear power flow (LPF) method is particularly useful for quick assessments of congestion and redispatch needs. PyPSA provides the `network.lpf()` function for running linear power flow. This method is significantly faster than a full non-linear AC power flow, making it suitable for real-time analysis or large network studies.
 
[…]
