+++ docs/source/outputs.rst
@@ -1,0 +1,172 @@
+.. SPDX-FileCopyrightText: ASSUME Developers
+..
+.. SPDX-License-Identifier: AGPL-3.0-or-later
+
+###############
+Storing Outputs
+###############
+
+This documentation describes the functionality of the output processing system used to store simulation outputs either in a database or as CSV files. It covers the configuration, data buffering, conversion, and storage mechanisms of the main class as well as the support class for database maintenance.
+
+
+WriteOutput Class
+==================
+
+The ``WriteOutput`` class is responsible for:
+
+- Collecting simulation output messages.
+- Buffering data until either a time-based interval or a memory limit is reached.
+- Converting raw data into structured pandas DataFrames.
+- Storing data into a database and/or exporting it as CSV files.
+- Managing performance trade-offs by allowing configuration of buffer sizes and storage intervals.
+
+Constructor Parameters
+------------------------
+- **simulation_id** (*str*, required): Unique identifier for the simulation.
+- **start** (*datetime*, required): The simulation start date and time.
+- **end** (*datetime*, required): The simulation end date and time.
+- **db_uri** (*str*, optional): URI for the database engine. Defaults to an empty string.
+- **export_csv_path** (*str*, optional): Path for exporting CSV files. When provided, ``save_frequency`` is disabled.
+- **save_frequency_hours** (*int*, optional): Interval (in hours) at which the buffered data is stored. Setting to ``null`` disables time-based writes. Default: 48 hours.
+- **outputs_buffer_size** (*int*, optional): Maximum allowed size of the output buffer in MB. Default: 300 MB.
+
+Data Processing Workflow
+--------------------------
+1. **Message Handling:**
+
+   - The method ``handle_output_message`` receives asynchronous messages containing simulation outputs.
+   - Data is buffered in lists corresponding to the intended database tables based on content type.
+
+2. **Buffer Size Management:**
+
+   - The memory usage of buffered data is monitored continuously.
+   - If memory usage exceeds the ``outputs_buffer_size`` threshold, buffered data is flushed to storage.
+
+3. **Data Conversion:**
+
+   - Helper methods like ``convert_rl_params`` and ``convert_market_results`` transform raw data into DataFrames suitable for storage.
+
+4. **Data Storage:**
+
+   - The method ``store_dfs`` flushes buffered DataFrames to storage based on time intervals (``save_frequency_hours``) or buffer size limits.
+
+Performance Considerations
+---------------------------
+- **Buffering Strategy:** Increasing ``outputs_buffer_size`` can improve performance by reducing write frequency. Default: 300 MB.
+- **Write Intervals:** Setting ``save_frequency_hours`` facilitates real-time observation with Grafana dashboards. Disabling this by setting it to ``null`` improves performance but prevents real-time monitoring.
+- **Optimal Configuration:** Increase ``outputs_buffer_size`` and disable ``save_frequency_hours`` for maximum performance. Ensure sufficient memory is available.
+
+.. note::
+  When storing results as CSV files, ``save_frequency`` is automatically disabled, meaning data is only saved at the end of the simulation or if the buffer size limit is reached. This prevents real-time observation through Grafana dashboards.
+
+Key Methods
+------------
+- **handle_output_message(self, content, meta):** Processes incoming messages and buffers them.
+- **store_dfs(self):** Flushes the DataFrames to storage.
+- **on_stop(self):** Ensures all remaining buffered data is stored at the end of the simulation.
+
+Database Tables Overview
+=========================
[…]

[…]
