### 2025-04-03 369a523
Update assume/common/outputs.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
