### 2025-04-03 8977122
adress comments and improve structure
