### 2025-04-03 f2a521a
Update assume/world.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
