+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -56,6 +56,10 @@ class TD3(RLAlgorithm):
         self.n_updates = 0
         self.grad_clip_norm = 1.0
 
+        self.use_shared_actor = (
+            True if actor_architecture == "contextual_mlp" else False
+        )
+
     def save_params(self, directory):
         """
         This method saves the parameters of both the actor and critic networks associated with the learning role. It organizes the
@@ -65,7 +69,11 @@ class TD3(RLAlgorithm):
             directory (str): The base directory for saving the parameters.
         """
         self.save_critic_params(directory=f"{directory}/critics")
-        self.save_actor_params(directory=f"{directory}/actors")
+
+        if self.use_shared_actor:
+            self.save_shared_actor_params(directory=f"{directory}/actors")
+        else:
+            self.save_actor_params(directory=f"{directory}/actors")
 
     def save_critic_params(self, directory):
         """
@@ -100,26 +108,26 @@ class TD3(RLAlgorithm):
             directory (str): The base directory for saving the parameters.
         """
         os.makedirs(directory, exist_ok=True)
-        # save a single shared actor for all agents
-        if self.actor_architecture == "contextual_mlp":
+        # save individual actors for each agent
+        for u_id, strategy in self.learning_role.rl_strats.items():
             obj = {
-                "actor": self.shared_actor.state_dict(),
-                "actor_target": self.shared_actor_target.state_dict(),
-                "actor_optimizer": self.shared_actor.optimizer.state_dict(),
+                "actor": strategy.actor.state_dict(),
+                "actor_target": strategy.actor_target.state_dict(),
+                "actor_optimizer": strategy.actor.optimizer.state_dict(),
             }
-            path = f"{directory}/shared_actor.pt"
+            path = f"{directory}/actor_{u_id}.pt"
             th.save(obj, path)
 
-        # else save individual actors for each agent
-        else:
-            for u_id, strategy in self.learning_role.rl_strats.items():
-                obj = {
-                    "actor": strategy.actor.state_dict(),
-                    "actor_target": strategy.actor_target.state_dict(),
-                    "actor_optimizer": strategy.actor.optimizer.state_dict(),
-                }
-                path = f"{directory}/actor_{u_id}.pt"
-                th.save(obj, path)
+    def save_shared_actor_params(self, directory):
+        os.makedirs(directory, exist_ok=True)
+        # save a single shared actor for all agents
+        obj = {
+            "actor": self.shared_actor.state_dict(),
+            "actor_target": self.shared_actor_target.state_dict(),
+            "actor_optimizer": self.shared_actor.optimizer.state_dict(),
+        }
+        path = f"{directory}/shared_actor.pt"
+        th.save(obj, path)
 
     def load_params(self, directory: str) -> None:
         """
@@ -132,7 +140,11 @@ class TD3(RLAlgorithm):
             directory (str): The directory from which the parameters should be loaded.
         """
         self.load_critic_params(directory)
-        self.load_actor_params(directory)
+
+        if self.use_shared_actor:
+            self.load_shared_actor_params(directory)
+        else:
+            self.load_actor_params(directory)
 
     def load_critic_params(self, directory: str) -> None:
         """
@@ -184,32 +196,46 @@ class TD3(RLAlgorithm):
             )
             return
 
-        if self.actor_architecture == "contextual_mlp":
+        for u_id, strategy in self.learning_role.rl_strats.items():
             try:
                 actor_params = self.load_obj(
-                    directory=f"{directory}/actors/shared_actor.pt"
+                    directory=f"{directory}/actors/actor_{str(u_id)}.pt"
                 )
-                self.shared_actor.load_state_dict(actor_params["actor"])
-                self.shared_actor_target.load_state_dict(actor_params["actor_target"])
-                self.shared_actor.optimizer.load_state_dict(
[…]
