### 2025-04-07 24d4edb
drop example
