+++ examples/examples.py
@@ -79,10 +79,6 @@ available_examples = {
     "small_learning_2_lstm": {"scenario": "example_02b", "study_case": "base_lstm"},
     # Further DRL example simulation showcasing learning features
     "small_learning_with_storage": {"scenario": "example_02e", "study_case": "base"},
-    "small_learning_with_storage_and_pp": {
-        "scenario": "example_02f",
-        "study_case": "base_longer",
-    },
     #
     # full year examples to show real-world scenarios
     "large_2019_eom": {"scenario": "example_03", "study_case": "base_case_2019"},
