+++ docs/source/outputs.rst
@@ -1,0 +1,151 @@
+.. SPDX-FileCopyrightText: ASSUME Developers
+..
+.. SPDX-License-Identifier: AGPL-3.0-or-later
+
+###############
+Storing Outputs
+###############
+
+This documentation describes the functionality of the output processing system used to store simulation outputs either in a database or as CSV files. It covers the configuration, data buffering, conversion, and storage mechanisms of the main class as well as the support class for database maintenance.
+
+
+WriteOutput Class
+==================
+
+The ``WriteOutput`` class is responsible for:
+
+- Collecting simulation output messages.
+- Buffering data until either a time-based interval or a memory limit is reached.
+- Converting raw data into structured pandas DataFrames.
+- Storing data into a database and/or exporting it as CSV files.
+- Managing performance trade-offs by allowing configuration of buffer sizes and storage intervals.
+
+This class is part of the :mod:`assume.common.outputs` module. For more details, refer to :py:class:`assume.common.outputs.WriteOutput`.
+
+Data Processing Workflow
+--------------------------
+1. **Message Handling:**
+
+   - The method ``handle_output_message`` receives asynchronous messages containing simulation outputs.
+   - Data is buffered in lists corresponding to the intended database tables based on content type.
+
+2. **Buffer Size Management:**
+
+   - The memory usage of buffered data is monitored continuously.
+   - If memory usage exceeds the ``outputs_buffer_size_mb`` threshold, buffered data is flushed to storage.
+
+3. **Data Conversion:**
+
+   - Helper methods like ``convert_rl_params`` and ``convert_market_results`` transform raw data into DataFrames suitable for storage.
+
+4. **Data Storage:**
+
+   - The method ``store_dfs`` flushes buffered DataFrames to storage based on time intervals (``save_frequency_hours``) or buffer size limits.
+
+Performance Considerations
+---------------------------
+- **Buffering Strategy:** Increasing ``outputs_buffer_size_mb`` can improve performance by reducing write frequency. Default: 300 MB.
+- **Write Intervals:** Setting ``save_frequency_hours`` facilitates real-time observation with Grafana dashboards. Disabling this by setting it to ``null`` improves performance but prevents real-time monitoring.
+- **Optimal Configuration:** Increase ``outputs_buffer_size_mb`` and disable ``save_frequency_hours`` for maximum performance. Ensure sufficient memory is available.
+
+.. note::
+  When storing results as CSV files, ``save_frequency`` is automatically disabled, meaning data is only saved at the end of the simulation or if the buffer size limit is reached. This prevents real-time observation through Grafana dashboards.
+
+
+Database Tables Overview
+=========================
+The simulation outputs are stored in various tables. Below is a breakdown of tables, their purpose, and what columns they contain.
+When the outputs are stored in CSV files, the data is organized in a similar structure, with each table represented as a separate CSV file.
+
+.. list-table::
+   :header-rows: 1
+
+   * - **Table Name**
+     - **Purpose**
+     - **Columns**
+
+   * - market_meta
+     - Market-related data.
+     - simulation, market_id, time, node, product_start, product_end, only_hours, price, max_price, min_price, supply_volume, supply_volume_energy, demand_volume, demand_volume_energy
+
+   * - market_dispatch
+     - Dispatch data for markets.
+     - simulation, market_id, datetime, unit_id, power
+
+   * - unit_dispatch
+     - Unit-level dispatch data.
+     - simulation, unit, power, heat, soc, energy_generation_costs, energy_cashflow, total_costs
+
[…]
