+++ examples/inputs/example_06b/config.yaml
@@ -26,89 +26,89 @@ base:
       price_unit: EUR/MWh
       market_mechanism: pay_as_clear
 
-    redispatch:
-      start_date: 2035-01-01 21:00
-      operator: network_operator
-      product_type: energy
-      products:
-        - duration: 1h
-          count: 24
-          first_delivery: 3h
-      opening_frequency: 24h
-      opening_duration: 2h
-      volume_unit: MWh
-      maximum_bid_volume: 100000
-      maximum_bid_price: 3000
-      minimum_bid_price: -500
-      price_unit: EUR/MWh
-      market_mechanism: redispatch
-      additional_fields:
-        - node
-        - min_power
-        - max_power
-      param_dict:
-        network_path: .
-        solver: glpk
-        payment_mechanism: pay_as_bid
-        backup_marginal_cost: 10000
+#     redispatch:
+#       start_date: 2035-01-01 21:00
+#       operator: network_operator
+#       product_type: energy
+#       products:
+#         - duration: 1h
+#           count: 24
+#           first_delivery: 3h
+#       opening_frequency: 24h
+#       opening_duration: 2h
+#       volume_unit: MWh
+#       maximum_bid_volume: 100000
+#       maximum_bid_price: 3000
+#       minimum_bid_price: -500
+#       price_unit: EUR/MWh
+#       market_mechanism: redispatch
+#       additional_fields:
+#         - node
+#         - min_power
+#         - max_power
+#       param_dict:
+#         network_path: .
+#         solver: glpk
+#         payment_mechanism: pay_as_bid
+#         backup_marginal_cost: 10000
 
 
-nodal_case:
-  start_date: 20135-01-01 00:00
-  end_date: 2035-01-06 00:00
-  time_step: 1h
-  save_frequency_hours: 24
+# nodal_case:
+#   start_date: 20135-01-01 00:00
+#   end_date: 2035-01-06 00:00
+#   time_step: 1h
+#   save_frequency_hours: 24
 
-  markets_config:
-    DAM:
-      operator: EOM_operator
-      product_type: energy
-      products:
-        - duration: 1h
-          count: 24
-          first_delivery: 24h
-      opening_frequency: 24h
-      opening_duration: 20h
-      volume_unit: MWh
-      maximum_bid_volume: 100000
-      maximum_bid_price: 3000
-      minimum_bid_price: -500
-      price_unit: EUR/MWh
-      market_mechanism: nodal
-      additional_fields:
-        - node
-        - min_power
-        - max_power
-      param_dict:
-        network_path: .
-        solver: glpk
-        payment_mechanism: pay_as_clear
-        backup_marginal_cost: 10000
+#   markets_config:
+#     DAM:
+#       operator: EOM_operator
+#       product_type: energy
+#       products:
+#         - duration: 1h
+#           count: 24
+#           first_delivery: 24h
+#       opening_frequency: 24h
+#       opening_duration: 20h
+#       volume_unit: MWh
+#       maximum_bid_volume: 100000
+#       maximum_bid_price: 3000
+#       minimum_bid_price: -500
+#       price_unit: EUR/MWh
+#       market_mechanism: nodal
+#       additional_fields:
+#         - node
+#         - min_power
+#         - max_power
+#       param_dict:
+#         network_path: .
+#         solver: glpk
+#         payment_mechanism: pay_as_clear
+#         backup_marginal_cost: 10000
 
-zonal_case:
-  start_date: 2035-01-01 00:00
-  end_date: 2035-01-06 00:00
-  time_step: 1h
-  save_frequency_hours: 24
+# zonal_case:
+#   start_date: 2035-01-01 00:00
+#   end_date: 2035-01-06 00:00
+#   time_step: 1h
[…]
