### 2025-04-08 47ca9e0
- fix optimizer initialization - move obs and context creation to the start of simulation
