+++ assume/reinforcement_learning/algorithms/matd3.py
@@ -4,8 +4,11 @@
 
 import logging
 import os
+from collections import defaultdict
 
+import numpy as np
 import torch as th
+from sklearn.cluster import AgglomerativeClustering, KMeans
 from torch.nn import functional as F
 from torch.optim import AdamW
 
@@ -40,6 +43,8 @@ class TD3(RLAlgorithm):
         target_policy_noise=0.2,
         target_noise_clip=0.5,
         actor_architecture="mlp",
+        actor_clustering_method=None,
+        clustering_method_kwargs={},
     ):
         super().__init__(
             learning_role,
@@ -59,6 +64,13 @@ class TD3(RLAlgorithm):
         self.use_shared_actor = (
             True if actor_architecture == "contextual_mlp" else False
         )
+        if self.use_shared_actor:
+            self.actor_clustering_method = actor_clustering_method
+            self.clustering_method_kwargs = clustering_method_kwargs
+
+            self.shared_actors = {}
+            self.shared_actor_targets = {}
+            self.clusters = defaultdict(list)
 
     def save_params(self, directory):
         """
@@ -120,14 +132,18 @@ class TD3(RLAlgorithm):
 
     def save_shared_actor_params(self, directory):
         os.makedirs(directory, exist_ok=True)
-        # save a single shared actor for all agents
-        obj = {
-            "actor": self.shared_actor.state_dict(),
-            "actor_target": self.shared_actor_target.state_dict(),
-            "actor_optimizer": self.shared_actor.optimizer.state_dict(),
-        }
-        path = f"{directory}/shared_actor.pt"
-        th.save(obj, path)
+
+        # iterate over the shared_actors and save them
+        for cluster_index in self.clusters.keys():
+            obj = {
+                "actor": self.shared_actors[cluster_index].state_dict(),
+                "actor_target": self.shared_actor_targets[cluster_index].state_dict(),
+                "actor_optimizer": self.shared_actors[
+                    cluster_index
+                ].optimizer.state_dict(),
+            }
+            path = f"{directory}/shared_actor_{cluster_index}.pt"
+            th.save(obj, path)
 
     def load_params(self, directory: str) -> None:
         """
@@ -228,12 +244,18 @@ class TD3(RLAlgorithm):
             return
 
         try:
-            actor_params = self.load_obj(
-                directory=f"{directory}/actors/shared_actor.pt"
-            )
-            self.shared_actor.load_state_dict(actor_params["actor"])
-            self.shared_actor_target.load_state_dict(actor_params["actor_target"])
-            self.shared_actor.optimizer.load_state_dict(actor_params["actor_optimizer"])
+            for cluster_index in self.clusters.keys():
+                actor_params = self.load_obj(
+                    directory=f"{directory}/actors/shared_actor_{cluster_index}.pt"
+                )
+                self.shared_actors[cluster_index].load_state_dict(actor_params["actor"])
+                self.shared_actor_targets[cluster_index].load_state_dict(
+                    actor_params["actor_target"]
+                )
+                self.shared_actors[cluster_index].optimizer.load_state_dict(
+                    actor_params["actor_optimizer"]
+                )
+
         except Exception:
             logger.warning("No shared actor values loaded")
 
@@ -252,7 +274,7 @@ class TD3(RLAlgorithm):
             self.check_policy_dimensions()
 
             if self.use_shared_actor:
-                self.create_shared_actor()
+                self.create_shared_actors()
             else:
                 self.create_individual_actors()
 
@@ -260,12 +282,24 @@ class TD3(RLAlgorithm):
 
         else:
             if self.use_shared_actor:
-                self.shared_actor = actors_and_critics["actors"]
[…]
