### 2025-04-09 5e0fdc2
- fix actor loading
