### 2025-04-09 2cc45cf
- fix cluster map creation
