### 2025-04-09 cac25c1
- fix max_size clustering
