### 2025-04-10 dd4e6c9
register agents later in output (#570)
this is required for distributed simulations, in which the output agent
might not yet be available if the clock manager is running somewhere
else

also fix the distributed simulation example
and switch to timescale official repository for docker image

This fixes reconnects and waiting for other containers in distributed
simulations, while not affecting anything else.
