### 2025-04-10 cd2490c
see PR, suggested using terminate learning in initialize policy function
