### 2025-04-11 10257b1
pulp and paper plant
