+++ assume/units/paper_pulp_plant.py
@@ -1,0 +1,235 @@
+# SPDX-FileCopyrightText: ASSUME Developers
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+
+import logging
+from datetime import datetime
+
+import pyomo.environ as pyo
+
+import logging
+from assume.common.base import SupportsMinMax
+from assume.common.forecasts import Forecaster
+from assume.units.dsm_load_shift import DSMFlex
+from assume.units.dst_components import HeatPump, Boiler, GenericStorage
+
+logger = logging.getLogger(__name__)
+
+class PaperPulpPlant(DSMFlex, SupportsMinMax):
+    """
+    Represents a paper and pulp plant in an energy system. This includes components like heat pumps,
+    boilers, and storage units for operational optimization.
+
+    Args:
+        id (str): Unique identifier for the paper and pulp plant.
+        unit_operator (str): The operator responsible for the plant.
+        bidding_strategies (dict): A dictionary of bidding strategies that define how the plant participates in energy markets.
+        forecaster (Forecaster): A forecaster used to get key variables such as fuel or electricity prices.
+        components (dict, optional): A dictionary describing the components of the plant, such as heat pumps and boilers.
+        objective (str, optional): The objective function of the plant, typically to minimize variable costs. Default is "min_variable_cost".
+        flexibility_measure (str, optional): The flexibility measure used for the plant, such as "max_load_shift". Default is "max_load_shift".
+        demand (float, optional): The production demand, representing how much product needs to be produced. Default is 0.
+        cost_tolerance (float, optional): The maximum allowable increase in cost when shifting load. Default is 10.
+        node (str, optional): The node location where the plant is connected within the energy network. Default is "node0".
+        location (tuple[float, float], optional): The geographical coordinates (latitude, longitude) of the paper and pulp plant. Default is (0.0, 0.0).
+        **kwargs: Additional keyword arguments to support more specific configurations or parameters.
+    """
+
+    required_technologies = ["heat_pump", "boiler"]
+    optional_technologies = ["thermal_storage"]
+
+    def __init__(
+        self,
+        id: str,
+        unit_operator: str,
+        bidding_strategies: dict,
+        forecaster: Forecaster,
+        components: dict[str, dict] = None,
+        technology: str = "paper_pulp_plant",
+        objective: str = "min_variable_cost",
+        flexibility_measure: str = "max_load_shift",
+        demand: float = 0,
+        cost_tolerance: float = 10,
+        node: str = "node0",
+        location: tuple[float, float] = (0.0, 0.0),
+        **kwargs,
+    ):
+        super().__init__(
+            id=id,
+            unit_operator=unit_operator,
+            technology=technology,
+            components=components,
+            bidding_strategies=bidding_strategies,
+            forecaster=forecaster,
+            node=node,
+            location=location,
+            **kwargs,
+        )
+
+        # Check if the required components are present in the components dictionary
+        for component in self.required_technologies:
+            if component not in components.keys():
+                raise ValueError(
+                    f"Component {component} is required for the paper and pulp plant unit."
+                )
+
+        # Check if the provided components are valid and do not contain any unknown components
+        for component in components.keys():
+            if (
+                component not in self.required_technologies
+                and component not in self.optional_technologies
+            ):
+                raise ValueError(
+                    f"Component {component} is not a valid component for the paper and pulp plant unit."
+                )
+            
+        # Inject schedule into long-term thermal storage if applicable
+        if "thermal_storage" in self.components:
+            storage_cfg = self.components["thermal_storage"]
+            storage_type = storage_cfg.get("storage_type", "short-term")
[…]

[…]
