+++ assume/units/dst_components.py
@@ -1770,7 +1770,6 @@ demand_side_technologies: dict = {
     "electric_vehicle": ElectricVehicle,
     "generic_storage": GenericStorage,
     "pv_plant": PVPlant,
-    "thermal_storage": GenericStorage,
 }
 
 
