### 2025-04-11 91f3186
- improve contextual architecture
