### 2025-04-11 a963e79
Final paper and plant version
