+++ assume/units/dsm_load_shift.py
@@ -6,6 +6,7 @@ import logging
 from collections.abc import Callable
 from datetime import datetime
 
+import matplotlib.pyplot as plt
 import pyomo.environ as pyo
 from pyomo.opt import (
     SolverFactory,
@@ -251,11 +252,10 @@ class DSMFlex:
                     total_power_input += self.model.dsm_blocks["heat_pump"].power_in[t]
 
                 # Add boiler power if it's electric
-                if (
-                    self.has_boiler
-                    and self.components["boiler"]["fuel_type"] == "electricity"
-                ):
-                    total_power_input += self.model.dsm_blocks["boiler"].power_in[t]
+                if self.has_boiler:
+                    boiler = self.components["boiler"]
+                    if boiler.fuel_type == "electricity":  # Access attribute directly
+                        total_power_input += self.model.dsm_blocks["boiler"].power_in[t]
 
                 # Apply flexibility adjustments
                 return (
@@ -567,6 +567,56 @@ class DSMFlex:
             pyo.value(instance.variable_cost[t]) for t in instance.time_steps
         ]
         self.variable_cost_series = FastSeries(index=self.index, value=variable_cost)
+
+        # Setup plotting
+        time_steps = list(instance.time_steps)
+        fig, axs = plt.subplots(3, 1, figsize=(12, 12), sharex=True)
+
+        # Plot 1: Heat Pump power input
+        if "heat_pump" in instance.dsm_blocks:
+            hp_power = [
+                pyo.value(instance.dsm_blocks["heat_pump"].power_in[t])
+                for t in time_steps
+            ]
+            axs[0].plot(time_steps, hp_power, label="Heat Pump Power In", color="green")
+            axs[0].set_ylabel("Power (kW)")
+            axs[0].legend()
+            axs[0].grid(True)
+
+        # Plot 2: Boiler fuel input
+        if "boiler" in instance.dsm_blocks and hasattr(
+            instance.dsm_blocks["boiler"], "natural_gas_in"
+        ):
+            boiler_gas = [
+                pyo.value(instance.dsm_blocks["boiler"].natural_gas_in[t])
+                for t in time_steps
+            ]
+            axs[1].plot(
+                time_steps, boiler_gas, label="Boiler Natural Gas In", color="orange"
+            )
+            axs[1].set_ylabel("Fuel Input (kW)")
+            axs[1].legend()
+            axs[1].grid(True)
+
+        # Plot 3: Thermal Storage SOC, Charge, Discharge (Discharge in 4th quadrant)
+        if "thermal_storage" in instance.dsm_blocks:
+            soc = [pyo.value(instance.dsm_blocks["thermal_storage"].soc[t]) for t in time_steps]
+            charge = [pyo.value(instance.dsm_blocks["thermal_storage"].charge[t]) for t in time_steps]
+            discharge = [pyo.value(instance.dsm_blocks["thermal_storage"].discharge[t]) for t in time_steps]
+
+            axs[2].fill_between(time_steps, soc, alpha=0.3, label="SOC", color="blue")
+            axs[2].plot(time_steps, charge, label="Charge", linestyle="--", color="green")
+            axs[2].plot(time_steps, [-x for x in discharge], label="Discharge", linestyle="--", color="red")
+            axs[2].axhline(0, color='black', linewidth=0.5)
+            axs[2].set_ylabel("Storage (kWh)")
+            axs[2].legend()
+            axs[2].grid(True)
+
+
+        axs[2].set_xlabel("Time Step")
+        plt.suptitle("Paper & Pulp Plant Operation Overview")
+        plt.tight_layout()
+        plt.show()
 
     def determine_optimal_operation_with_flex(self):
         """
