+++ examples/inputs/example_04/config.yaml
@@ -4,8 +4,8 @@
 
 base_case_2019:
   start_date: 2019-01-01 00:00
-  end_date: 2019-01-03 00:00
-  time_step: 1h
+  end_date: 2019-01-04 00:00
+  time_step: 15min
   save_frequency_hours: null
 
   markets_config:
