+++ assume/scenario/loader_csv.py
@@ -1076,8 +1076,6 @@ def run_learning(
         terminate_learning=True,
     )
 
-    world.learning_role.load_inter_episodic_data(inter_episodic_data)
-
 
 if __name__ == "__main__":
     data = read_grid(Path("examples/inputs/example_01d"))
