+++ assume/reinforcement_learning/algorithms/__init__.py
@@ -7,11 +7,13 @@ from torch import nn
 from assume.reinforcement_learning.neural_network_architecture import (
     MLPActor,
     ContextualMLPActor,
+    FilmedMLPActor,
     LSTMActor,
 )
 
 actor_architecture_aliases: dict[str, type[nn.Module]] = {
     "mlp": MLPActor,
     "contextual_mlp": ContextualMLPActor,
+    "filmed_mlp": FilmedMLPActor,
     "lstm": LSTMActor,
 }
