### 2025-04-14 43b1f54
This PR allows us to change the number of learning agents between training runs while still reusing previously trained critics in `continue_learning` mode. Useful for scenarios like training power plants first and later adding storage units or adjusting the number of agents.
Changes:
- Add weight transfer logic to adapt critic weights when agent count changes
- Perform architecture check before loading critics; if mismatched, safely transfer compatible weights
- Transfer logic copies:
  - Shared observation weights
  - Unique per-agent obs and action weights for overlapping agents only
- Add `
[…]
