### 2025-04-14 74c708b
- fix initial exploration disable logic
