### 2025-04-14 d904d3f
implement bidding strategy that approximates marginal utility curve for demand
