+++ examples/inputs/future_markets_paper/config.yaml
@@ -2,37 +2,38 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
-base:
-  start_date: 2019-03-01 00:00
-  end_date: 2019-03-02 00:00
+base_case_2013:
+  start_date: 2013-01-01 00:00
+  end_date: 2013-01-03 00:00
   time_step: 1h
-  save_frequency_hours: null
-  learning_mode: true
+  learning_mode: True
+  save_frequency_hours: Null
 
   learning_config:
-    continue_learning: false
+    continue_learning: False
     trained_policies_save_path: null
-    trained_policies_load_path: null
     max_bid_price: 100
     algorithm: matd3
     learning_rate: 0.001
-    training_episodes: 100
+    training_episodes: 10
     episodes_collecting_initial_experience: 5
-    train_freq: 100h
-    gradient_steps: 10
-    batch_size: 128
+    train_freq: 24h
+    gradient_steps: 1
+    batch_size: 256
     gamma: 0.99
     device: cpu
-    action_noise_schedule: linear
     noise_sigma: 0.1
     noise_scale: 1
     noise_dt: 1
     validation_episodes_interval: 5
+    early_stopping_steps: 10
+    early_stopping_threshold: 0.05
 
   markets_config:
     EOM:
       operator: EOM_operator
       product_type: energy
+      start_date: 2013-01-01 00:00
       products:
         - duration: 1h
           count: 1
