### 2025-04-15 8449c52
added test that ensures we do not get error in last run when training_episodes are defined too small
