### 2025-04-15 537d2b2
fix disabling initial exploration for loaded actors
