### 2025-04-15 c1b6aea
- move the check for min required episodes into the run_learning function
