### 2025-04-15 44f998a
- fix warning when in continue learning and startegies have alreayd been trained
