### 2025-04-15 e836f25
Fix loading sequence after DRL training (#573)
Closes #571 

## Description
This PR fixes the incorrect policy loading sequence after learning.
Previously, the best policies were mistakenly overwritten by the
policies from the last episode. This was initially mentioned in #572

## Changes Proposed
- Do not create the learning role and set it to `None` when not in
learning or evaluation mode.
- Remove the loading of inter-episodic data after learning is completed,
as it is redundant.
- Remove the loading of target networks from strategies, since this is
unused and is instead handled directly by
[…]
