+++ tests/test_matd3.py
@@ -1,0 +1,317 @@
+# SPDX-FileCopyrightText: ASSUME Developers
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+
+import os
+from copy import deepcopy
+
+import pytest
+import torch as th
+
+from assume.reinforcement_learning.learning_role import (
+    Learning,
+    LearningConfig,
+    LearningStrategy,
+)
+from assume.reinforcement_learning.learning_utils import get_hidden_sizes
+
+
+@pytest.fixture
+def base_learning_config() -> LearningConfig:
+    return {
+        "obs_dim": 10,
+        "act_dim": 3,
+        "train_freq": "1h",
+        "unique_obs_dim": 2,
+        "algorithm": "matd3",
+        "actor_architecture": "mlp",
+        "learning_mode": True,
+        "evaluation_mode": False,
+        "training_episodes": 1,
+        "episodes_collecting_initial_experience": 0,
+        "continue_learning": False,
+        "trained_policies_save_path": None,
+        "early_stopping_steps": 10,
+        "early_stopping_threshold": 0.05,
+        "learning_rate": 1e-4,
+        "batch_size": 100,
+        "tau": 0.005,
+        "gamma": 0.99,
+        "gradient_steps": 1,
+        "policy_delay": 2,
+        "target_policy_noise": 0.2,
+        "target_noise_clip": 0.5,
+    }
+
+
+@pytest.fixture(scope="function")
+def learning_role_n(base_learning_config) -> Learning:
+    config = base_learning_config.copy()
+    learn = Learning(config)
+    learn.rl_strats["agent_0"] = LearningStrategy(**config)
+    learn.rl_strats["agent_1"] = LearningStrategy(**config)
+    return learn
+
+
+@pytest.fixture(scope="function")
+def learning_role_n_plus_m(base_learning_config) -> Learning:
+    config = base_learning_config.copy()
+    learn = Learning(config)
+    learn.rl_strats["agent_0"] = LearningStrategy(**config)
+    learn.rl_strats["agent_1"] = LearningStrategy(**config)
+    learn.rl_strats["agent_2"] = LearningStrategy(**config)
+    return learn
+
+
+@pytest.fixture(scope="function")
+def saved_n_agent_model(learning_role_n, tmp_path) -> tuple[str, dict]:
+    learning_role_n.initialize_policy()
+    save_dir = tmp_path / "saved_model_n"
+    save_dir.mkdir(parents=True, exist_ok=True)
+    learning_role_n.rl_algorithm.save_params(directory=str(save_dir))
+    agent = learning_role_n.rl_strats["agent_0"]
+    return str(save_dir), {
+        "critic": agent.critics.state_dict(),
+        "actor": agent.actor.state_dict(),
+        "target_critic": agent.target_critics.state_dict(),
+        "target_actor": agent.actor_target.state_dict(),
+        "optimizer_critic": agent.critics.optimizer.state_dict(),
+        "optimizer_actor": agent.actor.optimizer.state_dict(),
+    }
+
+
+@pytest.fixture(scope="function")
+def saved_n_plus_m_agent_model(learning_role_n_plus_m, tmp_path) -> tuple[str, dict]:
+    learning_role_n_plus_m.initialize_policy()
+    save_dir = tmp_path / "saved_model_n_plus_m"
+    save_dir.mkdir(parents=True, exist_ok=True)
+    learning_role_n_plus_m.rl_algorithm.save_params(directory=str(save_dir))
+    agent = learning_role_n_plus_m.rl_strats["agent_0"]
+    return str(save_dir), {
+        "critic": agent.critics.state_dict(),
+        "actor": agent.actor.state_dict(),
+    }
+
+
+# -------------------- Utilities --------------------
+
+
[…]
