### 2025-04-15 8f6a0df
- fix torch in tests
