### 2025-04-15 3f0493a
- improve check for policy dimensionality - add test for this functionality
