### 2025-04-15 5f07cae
- add require learning marker
