### 2025-04-15 c0543e9
reformulate price calculation for bidding marginal utility curve of demand
