### 2025-04-15 c21595f
shift num_bids from demany.py to to demand_units.csv for better accesibility
