+++ assume/strategies/naive_strategies.py
@@ -5,6 +5,7 @@
 from assume.common.base import BaseStrategy, SupportsMinMax
 from assume.common.market_objects import MarketConfig, Order, Orderbook, Product
 
+
 class NaiveSingleBidStrategy(BaseStrategy):
     """
     A naive strategy that bids the marginal cost of the unit on the market.
@@ -82,8 +83,7 @@ class NaiveSingleBidStrategy(BaseStrategy):
         if "node" in market_config.additional_fields:
             return bids
         else:
-            #return self.remove_empty_bids(bids)
-            return bids
+            return self.remove_empty_bids(bids)
 
 
 class NaiveProfileStrategy(BaseStrategy):
@@ -359,12 +359,20 @@ class NaiveExchangeStrategy(BaseStrategy):
 
         return bids
 
-class NaiveMultiBidDemandStrategy(BaseStrategy):
+
+class ElasticDemandStrategy(BaseStrategy):
     """
-    A naive strategy of a Demand unit on the EOM that bids multiple bids with different prices and volumes
-    and hereby approximates a given marginal utility curve.
-    The linear (!) marginal utility curve is defined by the units max_price and elasticity.
+    A naive strategy of a demand unit that submits multiple bids to approximate
+    a marginal utility curve, based on linear or isoelastic demand theory.
+
+    - Linear model: P = P_max - slope * Q
+    - Isoelastic model: P = (Q/Q_max) ** (1/E)
+      (derived from log-log price elasticity of demand)
+
+    See:
+    - https://en.wikipedia.org/wiki/Price_elasticity_of_demand
     """
+
     def calculate_bids(
         self,
         unit: SupportsMinMax,
@@ -372,127 +380,96 @@ class NaiveMultiBidDemandStrategy(BaseStrategy):
         product_tuples: list[Product],
         **kwargs,
     ) -> Orderbook:
-        """
-        Takes information from a demand unit that the unit operator manages and
-        defines how it bids to the market.
-
-        The bids take the following form:
-        For each hour several bids (num_bids) are created with different prices and volumes.
-        The bids are created according to the marginal utility curve of the unit.
-
-        Args:
-            unit (SupportsMinMax): The unit to be buying energy.
-            market_config (MarketConfig): The market configuration.
-            product_tuples (list[Product]): The list of all products the unit can demand.
-            num_bids (int): The number of bids to be created for each hour.
-
-        Returns:
-            Orderbook: The bids consisting of the start time, end time, only hours, price and volume.
-        """
-        start = product_tuples[0][0]  # start time of the first product
-        end_all = product_tuples[-1][1]  # end time of the last product
-        
-        #min_power_values, max_power_values = unit.calculate_min_max_power(
-        #    start, end_all
-        #)  # minimum and maximum power demand of the unit between the start time of the first product and the end time of the last product
-        # this somehow does not yield reasonable results for this bidding strategy, so we use the units max_power instead
-        min_power_values = [unit.min_power] * len(product_tuples)
-        max_power_values = [unit.max_power] * len(product_tuples)
-
-        max_price = unit.max_price
-        elasticity = unit.elasticity
-        elasticity_model = 'isoelastic' #unit.elasticity_model
-        num_bids = unit.num_bids
+        if unit.elasticity == 0 or unit.num_bids == 1:
+            raise ValueError(
+                "ElasticDemandStrategy requires an elastic unit with num_bids > 1."
+            )
 
         bids = []
-        for product, min_power, max_power in zip(
-            product_tuples, min_power_values, max_power_values
-        ):
-            # for each product, calculate the  of the unit at the start time of the product
-            # and the volume of the product. Dispatch the order to the market.
-            start = product[0]
-            end = product[1]
-            
-            max_abs_power = max(abs(max_power), abs(min_power))
-            
[…]
