+++ assume/units/dst_components.py
@@ -1254,133 +1254,172 @@ class ElectricArcFurnace:
         return model_block
 
 
-class ElectricVehicle(GenericStorage):
-    """
-    A class to represent an Electric Vehicle (EV) unit in an energy system model.
-
-    The class encapsulates the parameters, variables, and constraints necessary to model
-    the behavior of an EV, including charging and discharging, battery storage limits, availability
-    profiles, and ramp rates.
-
-    Inherits from GenericStorage and adds EV-specific functionality, such as availability profiles
-    and predefined charging profiles.
-
-    Args:
-        max_capacity (float): Maximum capacity of the EV battery.
-        min_capacity (float): Minimum capacity of the EV battery.
-        max_power_charge (float): Maximum allowable charging power.
-        max_power_discharge (float): Maximum allowable discharging power. Defaults to 0 (no discharging allowed).
-        availability_profile (pd.Series): A pandas Series indicating the EV's availability, where 1 means available and 0 means unavailable.
-        time_steps (list[int]): A list of time steps over which the EV operates.
-        efficiency_charge (float, optional): Charging efficiency of the EV. Defaults to 1.0.
-        efficiency_discharge (float, optional): Discharging efficiency of the EV. Defaults to 1.0.
-        initial_soc (float, optional): Initial state of charge (SOC) of the EV, represented as a fraction of `max_capacity`. Defaults to 1.0.
-        ramp_up (float, optional): Maximum allowed increase in charging power per time step. Defaults to None (no ramp constraint).
-        ramp_down (float, optional): Maximum allowed decrease in charging power per time step. Defaults to None (no ramp constraint).
-        charging_profile (pd.Series | None, optional): A predefined charging profile. If provided, the EV follows this profile instead of optimizing the charge. Defaults to None.
-    """
-
+class ElectricVehicle:
     def __init__(
         self,
         max_capacity: float,
         time_steps: list[int],
-        availability_profile: pd.Series,
-        max_power_charge: float,
-        min_capacity: float = 0.0,
-        max_power_discharge: float = 0,
-        efficiency_charge: float = 1.0,
-        efficiency_discharge: float = 1.0,
-        initial_soc: float = 1.0,
+        max_power_charge: float | None = None,
+        max_power_discharge: float | None = None,
+        efficiency_charge: float = 0.95,
+        efficiency_discharge: float = 0.95,
+        initial_soc: float = 0.5,
+        mileage: float = 0.77,
+        availability_profile: pd.Series | None = None,
+        charging_profile: pd.Series | None = None,
+        power_flow_directionality: str = "unidirectional",  # or "bidirectional"
         ramp_up: float | None = None,
         ramp_down: float | None = None,
-        charging_profile: pd.Series | None = None,
-        storage_loss_rate: float = 0.0,
         **kwargs,
     ):
-        # Call the parent class (GenericStorage) __init__ method
-        super().__init__(
-            max_capacity=max_capacity,
-            time_steps=time_steps,
-            min_capacity=min_capacity,
-            max_power_charge=max_power_charge,
-            max_power_discharge=max_power_discharge,
-            efficiency_charge=efficiency_charge,
-            efficiency_discharge=efficiency_discharge,
-            initial_soc=initial_soc,
-            ramp_up=ramp_up,
-            ramp_down=ramp_down,
-            storage_loss_rate=storage_loss_rate,
-            **kwargs,
-        )
+        super().__init__()
 
-        # EV-specific attributes
+        # check if initial_soc is within the bounds [0, 1] and fix it if not
+        if initial_soc > 1:
+            initial_soc /= max_capacity
+            logger.warning(
+                f"Initial SOC is greater than 1.0. Setting it to {initial_soc}."
+            )
[…]
