+++ assume/units/dsm_load_shift.py
@@ -16,7 +16,7 @@ from pyomo.opt import (
 )
 
 from assume.common.fast_pandas import FastSeries
-from assume.units.dst_components import get_technology_class
+from assume.units.dst_components import demand_side_technologies
 
 SOLVERS = ["appsi_highs", "gurobi", "glpk", "cbc", "cplex"]
 
@@ -70,41 +70,36 @@ class DSMFlex:
 
         This method iterates over the provided components, instantiates their corresponding classes,
         and adds the respective blocks to the Pyomo model.
+
+        Args:
+            components (dict[str, dict]): A dictionary where each key is a technology name and
+                                        the value is a dictionary of parameters for the respective technology.
+                                        Each technology is mapped to a corresponding class in `demand_side_technologies`.
+
+        The method:
+        - Looks up the corresponding class for each technology in `demand_side_technologies`.
+        - Instantiates the class by passing the required parameters.
+        - Adds the resulting block to the model under the `dsm_blocks` attribute.
         """
         components = self.components.copy()
         self.model.dsm_blocks = pyo.Block(list(components.keys()))
 
         for technology, component_data in components.items():
-            try:
-                # Dynamically get the class for the technology
-                component_class = get_technology_class(technology)
+            if technology in demand_side_technologies:
+                # Get the class from the dictionary mapping (adjust `demand_side_technologies` to hold classes)
+                component_class = demand_side_technologies[technology]
 
-                # Instantiate the component with the required parameters
+                # Instantiate the component with the required parameters (unpack the component_data dictionary)
                 component_instance = component_class(
                     time_steps=self.model.time_steps, **component_data
                 )
-
                 # Add the component to the components dictionary
                 self.components[technology] = component_instance
 
-                # Optional external range binding
-                range_param_name = f"{technology}_range"
-                external_range = getattr(self.model, range_param_name, None)
-
-                if external_range is not None:
-                    component_instance.add_to_model(
-                        self.model,
-                        self.model.dsm_blocks[technology],
-                        external_range=external_range,
-                    )
-                else:
-                    component_instance.add_to_model(
-                        self.model, self.model.dsm_blocks[technology]
-                    )
-
-            except ValueError as e:
-                logger.error(f"Error initializing component {technology}: {e}")
-                raise
+                # Add the component's block to the model
+                component_instance.add_to_model(
+                    self.model, self.model.dsm_blocks[technology]
+                )
 
     def setup_model(self, presolve=True):
         # Initialize the Pyomo model
@@ -242,14 +237,17 @@ class DSMFlex:
                 if self.has_pv:
                     total_power_input -= self.model.dsm_blocks["pv_plant"].power[t]
 
-            elif self.technology == "bus_depot":
-                return m.total_power_input[t] + m.load_shift_pos[t] - m.load_shift_neg[
-                    t
-                ] == sum(
-                    self.model.dsm_blocks[ev].charge[t]
-                    - self.model.dsm_blocks[ev].discharge[t]
-                    for ev in self.model.dsm_blocks
-                    if ev.startswith("electric_vehicle")
+            elif self.technology == "paper_pulp_plant":
+                total_power = 0
+                if self.has_heat_pump:
+                    total_power += self.model.dsm_blocks["heat_pump"].power_in[t]
[…]
