+++ docs/source/release_notes.rst
@@ -12,6 +12,15 @@ Upcoming Release
   The features in this section are not released yet, but will be part of the next release! To use the features already you have to install the main branch,
   e.g. ``pip install git+https://github.com/assume-framework/assume``
 
+**New Features:**
+
+- **Bidding Strategy for Elastic Demand**: The new `ElasticDemandStrategy` enables demand units to submit multiple bids that approximate a marginal utility curve, using
+  either linear or isoelastic price elasticity models. Unlike other strategies, it does **not** rely on predefined volumes—bids are dynamically generated based on the
+  unit’s elasticity configuration. To use this strategy, set `bidding_strategy` to `"elastic_demand"` in the `demand_units.csv` file and specify the following
+  parameters: `elasticity` (must be negative), `elasticity_model` (`"linear"` or `"isoelastic"`), `num_bids`, and `price` (which acts as `max_price`). The `elasticity_model`
+  defines the shape of the demand curve, with `"linear"` producing a straight-line decrease and `"isoelastic"` generating a hyperbolic curve. `num_bids` determines how many
+  bid steps are submitted, allowing control over the granularity of demand flexibility.
+
 **Bug Fixes:**
 
 - **Last policy loading**: Fixed a bug where the last policy loaded after a training run was not the best policy, but rather the last policy.
