+++ tests/test_rl_strategies.py
@@ -9,7 +9,7 @@ import pytest
 
 try:
     from assume.reinforcement_learning.learning_role import LearningConfig
-    from assume.strategies.learning_strategies import RLStrategy
+    from assume.strategies.learning_strategies import RLStrategy, RLStrategySingleBid
 except ImportError:
     pass
 
@@ -21,7 +21,7 @@ end = datetime(2023, 7, 2)
 
 
 @pytest.fixture
-def power_plant_mcp() -> PowerPlant:
+def power_plant() -> PowerPlant:
     # Create a PowerPlant instance with some example parameters
     index = pd.date_range("2023-06-30 22:00:00", periods=48, freq="h")
     ff = NaiveForecast(index, availability=1, fuel_price=10, co2_price=10)
@@ -50,96 +50,65 @@ def power_plant_mcp() -> PowerPlant:
     )
 
 
-@pytest.fixture
-def power_plant_lstm() -> PowerPlant:
-    # Create a PowerPlant instance with some example parameters
-    index = pd.date_range("2023-06-30 22:00:00", periods=48, freq="h")
-    ff = NaiveForecast(index, availability=1, fuel_price=10, co2_price=10)
+@pytest.mark.require_learning
+@pytest.mark.parametrize(
+    "strategy_class, obs_dim, act_dim, actor_architecture, expected_bid_count, expected_volumes",
+    [
+        (RLStrategy, 50, 2, "mlp", 2, [200, 800]),
+        (RLStrategy, 50, 2, "lstm", 2, [200, 800]),
+        (RLStrategySingleBid, 74, 1, "mlp", 1, [1000]),
+    ],
+)
+def test_learning_strategies_parametrized(
+    mock_market_config,
+    power_plant,
+    strategy_class,
+    obs_dim,
+    act_dim,
+    actor_architecture,
+    expected_bid_count,
+    expected_volumes,
+):
+    product_index = pd.date_range("2023-07-01", periods=1, freq="h")
+    mc = mock_market_config
+    mc.product_type = "energy_eom"
+    product_tuples = [
+        (start, start + pd.Timedelta(hours=1), None) for start in product_index
+    ]
+
+    # Build learning config dynamically
     learning_config: LearningConfig = {
-        "observation_dimension": 50,
-        "action_dimension": 2,
+        "observation_dimension": obs_dim,
+        "action_dimension": act_dim,
         "algorithm": "matd3",
-        "actor_architecture": "lstm",
         "learning_mode": True,
         "training_episodes": 3,
-        "unit_id": "test_pp",
+        "unit_id": power_plant.id,
     }
+    if actor_architecture != "mlp":
+        learning_config["actor_architecture"] = actor_architecture
 
-    return PowerPlant(
-        id="test_pp",
-        unit_operator="test_operator",
-        technology="coal",
-        index=ff.index,
-        max_power=1000,
-        min_power=200,
-        efficiency=0.5,
-        additional_cost=10,
-        bidding_strategies={"EOM": RLStrategy(**learning_config)},
-        fuel_type="lignite",
-        emission_factor=0.5,
-        forecaster=ff,
-    )
+    # Override the strategy
+    power_plant.bidding_strategies["EOM"] = strategy_class(**learning_config)
 
+    strategy = power_plant.bidding_strategies["EOM"]
+    bids = strategy.calculate_bids(power_plant, mc, product_tuples=product_tuples)
 
-@pytest.mark.require_learning
-def test_learning_strategies(mock_market_config, power_plant_mcp):
-    product_index = pd.date_range("2023-07-01", periods=1, freq="h")
-    mc = mock_market_config
-    mc.product_type = "energy_eom"
-    product_tuples = [
[…]
