### 2025-04-16 57b7e32
add slightly more understandable version of error message
