### 2025-04-16 e894728
added explanation of P Q and E to avoid confusion with power P
