+++ tests/test_rl_strategies.py
@@ -11,9 +11,9 @@ try:
     from assume.reinforcement_learning.learning_role import LearningConfig
     from assume.strategies.learning_strategies import RLStrategy, RLStrategySingleBid
 
-    HAS_RL = True
 except ImportError:
-    HAS_RL = False
+    RLStrategy = None
+    RLStrategySingleBid = None
 
 from assume.common.forecasts import NaiveForecast
 from assume.units import PowerPlant
@@ -52,7 +52,7 @@ def power_plant() -> PowerPlant:
     )
 
 
-@pytest.mark.skipif(not HAS_RL, reason="Reinforcement learning module not available")
+@pytest.mark.require_learning
 @pytest.mark.parametrize(
     "strategy_class, obs_dim, act_dim, actor_architecture, expected_bid_count, expected_volumes",
     [
