### 2025-04-17 f215db6
- revert all exploratve changes to most smple integration - just add time dependend variables in obs space - increase exploration and increase training steps - increase NN size an ridicolous amount - all preparation  HiCORE run
