### 2025-04-17 43213ec
- cant work with docker  on HiCore
