+++ assume/reinforcement_learning/neural_network_architecture.py
@@ -146,7 +146,8 @@ class MLPActor(Actor):
         """Forward pass for action prediction."""
         x = F.relu(self.FC1(obs))
         x = F.relu(self.FC2(x))
-        x = F.softsign(self.FC3(x))
+        x = F.relu(self.FC3(x))
+        x = F.softsign(self.FC4(x))
 
         return x
 
