+++ examples/inputs/future_markets_paper/config.yaml
@@ -4,7 +4,7 @@
 
 base_case_2013:
   start_date: 2013-01-01 00:00
-  end_date: 2013-01-31 00:00
+  end_date: 2013-01-02 00:00
   time_step: 1h
   learning_mode: True
   save_frequency_hours: Null
