### 2025-04-18 2a7536b
- push monthly config to haicore
