### 2025-04-23 #7666 @kyleniemeyer
Hello @imcatta @SebastianBoblest and @KriNiTi would you be available to review this submission for JOSS? I'm reaching out specifically because of your past engagement with the AMIRIS submission (both as reviewer and author)—thanks!
