### 2025-04-24 #7666 @SebastianBoblest
Hi,

sorry I cannot review this time.

Sebastian Boblest
Schillerstrasse 19
73105 Dürnau
Tel.: 07164/9152167

Kyle Niemeyer ***@***.***> schrieb am Mi., 23. Apr. 2025,
19:37:

> *kyleniemeyer* left a comment (openjournals/joss-reviews#7666)
> <https://github.com/openjournals/joss-reviews/issues/7666#issuecomment-2825042595>
>
> Hello @imcatta <https://github.com/imcatta> @SebastianBoblest
> <https://github.com/SebastianBoblest> and @KriNiTi
> <https://github.com/KriNiTi> would you be available to review this
> submission for JOSS? I'm reaching out specifically because of your past
> engagement with the AMIRIS submission (both as reviewer and author)—thanks!
>
