### 2025-04-28 46d52cd
- store a map of the order of the units when saving critics - use the map for loading and weight transfer if needed when loading the critics - added a test to check the weight transfer logic for different configurations
