+++ assume/reinforcement_learning/learning_utils.py
@@ -3,14 +3,12 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import logging
-import math
 from collections.abc import Callable
 from datetime import datetime
 from typing import TypedDict
 
 import numpy as np
 import torch as th
-from torch import nn
 
 logger = logging.getLogger(__name__)
 
@@ -184,36 +182,6 @@ def get_hidden_sizes(state_dict: dict, prefix: str) -> list[int]:
     return sizes[:-1]  # exclude the final output layer if needed
 
 
-def infer_n_agents(state_dict, obs_base, act_dim, unique_obs_dim):
-    key = "q1_layers.0.weight"
-    if key not in state_dict:
-        raise KeyError(f"Missing expected key '{key}' in state_dict.")
-
-    input_dim = state_dict[key].shape[1]
-
-    # If both are 0, cannot infer
-    denom = unique_obs_dim + act_dim
-    if denom == 0:
-        raise ValueError("Cannot infer N: unique_obs_dim and act_dim both zero.")
-
-    # Proper formula derived from CriticTD3 logic
-    num = input_dim - obs_base + unique_obs_dim
-    n_est = num / denom
-    n_rounded = round(n_est)
-
-    # Reconstruct the original input_dim as CriticTD3 would
-    reconstructed_input = (
-        obs_base + unique_obs_dim * (n_rounded - 1) + act_dim * n_rounded
-    )
-
-    if not math.isclose(reconstructed_input, input_dim) or n_rounded <= 0:
-        raise ValueError(
-            f"Inferred N={n_rounded} is invalid (Recalculated input={reconstructed_input}, got={input_dim})"
-        )
-
-    return n_rounded
-
-
 def copy_layer_data(dst, src):
     for k in dst:
         if k in src and dst[k].shape == src[k].shape:
@@ -221,75 +189,80 @@ def copy_layer_data(dst, src):
 
 
 def transfer_weights(
-    model: nn.Module,
+    model: th.nn.Module,
     old_state: dict,
-    old_n_agents: int,
-    new_n_agents: int,
+    old_id_order: list[str],
+    new_id_order: list[str],
     obs_base: int,
     act_dim: int,
     unique_obs: int,
 ) -> dict | None:
-    # Check architecture compatibility (extract hidden sizes from state_dicts)
+    """
+    Copy only those obs_ and action-slices for matching IDs.
+    New IDs keep their original (random) weights.
+    """
+    # 1) Architecture check
+    new_state = model.state_dict()
     old_hidden = get_hidden_sizes(old_state, prefix="q1_layers")
-    new_hidden = get_hidden_sizes(model.state_dict(), prefix="q1_layers")
+    new_hidden = get_hidden_sizes(new_state, prefix="q1_layers")
     if old_hidden != new_hidden:
         logger.warning(
             f"Cannot transfer weights: architecture mismatch.\n"
-            f"Old sizes of hidden layers: {old_hidden}, New sizes of hidden layers: {new_hidden}. Skipping transfer."
+            f"Old sizes: {old_hidden}, New sizes: {new_hidden}."
         )
         return None
 
-    new_template = model.state_dict()
-    new_state = {k: v.clone() for k, v in new_template.items()}
+    # 2) Compute total dims
+    old_n = len(old_id_order)
+    new_n = len(new_id_order)
+    old_obs_tot = obs_base + unique_obs * max(0, old_n - 1)
+    new_obs_tot = obs_base + unique_obs * max(0, new_n - 1)
 
-    # Calculate complete dimensions from both old and new agent counts
-    old_obs = obs_base + unique_obs * max(0, old_n_agents - 1)
-    new_obs = obs_base + unique_obs * max(0, new_n_agents - 1)
+    # 3) Clone new state
[…]
