### 2025-04-28 608f312
- add require learning attribute
