### 2025-04-28 11cfb11
- improve logging
