### 2025-04-28 8140fd0
add better description of first_block_bid and isoelastic demand
