+++ assume/strategies/naive_strategies.py
@@ -366,12 +366,15 @@ class ElasticDemandStrategy(BaseStrategy):
     a marginal utility curve, based on linear or isoelastic demand theory. 
     P = Price, Q = Quantity, E = Elasticity.
 
-    - Linear model: P = P_max - slope * Q
-    - Isoelastic model: P = (Q/Q_max) ** (1/E)
+    - Linear model: P = P_max + slope * Q (slope is only defined by P_max and Q_max, negative value)
+    - Isoelastic model: P = (Q/Q_max) ** (1/E) (E is negative)
       (derived from log-log price elasticity of demand)
 
     See:
     - https://en.wikipedia.org/wiki/Price_elasticity_of_demand
+    - Arnold, Fabian. 2023. https://hdl.handle.net/10419/286380
+    - Hirth, Lion et al. 2024. https://doi.org/10.1016/j.eneco.2024.107652.
+
     """
 
     def calculate_bids(
@@ -396,59 +399,91 @@ class ElasticDemandStrategy(BaseStrategy):
             start, end, only_hours = product
             max_abs_power = max(abs(unit.min_power), abs(unit.max_power))
 
-            first_bid_volume = self.find_first_block_bid(
-                elasticity, max_price, max_abs_power
-            )
-            bids.append(
-                {
-                    "start_time": start,
-                    "end_time": end,
-                    "only_hours": only_hours,
-                    "price": max_price,
-                    "volume": -first_bid_volume,
-                    "node": unit.node,
-                }
-            )
+            if elasticity_model == "isoelastic":
+                # ISOELASTIC model (constant elasticity of demand):
+                # P = (Q / Q_max)^(1 / E)
+                # Derived from:
+                #   E = (dQ/dP) * (P / Q(P)), E == constant
+                #   E = Q'(P) * (P / Q(P))
+                #   Q'(P) = E * 1/P * Q(P)
+                #   dQ/dP = E * (1 / P) * Q
+                #   integrate:
+                #   \int 1/Q dQ = E * \int 1/P dP
+                #   ln(Q) = E * ln(P) + C
+                #   exp(ln(Q)) = exp(E * ln(p) + C)  
+                #   Q(p) = (exp(ln(p))^E) * exp(C)
+                #   Q(p) = P^E * exp(C)
+                #   possibly C = 0, C = 1 or C >= 1. We assume C >= 1.
+                #   C shifts the demand curve (demand vs. price) up / down and
+                #   can be interpreted as the demand that will always be served
+                #   we set C = ln(Q_max) and derive the demand curve from there:
+                #   P = Q^(1/E) * exp(-C/E) and C = ln(Q_max) 
+                #   => P = Q^(1/E) * exp(-ln(Q_max)/E) and because exp(-ln(Q_max)/E) = Q_max^(-1/E)
+                #   => P = Q^(1/E) * Q_max^(-1/E)
+                #   finally
+                #   => P = (Q / Q_max)^(1/E)
+                # This yields decreasing price as volume increases
 
-            remaining_volume = max_abs_power - first_bid_volume
-            if remaining_volume < 0:
-                raise ValueError(
-                    "Negative remaining volume for bidding. Check max_power and elasticity."
+                # calculate first bid in isoelastic model (the volume that is bid at max price)
+                first_bid_volume = self.find_first_block_bid(
+                    elasticity, max_price, max_abs_power
                 )
+                bids.append(
+                    {
+                        "start_time": start,
+                        "end_time": end,
+                        "only_hours": only_hours,
+                        "price": max_price,
+                        "volume": -first_bid_volume,
+                        "node": unit.node,
+                    }
+                )
+            
+                remaining_volume = max_abs_power - first_bid_volume
+                if remaining_volume < 0:
[…]
