### 2025-04-28 bf69173
remove 0 from range obj to pass test
