+++ assume/reinforcement_learning/algorithms/__init__.py
@@ -6,14 +6,14 @@ from torch import nn
 
 from assume.reinforcement_learning.neural_network_architecture import (
     MLPActor,
-    ContextualMLPActor,
-    FilmActor,
+    ContextualLateFusionActor,
+    ContextualFiLMActor,
     LSTMActor,
 )
 
 actor_architecture_aliases: dict[str, type[nn.Module]] = {
     "mlp": MLPActor,
-    "contextual_mlp": ContextualMLPActor,
-    "film_actor": FilmActor,
+    "contextual_late_fusion": ContextualLateFusionActor,
+    "contextual_film": ContextualFiLMActor,
     "lstm": LSTMActor,
 }
