### 2025-04-29 7aebf11
Merge continue learning into the branch
