### 2025-04-29 67b7b16
added check for attribute elasticity_model and elasticity if 'isoelastic' model
