### 2025-04-29 f27bbcf
make checks concise, better follow of object oriented paradigm
