### 2025-04-29 ff42c37
add tests for valid and invalid linear model
