+++ assume/strategies/learning_strategies.py
@@ -78,7 +78,7 @@ class BaseLearningStrategy(LearningStrategy):
         if self.actor_architecture in ["mlp", "lstm"]:
             directory = f"{load_path}/actors/actor_{self.unit_id}.pt"
 
-        elif self.actor_architecture in ["contextual_late_fusion", "contextual_film"]:
+        elif self.actor_architecture in ["contextual_fusion", "contextual_film"]:
             # first load the cluster mapping json
             cluster_mapping_path = f"{load_path}/actors/cluster_mapping.json"
             with open(cluster_mapping_path) as f:
