+++ assume/reinforcement_learning/algorithms/__init__.py
@@ -13,7 +13,7 @@ from assume.reinforcement_learning.neural_network_architecture import (
 
 actor_architecture_aliases: dict[str, type[nn.Module]] = {
     "mlp": MLPActor,
-    "contextual_late_fusion": ContextualLateFusionActor,
+    "contextual_fusion": ContextualLateFusionActor,
     "contextual_film": ContextualFiLMActor,
     "lstm": LSTMActor,
 }
