### 2025-04-30 20e2405
Improve shared-actor loading and clustering robustness
- Add mismatch detection in load_shared_actor_params: log and revert units moved by new clustering
- Consolidate cluster mapping rebuild and actor instantiation helpers
- Handle missing actors/ directory and missing cluster_mapping.json gracefully
- Ensure brand-new clusters remain uninitialized while old clusters load weights
- Add comprehensive pytest coverage for shared-actor edge cases (no mapping, no actors dir, no mismatches, new clusters, single-strategy & max-size clustering)
