+++ tests/test_matd3.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import json
+import logging
 import os
 from copy import deepcopy
 
@@ -534,3 +535,360 @@ def test_transfer_weights_various_orders(prefix, new_id_order):
         b_key = f"{prefix}.{i}.bias"
         assert th.equal(new_state[w_key], old_state[w_key])
         assert th.equal(new_state[b_key], old_state[b_key])
+
+
+@pytest.fixture
+def clustering_config(base_learning_config):
+    cfg = base_learning_config.copy()
+    cfg.update(
+        {
+            "actor_architecture": "contextual_fusion",
+            "actor_clustering_method": "k-means",
+            "clustering_method_kwargs": {"n_clusters": 2, "random_state": 42},
+            "context_dim": 2,
+        }
+    )
+    return cfg
+
+
+@pytest.mark.require_learning
+def test_shared_actor_clustering_and_mapping(clustering_config):
+    # 1) build Learning + 3 strategies
+    learn = Learning(clustering_config)
+    contexts = {
+        "A": th.tensor([1.0, 0.0]),
+        "B": th.tensor([1.0, 0.0]),
+        "C": th.tensor([0.0, 1.0]),
+    }
+    for uid, ctx in contexts.items():
+        strategy = LearningStrategy(**clustering_config)
+        strategy.unit_id = uid
+        strategy.context = ctx
+        learn.rl_strats[uid] = strategy
+
+    # 2) initialize → creates clusters on rl_algorithm
+    learn.initialize_policy()
+
+    alg = learn.rl_algorithm
+
+    # 3) mapping covers all unit_ids
+    assert set(alg.cluster_mapping.keys()) == set(contexts.keys())
+
+    # 4) we expect two clusters by KMeans
+    assert len(alg.clusters) == 2
+
+    # 5) A and B in same cluster; C in a different one
+    clA = alg.cluster_mapping["A"]
+    clB = alg.cluster_mapping["B"]
+    clC = alg.cluster_mapping["C"]
+    assert clA == clB
+    assert clC != clA
+
+    # 6) clusters dict matches mapping
+    assert {s.unit_id for s in alg.clusters[clA]} == {"A", "B"}
+    assert {s.unit_id for s in alg.clusters[clC]} == {"C"}
+
+    # 7) shared_actors keys match clusters, and instances are shared
+    assert set(alg.shared_actors.keys()) == set(alg.clusters.keys())
+    actorAB = alg.shared_actors[clA]
+    actorC = alg.shared_actors[clC]
+    assert learn.rl_strats["A"].actor is actorAB
+    assert learn.rl_strats["B"].actor is actorAB
+    assert learn.rl_strats["C"].actor is actorC
+    assert actorAB is not actorC
+
+
+@pytest.mark.require_learning
+def test_load_shared_actor_params_mismatch_and_new_unit(
+    clustering_config, tmp_path, caplog
+):
+    # --- 1) First run: create and save model with 3 units ---
+    learn1 = Learning(clustering_config)
+    contexts1 = {
+        "u1": th.tensor([1.0, 0.0]),
+        "u2": th.tensor([1.0, 0.0]),
+        "u3": th.tensor([0.0, 1.0]),
+    }
+    for uid, ctx in contexts1.items():
+        strategy = LearningStrategy(**clustering_config)
+        strategy.unit_id = uid
+        strategy.context = ctx
+        learn1.rl_strats[uid] = strategy
+    learn1.initialize_policy()
+
+    save_dir = tmp_path / "shared_model"
+    save_dir.mkdir()
+    learn1.rl_algorithm.save_params(directory=str(save_dir))
+    old_map = learn1.rl_algorithm.cluster_mapping.copy()
+
[…]
