### 2025-04-30 0091e52
- fix tests - improve logging
