### 2025-04-30 #7666 @KriNiTi
Hello @kyleniemeyer,

Thank you for reaching out. Unfortunately, I do not have sufficient software development skills to review the submission. Please contact my colleague Christoph Schimeczek, dlr-cjs (cc’ed), if a joint publication with the first author of the submission, Florian Maurer, is not a reason for exclusion (see https://ieeexplore.ieee.org/stamp/stamp.jsp?arnumber=10609021).

Kind regards,

Kristina

Von: Kyle Niemeyer ***@***.***>
Gesendet: Mittwoch, 23. April 2025 19:37
An: openjournals/joss-reviews ***@***.***>
Cc: Nienhaus, Kristina ***@***.***>; Mention ***@***.***>
Betreff: Re: [openjournals/joss-reviews] [PRE REVIEW]: ASSUME - Agent-based Simulation for Studying and Understanding Market Evolution (Issue #7666)

[https://avatars.githubusercontent.com/u/190432?s=20&v=4]kyleniemeyer left a comment (openjournals/joss-reviews#7666)<https://github.com/openjournals/joss-reviews/issues/7666#issuecomment-2825042595>

Hello @imcatta<https://github.com/imcatta> @SebastianBoblest<https://github.com/SebastianBoblest> and @KriNiTi<https://github.com/KriNiTi> would you be available to review this submission for JOSS? I'm reaching out specifically because of your past engagement with the AMIRIS submission (both as reviewer and author)—thanks!
