### 2025-05-01 121771e
- add args to docstrings - renamed old to loaded as it seems more speaking - added comments and error messages, to be clearer to first viewers of code - added a doc section that explains general concept of continue learning
