### 2025-05-02 f2279bc
added docu of elastic_demand strategy
