### 2025-05-02 7db86d1
modified:   assume/strategies/naive_strategies.py
