### 2025-05-02 38cf098
modified:   assume/units/powerplant.py
