### 2025-05-02 e70ffae
modified:   assume/strategies/__init__.py 	modified:   assume/strategies/naive_strategies.py 	modified:   examples/examples.py 	new example added:   examples/inputs/example_05e
