+++ examples/examples.py
@@ -13,7 +13,7 @@ log = logging.getLogger(__name__)
 
 # define the path where the simulation results in form of CSV files will be stored, for example: "examples/outputs"
 # "" means no CSV files will be stored
-csv_path = ""
+csv_path = "outputs"
 
 os.makedirs("./examples/local_db", exist_ok=True)
 
@@ -105,6 +105,8 @@ available_examples = {
         "scenario": "example_04",
         "study_case": "base_case_2019",
     },
+    # redispatch for DSM validation example
+    "redisp_dsm_3_nodes": {"scenario": "example_05e", "study_case": "base"},
 }
 
 
@@ -121,7 +123,7 @@ if __name__ == "__main__":
     data_format = "local_db"  # "local_db" or "timescale"
 
     # select the example to run from the available examples above
-    example = "paper_plant_use_case"
+    example = "redisp_dsm_3_nodes"
 
     if data_format == "local_db":
         db_uri = "sqlite:///./examples/local_db/assume_db.db"
