+++ assume/strategies/naive_strategies.py
@@ -292,7 +292,74 @@ class NaiveRedispatchStrategy(BaseStrategy):
 
         return bids
 
+class NaiveRedispatchSteelplantStrategy(BaseStrategy):
+    def calculate_bids(
+        self,
+        unit: SupportsMinMax,
+        market_config: MarketConfig,
+        product_tuples: list[Product],
+        **kwargs,
+    ) -> Orderbook:
+        # calculate the optimal operation of the unit according to the objective function
+        unit.calculate_optimal_operation_if_needed()
 
+        bids = []
+        for product in product_tuples:
+            """
+            for each product, calculate the marginal cost of the unit at the start time of the product
+            and the volume of the product. Dispatch the order to the market.
+            """
+            start = product[0]
+            volume = abs(unit.outputs["energy"].at[start])
+            # volume = unit.opt_power_requirement.loc[start]
+            volume_flex = unit.flex_power_requirement.loc[start]
+            marginal_price = unit.calculate_marginal_cost(start, volume_flex)
+            if volume > volume_flex:
+                bids.append(
+                    {
+                        "start_time": product[0],
+                        "end_time": product[1],
+                        "only_hours": product[2],
+                        "price": marginal_price,
+                        "volume": -volume,
+                        # "max_power": -max(volume, volume_flex),
+                        # "min_power": -min(volume, volume_flex),
+                        "max_power": -volume,
+                        "min_power": -volume_flex,
+                        "node": unit.node,
+                    }
+                )
+            elif volume < volume_flex:
+                bids.append(
+                    {
+                        "start_time": product[0],
+                        "end_time": product[1],
+                        "only_hours": product[2],
+                        "price": marginal_price,
+                        "volume": -volume,
+                        # "max_power": -max(volume, volume_flex),
+                        # "min_power": -min(volume, volume_flex),
+                        "max_power": -volume_flex,
+                        "min_power": -volume,
+                        "node": unit.node,
+                    }
+                )
+            else:
+                bids.append(
+                    {
+                        "start_time": product[0],
+                        "end_time": product[1],
+                        "only_hours": product[2],
+                        "price": marginal_price,
+                        "volume": -volume,
+                        "max_power": 0,
+                        "min_power": 0,
+                        "node": unit.node,
+                    }
+                )
+
+        return bids
+    
 class NaiveExchangeStrategy(BaseStrategy):
     """
     A naive strategy for an exchange unit that bids the defined import and export prices on the market.
