### 2025-05-05 b0541eb
- improve cluster loading
