### 2025-05-05 3720575
- improve pre-commit checks
