### 2025-06-30 #8515 @editorialbot
[REVIEW]: ASSUME - Agent-based Simulation for Studying and Understanding Market Evolution
**Submitting author:** @maurerle (<a href="http://orcid.org/0000-0001-8345-3889">Florian Maurer</a>)
**Repository:** https://github.com/assume-framework/assume
**Branch with paper.md** (empty if default branch): main
**Version:** v0.5.5
