### 2025-07-01 #8515 @dlr-cjs
@kyleniemeyer I further want to disclose that Manish Khanra and I (as well as a third person) hosted a 1.5-hour interactive workshop at a conference this year and made the results available at [zenodo](https://zenodo.org/records/15172712). I am unsure if this counts as publication with regard to a conflict of interest.
