### 2025-07-05 a17f24f
steam generation plant, Bidding strategies and flexibility measure (#593)
Introduced a 'SteamGenerationPlant' class to
model steam generation processes. This class supports both electric and
thermal inputs, allowing for flexible operation based on available
resources. The plant can be configured with various components, such as
heat pumps and boilers, to optimize steam production.

**New Demand Side Flexibility Measure** Implemented
'symmetric_flexible_block' flexibility measure for demand side units.
This measure allows users to define a symmetric block of flexibility,
enabling to construct a
[…]
