+++ assume/strategies/naive_strategies.py
@@ -2,6 +2,7 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
+
 from assume.common.base import BaseStrategy, SupportsMinMax
 from assume.common.market_objects import MarketConfig, Order, Orderbook, Product
 
@@ -169,8 +170,9 @@ class NaiveDADSMStrategy(BaseStrategy):
         """
 
         # check if unit has opt_power_requirement attribute
-        if not hasattr(unit, "opt_power_requirement"):
-            unit.determine_optimal_operation_without_flex()
+        if unit.optimisation_counter == 0:
+            unit.determine_optimal_operation_with_flex()
+            unit.optimisation_counter = 1
 
         bids = []
         for product in product_tuples:
@@ -508,3 +510,79 @@ class ElasticDemandStrategy(BaseStrategy):
                 f"Calculated first block bid volume ({volume}) exceeds max power ({max_power})."
             )
         return volume
+
+
+class DSM_PosCRM_Strategy(BaseStrategy):
+    """
+    Strategy for Positive CRM Reserve (Demand Side, i.e., up & down, symmetric).
+    """
+
+    def calculate_bids(self, unit, market_config, product_tuples, **kwargs):
+        bids = []
+        max_power = unit.max_plant_capacity
+        min_power = unit.min_plant_capacity
+
+        for product in product_tuples:
+            start, end, only_hours = product
+            block_times = [dt for dt in unit.index.get_date_list() if start <= dt < end]
+
+            # For all time steps in block, calculate possible symmetric bid
+            up_caps = []
+            down_caps = []
+            for t in block_times:
+                flex = unit.flex_power_requirement.at[t]
+                up_caps.append(max_power - flex)
+                down_caps.append(flex - min_power)
+            # The symmetric bid is the minimum capacity that is possible in *all* timesteps in the block
+            symmetric_capacity = min(min(up_caps), min(down_caps))
+            if symmetric_capacity > 0:
+                bids.append(
+                    {
+                        "start_time": start,
+                        "end_time": end,
+                        "only_hours": only_hours,
+                        "price": 0,  # or unit.calculate_marginal_cost(...)
+                        "volume": symmetric_capacity,
+                        "unit_id": unit.id,
+                        "market_id": "CRM_pos",
+                    }
+                )
+        return self.remove_empty_bids(bids)
+
+
+class DSM_NegCRM_Strategy(BaseStrategy):
+    """
+    Strategy for Negative CRM Reserve (Demand Side, i.e., up & down, symmetric).
+    """
+
+    def calculate_bids(self, unit, market_config, product_tuples, **kwargs):
+        # IDENTICAL LOGIC as POS, since symmetric in Germany (volume is symmetric)
+        # If you ever want to do *only* neg or pos (asymmetric), just change which cap you use!
+        bids = []
+        max_power = unit.max_plant_capacity
+        min_power = unit.min_plant_capacity
+
+        for product in product_tuples:
+            start, end, only_hours = product
+            block_times = [dt for dt in unit.index.get_date_list() if start <= dt < end]
+
+            up_caps = []
+            down_caps = []
+            for t in block_times:
+                flex = unit.flex_power_requirement.at[t]
+                up_caps.append(max_power - flex)
+                down_caps.append(flex - min_power)
+            symmetric_capacity = min(min(up_caps), min(down_caps))
+            if symmetric_capacity > 0:
+                bids.append(
+                    {
+                        "start_time": start,
+                        "end_time": end,
+                        "only_hours": only_hours,
+                        "price": 0,
+                        "volume": symmetric_capacity,
+                        "unit_id": unit.id,
+                        "market_id": "CRM_neg",
+                    }
+                )
+        return self.remove_empty_bids(bids)
