### 2025-07-06 assume-framework/assume#597 @dlr-cjs
README: Add a statement of need
Currently, the README only gives an abstract impression on that ASSUME can do.

As asked by the JOSS checklist [here](https://github.com/openjournals/joss-reviews/issues/8515#issuecomment-3020070777), I suggest to add a few senteces on "what problems the software is designed to solve" as wellas "who the target audience is" under the introduction section of the README main page.
