### 2025-07-06 #8515 @dlr-cjs
I ask to extend the "statement of need" part in the README and docs [here](https://github.com/assume-framework/assume/issues/597)
