### 2025-07-06 assume-framework/assume#598 @dlr-cjs
CONTRIBUTING.md
As part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8515) I suggest to add a "CONTRIBUTING.md" file to the repository that convers:

* How to contribute to the software
* How to report issues
* How to seek support

The first of these aspects is currently covered in the README, but in my opinion would be best addressed in a dedicated file to avoid cluttering the README file (see e.g. [here](https://fosdem.org/2025/schedule/event/fosdem-2025-4990-how-to-write-a-killer-readme/)).

In addition, you could also enhance the CONTRIBUTING.md following, e.g., this [guidline](
https://mozillascience.github.io/working-open-workshop/contributing).

Alternatively, consider placing a corresponding section in the documentation.
