### 2025-07-06 #8515 @dlr-cjs
[Here](https://github.com/assume-framework/assume/issues/598), I ask to create an explicit section on how to contribute - either the repository base folder or the documentation.
