### 2025-07-06 assume-framework/assume#600 @dlr-cjs
Example3 Results: Columns mixed in unit_dispatch.csv
This issue is part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8515).

I ran the example3 dam_case and the resulting "unit_dispatch.csv" file seems to be broken. About half-way in the file, the rows have a different column order.

I ran `assume -s example_03 -c dam_case_2019 -db "sqlite:///./examples/local_db/assume_db.db" -csv outputs` using
* Windows,
* Python=3.12,
* assume-framework=0.5.3, and
* data from assume commit a17f24f

The resulting `unit_dispatch.csv` starts with a plausible structure:

```csv
time,power,heat,energy_generation_costs,unit,soc,energy_cashflow,total_costs,simulation
2019-01-01 12:00:00,0,0,0,LUENEN 2 STUMMHAFEN,,,,example_03_dam_case_2019
```
but beginning with line 142,859 the column ordering seems to be off:
```csv
2019-01-21 13:00:00,750,0,28508,34058,28508,LUENEN 2 STUMMHAFEN,,example_03_dam_case_2019
```

Maybe this is related to #539 or #501 ?
