### 2025-07-06 assume-framework/assume#600 @dlr-cjs
This problems seems also to occur with the "base_case" in the same scenario.
