### 2025-07-07 assume-framework/assume#601 @dlr-cjs
Docs/Installation: Add hint about "network" installation
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8515).

Currently the [installation documentation page ](https://assume.readthedocs.io/en/latest/installation.html) does not cover all installation variants necessary to run the examples: At least the "[network]" variant installation is missing, that is required to run "examples01d".

Also, it would be helpful to list the required assume components on the [examples overview table](https://assume.readthedocs.io/en/latest/example_simulations.html) in an extra column.
