### 2025-07-07 assume-framework/assume#602 @dlr-cjs
Example_01d Crash
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8515).

When running example_01d like so `assume -s example_01d -c base -db "sqlite:///./examples/local_db/assume_db.db" -csv outputs`, I encounter a crash about one day into the simulation, where the original exception is raised by some function in linopy.

The full log is attached here:

[example_01d_log.txt](https://github.com/user-attachments/files/21097944/example_01d_log.txt)

I am using Windows 11, assume-framework 0.5.3 with extensions [optimize,learning,network], see also the [package_list.txt](https://github.com/user-attachments/files/21097979/package_list.txt)

The "zonal_case" example does not crash with the same setup.
