### 2025-07-08 assume-framework/assume#600 @dlr-cjs
I confirm - problem has been solved for me.
