### 2025-07-23 assume-framework/assume#619 @igarizio
Error when running 10_DSU_and_flexibility
When running this example ([10_DSU_and_flexibility](https://github.com/assume-framework/assume/blob/main/examples/notebooks/10_DSU_and_flexibility.ipynb)) the code fails in the section "cleaning dispatch data from the simulation to feed it to the PyPSA network". The actual [line that fails is](https://github.com/assume-framework/assume/blob/main/examples/notebooks/10_DSU_and_flexibility.ipynb?short_path=9c34160#L2811) (the next cell contains the same problem):
`redisp_p_set = redisp_p_set.drop(columns=["demand_EOM1", "demand_EOM2"])`
> KeyError: "['demand_EOM1', 'demand_EOM2'] not found in axis"

I think this is just because the previous line creates an empty dataframe:
```python
redispatches = market_dispatch_total[
    market_dispatch_total["market_id"].str.contains("redispatch", case=False, na=False)
]
```
Is it possible that `market_dispatch_total["market_id"]` should be `market_dispatch_total["simulation"]`?
