### 2025-07-23 #8515 @igarizio
Hi, sorry for the delay. I have only found a [single issue](https://github.com/assume-framework/assume/issues/619) in one of the examples (looks very simple to fix, but I could be wrong). Everything else looks great.
