### 2025-07-30 #8515 @kyleniemeyer
Great to hear, @igarizio! I can see the interaction over in that issue. @maurerle, have you considered expanding the CI to cover multiple operating systems? Just wondering if that could be a reason for different behavior.
