### 2025-07-31 #8515 @maurerle
Good idea @kyleniemeyer
I had in mind that I could reproduce this issue on linux as well, but since my fresh env install it does not happen anymore..
I am quite sure it is a dependency which was broken.
But I dont have the broken env anymore :/
