### 2025-08-01 assume-framework/assume#619 @maurerle
Hi @igarizio ,

I tried again today to reproduce this issue but could not do this in any of my older python envs.
Maybe reinstalling `pandas`, `numpy` and/or `pypsa` did help somehow..?

If you can still reproduce this issue, can you send a `pip freeze` output of your env and maybe try if it works in a fresh env on your side?

Thank you very much for your efforts!
