### 2025-08-02 assume-framework/assume#619 @igarizio
Hi @maurerle, thanks for looking into this!

Sorry, I should have given you way more details when creating this issue (I assumed it was just a column name mix-up, but I think it's something else).
I was running this example in colab (from the link provided [here](https://assume.readthedocs.io/en/latest/examples/10_DSU_and_flexibility.html)). This is what I did: I ran the first cell, waited until it finished, restarted the session, and ran everything again. I saved the outputs [here](https://colab.research.google.com/drive/1unmegSsIabM_kWuxb-_A1cFE4UKa_sc0#scrollTo=CfQv2GupDkKa).

As I mentioned above, the error that killed the notebook was `KeyError: "['demand_EOM1', 'demand_EOM2'] not found in axis"` but I had not realized that the previous cell had a silent error (I believe from `world.run()`). Here's that error:
```
tutorial_10_redispatch_case 2019-01-01 00:00:00:  21%|██        | 72001/342000 [00:00<00:02, 100304.39it/s]ERROR:assume.markets.base_market:clearing failed: fileno
ERROR:mango.util.scheduling:got exception in scheduled event
Traceback (most recent call last):
  File "/usr/local/lib/python3.11/dist-packages/mango/util/scheduling.py", line 31, in _raise_exceptions
    raise fut.exception()
  File "/usr/lib/python3.11/asyncio/tasks.py", line 277, in __step
    result = coro.send(None)
             ^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/mango/util/scheduling.py", line 211, in run
    return await self._coro
           ^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/assume/markets/base_market.py", line 638, in clear_market
    raise e
  File "/usr/local/lib/python3.11/dist-packages/assume/markets/base_market.py", line 633, in clear_market
    (accepted_orderbook, rejected_orderbook, market_meta, flows) = self.clear(
                                                                   ^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/assume/markets/clearing_algorithms/redispatch.py", line 194, in clear
    with suppress_output():
  File "/usr/lib/python3.11/contextlib.py", line 137, in __enter__
    return next(self.gen)
           ^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/assume/common/utils.py", line 640, in suppress_output
    original_stdout_fd = sys.stdout.fileno()
                         ^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/ipykernel/iostream.py", line 311, in fileno
    raise io.UnsupportedOperation("fileno")
io.UnsupportedOperation: fileno
tutorial_10_redispatch_case 2019-01-02 00:00:00:  46%|████▋     | 158401/342000 [00:01<00:01, 121387.75it/s]ERROR:assume.markets.base_market:clearing failed: fileno
ERROR:mango.util.scheduling:got exception in scheduled event
Traceback (most recent call last):
  File "/usr/local/lib/python3.11/dist-packages/mango/util/scheduling.py", line 31, in _raise_exceptions
    raise fut.exception()
  File "/usr/lib/python3.11/asyncio/tasks.py", line 277, in __step
    result = coro.send(None)
             ^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/mango/util/scheduling.py", line 211, in run
    return await self._coro
           ^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/assume/markets/base_market.py", line 638, in clear_market
    raise e
  File "/usr/local/lib/python3.11/dist-packages/assume/markets/base_market.py", line 633, in clear_market
    (accepted_orderbook, rejected_orderbook, market_meta, flows) = self.clear(
                                                                   ^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/assume/markets/clearing_algorithms/redispatch.py", line 194, in clear
    with suppress_output():
  File "/usr/lib/python3.11/contextlib.py", line 137, in __enter__
    return next(self.gen)
           ^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/assume/common/utils.py", line 640, in suppress_output
    original_stdout_fd = sys.stdout.fileno()
                         ^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/ipykernel/iostream.py", line 311, in fileno
    raise io.UnsupportedOperation("fileno")
io.UnsupportedOperation: fileno
```

I have been trying to understand why this happened, but I have not been successful. Please let me know if I missed something. Thanks!
