### 2025-08-11 838ea98
correctly suppress highs output in linopy (#622)
# Pull Request

The `suppress_output` method redirects the output buffers.
This is generally not a good idea, instead, the underlying source of
unwanted output should be tackled.

This PR fixes the output from highs in the redispatch market.

## Related Issue
This fixes #619 as this produced errors in Colab.

## Description
There is still a line `Running HiGHS 1.11.0 (git hash: 364c83a):
Copyright (c) 2025 HiGHS under MIT licence terms` which is printed for
every solving.
This is going to be fixed in linopy with
https://github.com/PyPSA/linopy/p
[…]
