### 2025-08-11 assume-framework/assume#619 @maurerle
Hi @igarizio, this is now fixed.
Thank you!
