### 2025-08-14 #8515 @kyleniemeyer
@maurerle thanks - can you now archive this state of the repo (using Zenodo, for example) and work through the other steps above? I will review the paper for any final edits.
