### 2025-08-14 #8515 @kyleniemeyer
@maurerle can you edit the author metadata on the archive to match the paper? I see some github handles rather than names.
