### 2025-08-14 #8515 @maurerle
@kyleniemeyer Thanks. I did update the names and added ORCID's where possible.
Is it fine that the zenodo metadata does even include minor contributions to the software, through single PRs?
Or should it match the paper metadata exactly?
