### 2025-08-14 #8515 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
